% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{AWMSI}
\alias{AWMSI}
\title{Area-Weighted Mean Shape Index}
\usage{
AWMSI(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the area-weighted mean shape index.
}
\description{
Calculates the area-weighted mean shape index.
}
\details{
The value must be greater than or equal to 1.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
