% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CA}
\alias{CA}
\title{Area in hectare of patches}
\usage{
CA(x)
}
\arguments{
\item{x}{is a matrix of data with patches identified as classes.}
}
\value{
A table with the values in hectares of each class and their proportion in percentage in the area.
}
\description{
Get the area in hectares of all raster classes.
}
