% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CACV_60}
\alias{CACV_60}
\title{Core Area Coefficient of Variation with edge of 60 meters}
\usage{
CACV_60(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the core area coefficient of variation in percentage, with subtracted edge of 60 meters.
}
\description{
Calculates the patch core area coefficient of variation in pecentage, when excluded the edge from 60 meters.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
