% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\alias{calculate_metrics}
\alias{calculate_metrics.RasterLayer}
\alias{calculate_metrics}
\alias{calculate_metrics.RasterStack}
\alias{calculate_metrics}
\alias{calculate_metrics.RasterBrick}
\alias{calculate_metrics}
\alias{calculate_metrics.list}
\title{Calculate metrics}
\usage{
calculate_metrics(landscape, what, ...)

\method{calculate_metrics}{RasterLayer}(landscape, what = "all", ...)

\method{calculate_metrics}{RasterStack}(landscape, what = "all", ...)

\method{calculate_metrics}{RasterBrick}(landscape, what = "all", ...)

\method{calculate_metrics}{list}(landscape, what = "all", ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Calculate all metrics of level
}
\examples{
calculate_metrics(landscape)
calculate_metrics(landscape_stack, what = "patch")

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
