% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccl_borders}
\alias{ccl_borders}
\title{8-way connectedness border states.}
\usage{
ccl_borders(m)
}
\arguments{
\item{m}{A two dimensional numeric matrix.}
}
\value{
Numeric matrix of the same dimensions as mat with
values indicating 8-way connected border states.
}
\description{
8-way neighborhood connected component border
(edge/vertice) states.
}
\details{
A cell is considered to have 8 distinct borders indexed
in counter clockwise order with values of 2^c(0:7) beginning
with the North edge and ending with the Northeast vertex. A cell
fully enclosed by similar neighbors has no borders ( a 0 value )
whereas when fully enclosed by dissimilar neighbors is fully
bordered ( a 255 value ).
}
\author{
Thell Fowler \email{thell@tbfowler.name}
}
\keyword{internal}
