% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccl_labels}
\alias{ccl_labels}
\title{8-way connected component labelling.}
\usage{
ccl_labels(m)
}
\arguments{
\item{m}{A two dimensional numeric matrix.}
}
\value{
Numeric matrix of the same dimensions as mat with
values indicating component identification label where values
increase in column major order.
}
\description{
8-way neighborhood connected component labelling of
all given values.
}
\author{
Thell Fowler \email{thell@tbfowler.name}
}
\keyword{internal}
