% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_buffer.R
\name{construct_buffer}
\alias{construct_buffer}
\title{construct_buffer}
\usage{
construct_buffer(coords, shape, size, return_vec = TRUE, verbose = TRUE)
}
\arguments{
\item{coords}{SpatVector, sf object or 2-column matrix with coordinates of sample points}

\item{shape}{String specifying plot shape. Either "circle" or "square"}

\item{size}{Size of sample plot. Equals the radius for circles or the
side-length for squares in map units}

\item{return_vec}{If TRUE, vector objects are returned.}

\item{verbose}{Print warning messages.}
}
\value{
matrix or SpatVector object
}
\description{
Internal function to construct plot area around coordinates
}
\examples{
coords <- matrix(c(10, 5, 25, 15, 5, 25), ncol = 2, byrow = TRUE)
construct_buffer(coords = coords, shape = "square", size = 5)

}
\keyword{internal}
