% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{data_info}
\alias{data_info}
\title{data_info}
\usage{
data_info(landscape)
}
\arguments{
\item{landscape}{RasterLayer}
}
\value{
raster
}
\description{
Data info
}
\details{
Check class of raster values and number of classes
}
\examples{
augusta_nlcd <- terra::rast(augusta_nlcd)
data_info(augusta_nlcd)
podlasie_ccilc <- terra::rast(podlasie_ccilc)
data_info(podlasie_ccilc)
landscape <- terra::rast(landscape)
data_info(landscape)

}
\keyword{internal}
