% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{ED}
\alias{ED}
\title{Edge Density}
\usage{
ED(x, a)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}

\item{a}{Total area of the landscape.}
}
\value{
The edge density in meteres per hectare (m/ha).
}
\description{
Calculates the total weighted class edges by the landscape area in meters per hectare (m/ha).
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
