% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lsm.R
\name{extract_lsm}
\alias{extract_lsm}
\alias{extract_lsm.RasterLayer}
\alias{extract_lsm.RasterStack}
\alias{extract_lsm.RasterBrick}
\alias{extract_lsm.stars}
\alias{extract_lsm.list}
\title{extract_lsm}
\usage{
extract_lsm(landscape, y, directions = 8, count_boundary,
  consider_boundary, classes_max, neighbourhood, ordered, base, full_name,
  verbose, progress, ...)

\method{extract_lsm}{RasterLayer}(landscape, y, directions = 8,
  count_boundary = FALSE, consider_boundary = FALSE,
  classes_max = NULL, neighbourhood = 4, ordered = TRUE,
  base = "log2", full_name = FALSE, verbose = TRUE,
  progress = FALSE, ...)

\method{extract_lsm}{RasterStack}(landscape, y, directions = 8,
  count_boundary = FALSE, consider_boundary = FALSE,
  classes_max = NULL, neighbourhood = 4, ordered = TRUE,
  base = "log2", full_name = FALSE, verbose = TRUE,
  progress = FALSE, ...)

\method{extract_lsm}{RasterBrick}(landscape, y, directions = 8,
  count_boundary = FALSE, consider_boundary = FALSE,
  classes_max = NULL, neighbourhood = 4, ordered = TRUE,
  base = "log2", full_name = FALSE, verbose = TRUE,
  progress = FALSE, ...)

\method{extract_lsm}{stars}(landscape, y, directions = 8,
  count_boundary = FALSE, consider_boundary = FALSE,
  classes_max = NULL, neighbourhood = 4, ordered = TRUE,
  base = "log2", full_name = FALSE, verbose = TRUE,
  progress = FALSE, ...)

\method{extract_lsm}{list}(landscape, y, directions = 8,
  count_boundary = FALSE, consider_boundary = FALSE,
  classes_max = NULL, neighbourhood = 4, ordered = TRUE,
  base = "log2", full_name = FALSE, verbose = TRUE,
  progress = FALSE, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{y}{Spatial object ( Spatialy*; SpatialPolygons*; SpatialLines; Extent or sf equivalents); two-column matrix/data.frame/tibble or cellnumbers that are used to extract landscapemetrics.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}

\item{count_boundary}{Include landscape boundary in edge length}

\item{consider_boundary}{Logical if cells that only neighbour the landscape
boundary should be considered as core}

\item{classes_max}{Potential maximum number of present classes}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}

\item{ordered}{The type of pairs considered. Either ordered (TRUE) or unordered (FALSE).
The default is TRUE.}

\item{base}{The unit in which entropy is measured. The default is "log2",
which compute entropy in "bits". "log" and "log10" can be also used.}

\item{full_name}{Should the full names of all functions be included in the
tibble.}

\item{verbose}{Print warning message if not sufficient patches are present}

\item{progress}{Print progress report}

\item{...}{Arguments passed to \code{raster::extract}.}
}
\value{
tibble
}
\description{
Extract landscapemetrics for certain cells
}
\details{
Extracts landscapemetrics for spatial objects. Returns a tibble with
landscapemetrics on patch level and the ID for the spatial object that was
used for extracting.
}
\examples{
points <- raster::sampleRandom(landscape, 20, sp = TRUE)
extract_lsm(landscape, points)

}
\seealso{
\code{\link{lsm_c_enn_mn}},
\code{\link{lsm_c_enn_sd}},
\code{\link{lsm_c_enn_cv}}, \cr
\code{\link{lsm_l_enn_mn}},
\code{\link{lsm_l_enn_sd}},
\code{\link{lsm_l_enn_cv}}
}
