% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragstats_class_podlasie}
\alias{fragstats_class_podlasie}
\title{Fragstats results for landscapemetrics::podlasie (class level)}
\format{A tibble object.}
\usage{
fragstats_class_podlasie
}
\description{
A single tibble for every spatial dataset included in landscapemetrics
that contains the FRAGSTAT results for every implemented metric on class
level.
}
\keyword{datasets}
