% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragstats_landscape_landscapestack}
\alias{fragstats_landscape_landscapestack}
\title{Fragstats results for landscapemetrics::landscape_stack (landscape level)}
\format{A tibble object.}
\usage{
fragstats_landscape_landscapestack
}
\description{
A single tibble for every spatial dataset included in landscapemetrics
that contains the FRAGSTAT results for every implemented metric on landscape
level.
}
\keyword{datasets}
