% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragstats_patch_augusta_nlcd}
\alias{fragstats_patch_augusta_nlcd}
\title{Fragstats results for landscapemetrics::augusta_nlcd (patch level)}
\format{A tibble object.}
\usage{
fragstats_patch_augusta_nlcd
}
\description{
A single tibble for every spatial dataset included in landscapemetrics
that contains the FRAGSTAT results for every implemented metric on patch
level.
}
\keyword{datasets}
