% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{get_boundaries}
\usage{
get_boundaries(landscape, directions = 4)
}
\arguments{
\item{directions}{4 (rooks case), 8 (queens case)}

\item{matrix}{matrix with values.}
}
\value{
matrix
}
\description{
Streamlined version of raster::boundaries that takes a matrix as input
}
\details{
Stripped down version of raster::boundaries so that we internally can parse
a matrix.
}
\examples{
test_matrix <- get_patches(landscape, class = 1, return_raster = FALSE)[[1]]
get_boundaries(test_matrix, directions = 4)

}
\references{
Robert J. Hijmans (2018). raster: Geographic Data Analysis and Modeling. R package version 2.8-4.
https://CRAN.R-project.org/package=raster
}
\keyword{internal}
