% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_extras.R
\name{get_core_patch}
\alias{get_core_patch}
\title{get_core_patch}
\usage{
get_core_patch(
  landscape_mat,
  classes,
  class_patches,
  directions,
  consider_boundary,
  edge_depth,
  resolution
)
}
\arguments{
\item{landscape_mat}{A matrix object}

\item{classes}{A vector with unique values (output of get_unique_values_int)}

\item{class_patches}{A list with matrices of patches for each class (output of get_class_patches)}

\item{directions}{The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).}

\item{consider_boundary}{Logical if cells that only neighbour the landscape boundary should be considered as core}

\item{edge_depth}{Distance (in cells) a cell has to be away from the patch edge to be considered as core cell}

\item{resolution}{A vector with two numbers (usually calculated using terra::res)}
}
\value{
A named vector with core area values for each patch (names are class IDs)
}
\description{
Core area of each patch in each class
}
\details{
Calculate core area of each patch in each class. Core area is the area within a patch
that is not on the edge. A cell is defined as core area if the cell has no neighbour
with a different value than itself (rook's case).
}
\examples{
landscape <- terra::rast(landscapemetrics::landscape)
landscape_mat <- terra::as.matrix(landscape, wide = TRUE)
classes <- landscapemetrics:::get_unique_values_int(landscape_mat)
class_patches <- get_class_patches(landscape_mat, classes, directions = 8)
core <- get_core_patch(landscape_mat, classes, class_patches, 8, FALSE, 1, terra::res(landscape))

}
\keyword{internal}
