% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lsm.R
\name{get_lsm}
\alias{get_lsm}
\alias{get_lsm.RasterLayer}
\alias{get_lsm.RasterStack}
\alias{get_lsm.RasterBrick}
\alias{get_lsm.stars}
\alias{get_lsm.list}
\title{get_lsm}
\usage{
get_lsm(landscape, metric, name, type, what, directions, progress, ...)

\method{get_lsm}{RasterLayer}(landscape, metric = NULL, name = NULL,
  type = NULL, what = NULL, directions = 8, progress = FALSE, ...)

\method{get_lsm}{RasterStack}(landscape, metric = NULL, name = NULL,
  type = NULL, what = NULL, directions = 8, progress = FALSE, ...)

\method{get_lsm}{RasterBrick}(landscape, metric = NULL, name = NULL,
  type = NULL, what = NULL, directions = 8, progress = FALSE, ...)

\method{get_lsm}{stars}(landscape, metric = NULL, name = NULL,
  type = NULL, what = NULL, directions = 8, progress = FALSE, ...)

\method{get_lsm}{list}(landscape, metric = NULL, name = NULL,
  type = NULL, what = NULL, directions = 8, progress = FALSE, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{metric}{Abbreviation of metrics (e.g. 'area').}

\item{name}{Full name of metrics (e.g. 'core area')}

\item{type}{Type according to FRAGSTATS grouping (e.g. 'aggregation metrics').}

\item{what}{Selected level of metrics: either "patch", "class" or "landscape".
It is also possible to specify functions as a vector of strings, e.g. \code{what = c("lsm_c_ca", "lsm_l_ta")}.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}

\item{progress}{Print progress report.}

\item{...}{Arguments passed on to \code{calculate_lsm()}.}
}
\value{
list
}
\description{
Get landscape metric values
}
\details{
The functions returns a nested list with \code{RasterLayer}s. The first level
contains each input layer (only one element if \code{RasterLayer} was provided).
The second level contains a \code{RasterLayer} for each selected metric
(see \code{list_lsm} for details) where each cell has the landscape metric
value of the patch it belongs to. Only patch level metrics are allowed.
}
\examples{
get_lsm(landscape, what = "lsm_p_area")

}
\seealso{
\code{\link{list_lsm}} \cr
\code{\link{show_lsm}}
}
