% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_extras.R
\name{get_points}
\alias{get_points}
\title{get_points}
\usage{
get_points(landscape_mat, resolution)
}
\arguments{
\item{landscape_mat}{A matrix object}

\item{resolution}{A vector with two numbers (usually calculated using terra::res)}
}
\value{
matrix with three columns: col, row, value
}
\description{
Raster to col, row, value
}
\details{
The col and row values are multiplied by the resolution to get the (internal) coordinates of the points.
}
\examples{
landscape <- terra::rast(landscapemetrics::landscape)
landscape_mat <- terra::as.matrix(landscape, wide = TRUE)
resolution <- terra::res(landscape)
get_points(landscape_mat, resolution)

}
\keyword{internal}
