% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_values.R
\name{get_unique_values}
\alias{get_unique_values}
\alias{get_unique_values.vector}
\alias{get_unique_values.matrix}
\alias{get_unique_values.RasterLayer}
\alias{get_unique_values.RasterStack}
\alias{get_unique_values.RasterBrick}
\title{get_unique_values}
\usage{
get_unique_values(x)

\method{get_unique_values}{vector}(x)

\method{get_unique_values}{matrix}(x)

\method{get_unique_values}{RasterLayer}(x)

\method{get_unique_values}{RasterStack}(x)

\method{get_unique_values}{RasterBrick}(x)
}
\arguments{
\item{x}{vector, matrix or Raster* object}
}
\description{
This function returns the unique values of an object.
}
\details{
Fast and memory friendly Rcpp implementation to find the unique values of an object.
}
\examples{
get_unique_values(landscape)

}
