% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irregular_areas}
\alias{irregular_areas}
\title{Irregular polygons}
\format{
An sf object.
}
\usage{
irregular_areas
}
\description{
An example irregular polygons for the area in \code{augusta_nlcd}
}
\keyword{datasets}
