% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape}
\alias{landscape}
\title{Example map (random cluster neutral landscape model).}
\format{
A raster object.
}
\source{
Simulated neutral landscape model with R. https://github.com/ropensci/NLMR/
}
\usage{
landscape
}
\description{
An example map to show landscapemetrics functionality
generated with the \code{nlm_randomcluster()} algorithm.
}
\keyword{datasets}
