% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_as_list.R
\name{landscape_as_list}
\alias{landscape_as_list}
\alias{landscape_as_list.RasterLayer}
\alias{landscape_as_list.RasterStack}
\alias{landscape_as_list.RasterBrick}
\alias{landscape_as_list.stars}
\alias{landscape_as_list.SpatRaster}
\alias{landscape_as_list.list}
\alias{landscape_as_list.matrix}
\alias{landscape_as_list.numeric}
\title{Landscape as list}
\usage{
landscape_as_list(landscape)

\method{landscape_as_list}{RasterLayer}(landscape)

\method{landscape_as_list}{RasterStack}(landscape)

\method{landscape_as_list}{RasterBrick}(landscape)

\method{landscape_as_list}{stars}(landscape)

\method{landscape_as_list}{SpatRaster}(landscape)

\method{landscape_as_list}{list}(landscape)

\method{landscape_as_list}{matrix}(landscape)

\method{landscape_as_list}{numeric}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick, Stars or a list of rasterLayers}
}
\value{
list
}
\description{
Convert raster input to list
}
\details{
Mainly for internal use
}
\examples{
landscape_as_list(raster::stack(landscape, landscape))

}
