% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_shdi.R
\name{landscape_shdi}
\alias{landscape_shdi}
\title{Shannon's Diversity Index}
\usage{
landscape_shdi(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 0
}
\description{
Number of patch types (classes) in the landscape
}
\examples{
library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))
landscape_shdi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
 PNW-351.
}
