% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapemetrics-package.R
\docType{package}
\name{landscapemetrics-package}
\alias{landscapemetrics}
\alias{landscapemetrics-package}
\title{landscapemetrics: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/landscapemetrics}
  \item Report bugs at \url{https://github.com/marcosci/landscapemetrics/issues}
}

}
\keyword{internal}
