% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapemetrics-package.R
\docType{package}
\name{landscapemetrics}
\alias{landscapemetrics}
\alias{landscapemetrics-package}
\alias{landscapemetrics-package}
\title{landscapemetrics}
\description{
xxx
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/landscapemetrics}
  \item Report bugs at \url{https://github.com/marcosci/landscapemetrics/issues}
}

}
