% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsm_abbreviations_names}
\alias{lsm_abbreviations_names}
\title{Tibble of abbreviations coming from FRAGSTATS}
\format{
A tibble object.
}
\usage{
lsm_abbreviations_names
}
\description{
A single tibble for every abbreviation of every metric that is
reimplemented in landscapemetrics and its corresponding full name
in the literature.
}
\details{
Can be used after calculating the metric(s) with a join to have
a more readable results tibble or for visualizing your results.
}
\examples{
landscape <- terra::rast(landscape)
patch_area <- lsm_p_area(landscape)
patch_area <- merge(x = patch_area, y = lsm_abbreviations_names, by = c("level", "metric"))

}
\keyword{datasets}
