% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_area_cv.R
\name{lsm_c_area_cv}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.RasterLayer}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.RasterStack}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.RasterBrick}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.list}
\title{AREA_CV}
\usage{
lsm_c_area_cv(landscape)

\method{lsm_c_area_cv}{RasterLayer}(landscape)

\method{lsm_c_area_cv}{RasterStack}(landscape)

\method{lsm_c_area_cv}{RasterBrick}(landscape)

\method{lsm_c_area_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers}
}
\value{
tibble
}
\description{
Coeffiecent of variation of patch area (class level)
}
\details{
\deqn{AREA_CV = cv(AREA[patch_{ij}])}
where \eqn{AREA[patch_{ij}]} is the area of each patch in hectares

AREA_CV is an 'Area and Edge metric'. The metric summarises each class
as the coeffiecent of variation of all patch areas belonging to class i

\subsection{Units}{Hectares}
\subsection{Range}{AREA_CV >= 0}
\subsection{Behaviour}{Increases as the variation of patch areas increases}
}
\examples{
lsm_c_area_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}} and \code{\link{cv}}
}
