% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_area_mn.R
\name{lsm_c_area_mn}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.RasterLayer}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.RasterStack}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.RasterBrick}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.list}
\title{AREA_MN}
\usage{
lsm_c_area_mn(landscape)

\method{lsm_c_area_mn}{RasterLayer}(landscape)

\method{lsm_c_area_mn}{RasterStack}(landscape)

\method{lsm_c_area_mn}{RasterBrick}(landscape)

\method{lsm_c_area_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers}
}
\value{
tibble
}
\description{
Mean of patch area (class level)
}
\details{
\deqn{AREA_MN = mean(AREA[patch_{ij}])}
where \eqn{AREA[patch_{ij}]} is the area of each patch in hectares

AREA_MN is an 'Area and Edge metric'. The metric summarises each class
as the mean of all patch areas belonging to class i

\subsection{Units}{Hectares}
\subsection{Range}{AREA_MN > 0}
\subsection{Behaviour}{Increases as the patch areas increase}
}
\examples{
lsm_c_area_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}} and \code{\link{mean}}
}
