% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ca.R
\name{lsm_c_ca}
\alias{lsm_c_ca}
\alias{lsm_c_ca}
\alias{lsm_c_ca.RasterLayer}
\alias{lsm_c_ca}
\alias{lsm_c_ca.RasterStack}
\alias{lsm_c_ca}
\alias{lsm_c_ca.RasterBrick}
\alias{lsm_c_ca}
\alias{lsm_c_ca.list}
\title{CA (class level)}
\usage{
lsm_c_ca(landscape)

\method{lsm_c_ca}{RasterLayer}(landscape)

\method{lsm_c_ca}{RasterStack}(landscape)

\method{lsm_c_ca}{RasterBrick}(landscape)

\method{lsm_c_ca}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total (class) area (Area and edge metric)
}
\details{
\deqn{CA = sum(AREA[patch_{ij}])}
where \eqn{AREA[patch_{ij}]} is the area of each patch in hectares.

CA is an 'Area and egdge metric' and a measure of composition.
The total (class) area sums the area of all patches belonging to class i.
It shows if the landcsape is e.g. dominated by one class or if all classes
are equally present. CA is an absolute measure, making comparisons among landscapes with different
total areas difficult.

\subsection{Units}{Hectares}
\subsection{Range}{CA > 0}
\subsection{Behaviour}{Aprroaches 0 as the patch areas of class i become small.
Increases, without limit, as the patch areas of class i become large. CA = TA if only
one class is present.}
}
\examples{
lsm_c_ca(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{sum}}, \cr
\code{\link{lsm_l_ta}}
}
