% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cai_cv.R
\name{lsm_c_cai_cv}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.RasterLayer}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.RasterStack}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.RasterBrick}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.list}
\title{CAI_CV}
\usage{
lsm_c_cai_cv(landscape)

\method{lsm_c_cai_cv}{RasterLayer}(landscape)

\method{lsm_c_cai_cv}{RasterStack}(landscape)

\method{lsm_c_cai_cv}{RasterBrick}(landscape)

\method{lsm_c_cai_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coefficient of variation of the core area index (class level)
}
\details{
\deqn{CAI_CV = cv(CAI[patch_{ij}]}
where \eqn{CAI[patch_{ij}]} is the core area index of each patch

CAI_CV is a 'Core area metric'. The metric summarises each class
as the coeffiecent of variation of the core area index of all patches
belonging to class i. The core area index is the percentag of core area
in relation to patch area

\subsection{Units}{Percent}
\subsection{Range}{CAI_CV >= 0}
\subsection{Behaviour}{Increases as the variation of the core area indices increases}
}
\examples{
lsm_c_cai_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_cai}} and \code{\link{cv}}
}
