% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cai_mn.R
\name{lsm_c_cai_mn}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterLayer}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterStack}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterBrick}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.list}
\title{CAI_MN}
\usage{
lsm_c_cai_mn(landscape)

\method{lsm_c_cai_mn}{RasterLayer}(landscape)

\method{lsm_c_cai_mn}{RasterStack}(landscape)

\method{lsm_c_cai_mn}{RasterBrick}(landscape)

\method{lsm_c_cai_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of the core area index (class level)
}
\details{
\deqn{CAI_MN = mean(CAI[patch_{ij}]}
where \eqn{CAI[patch_{ij}]} is the core area index of each patch

CAI_MN is a 'Core area metric'. The metric summarises each class
as the mean of the core area index of all patches belonging to class i.
The core area index is the percentag of core area in relation to patch area

\subsection{Units}{Percent}
\subsection{Range}{CAI_MN > 0}
\subsection{Behaviour}{Increases as the corea area indices increase}
}
\examples{
lsm_c_cai_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_cai}} and \code{\link{mean}}
}
