% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cai_sd.R
\name{lsm_c_cai_sd}
\alias{lsm_c_cai_sd}
\alias{lsm_c_cai_sd}
\alias{lsm_c_cai_sd.RasterLayer}
\alias{lsm_c_cai_sd}
\alias{lsm_c_cai_sd.RasterStack}
\alias{lsm_c_cai_sd}
\alias{lsm_c_cai_sd.RasterBrick}
\alias{lsm_c_cai_sd}
\alias{lsm_c_cai_sd.list}
\title{Core area index distribution (class level)}
\usage{
lsm_c_cai_sd(landscape, directions)

\method{lsm_c_cai_sd}{RasterLayer}(landscape, directions = 8)

\method{lsm_c_cai_sd}{RasterStack}(landscape, directions = 8)

\method{lsm_c_cai_sd}{RasterBrick}(landscape, directions = 8)

\method{lsm_c_cai_sd}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{???
Equals the standard deviation of the core area index of class i.
The core area index equals the percentage of a patch that is core area
\deqn{CAI_SD = sd(CAI[patch_i]}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}}
}
\value{
tibble
}
\description{
Standard deviation of core area index (class level)
}
\examples{
lsm_c_cai_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
