% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_circle_sd.R
\name{lsm_c_circle_sd}
\alias{lsm_c_circle_sd}
\alias{lsm_c_circle_sd}
\alias{lsm_c_circle_sd.RasterLayer}
\alias{lsm_c_circle_sd}
\alias{lsm_c_circle_sd.RasterStack}
\alias{lsm_c_circle_sd}
\alias{lsm_c_circle_sd.RasterBrick}
\alias{lsm_c_circle_sd}
\alias{lsm_c_circle_sd.list}
\title{Related Circumscribing Circle distribution (CIRCLE)}
\usage{
lsm_c_circle_sd(landscape)

\method{lsm_c_circle_sd}{RasterLayer}(landscape)

\method{lsm_c_circle_sd}{RasterStack}(landscape)

\method{lsm_c_circle_sd}{RasterBrick}(landscape)

\method{lsm_c_circle_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of related circumscribing circle (class)
}
\details{
Equals the standard deviation of the related circumscribing circle of
class i. Equals the 1 - the patch area (m^2) divided by the area (m^2) of the smallest
circumscribing circle.
\deqn{CIRCLE_SD = sd(CIRCLE[patch_i])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{????}
}
\examples{
lsm_c_circle_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
