% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cohesion.R
\name{lsm_c_cohesion}
\alias{lsm_c_cohesion}
\alias{lsm_c_cohesion}
\alias{lsm_c_cohesion.RasterLayer}
\alias{lsm_c_cohesion}
\alias{lsm_c_cohesion.RasterStack}
\alias{lsm_c_cohesion}
\alias{lsm_c_cohesion.RasterBrick}
\alias{lsm_c_cohesion}
\alias{lsm_c_cohesion.list}
\title{Patch Cohesion Index (class level)}
\usage{
lsm_c_cohesion(landscape)

\method{lsm_c_cohesion}{RasterLayer}(landscape)

\method{lsm_c_cohesion}{RasterStack}(landscape)

\method{lsm_c_cohesion}{RasterBrick}(landscape)

\method{lsm_c_cohesion}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Percentage of Like Adjacencies (class level)
}
\details{
???

\deqn{???}
\subsection{Units}{???}
\subsection{Ranges}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_cohesion(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
