% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core.R
\name{lsm_c_core}
\alias{lsm_c_core}
\alias{lsm_c_core}
\alias{lsm_c_core.RasterLayer}
\alias{lsm_c_core}
\alias{lsm_c_core.RasterStack}
\alias{lsm_c_core}
\alias{lsm_c_core.RasterBrick}
\alias{lsm_c_core}
\alias{lsm_c_core.list}
\title{Total area of core areas  (class level)}
\usage{
lsm_c_core(landscape, directions)

\method{lsm_c_core}{RasterLayer}(landscape, directions = 4)

\method{lsm_c_core}{RasterStack}(landscape, directions = 4)

\method{lsm_c_core}{RasterBrick}(landscape, directions = 4)

\method{lsm_c_core}{list}(landscape, directions = 4)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{...}
}
\value{
tibble
}
\description{
Area of corea area (class level)
}
\examples{
lsm_c_core(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
