% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core_cv.R
\name{lsm_c_core_cv}
\alias{lsm_c_core_cv}
\alias{lsm_c_core_cv}
\alias{lsm_c_core_cv.RasterLayer}
\alias{lsm_c_core_cv}
\alias{lsm_c_core_cv.RasterStack}
\alias{lsm_c_core_cv}
\alias{lsm_c_core_cv.RasterBrick}
\alias{lsm_c_core_cv}
\alias{lsm_c_core_cv.list}
\title{Core area distribution (class level)}
\usage{
lsm_c_core_cv(landscape)

\method{lsm_c_core_cv}{RasterLayer}(landscape)

\method{lsm_c_core_cv}{RasterStack}(landscape)

\method{lsm_c_core_cv}{RasterBrick}(landscape)

\method{lsm_c_core_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coefficient of variation (class level)
}
\examples{
lsm_c_core_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
