% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core_mn.R
\name{lsm_c_core_mn}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterLayer}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterStack}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterBrick}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.list}
\title{CORE_MN (class level)}
\usage{
lsm_c_core_mn(landscape)

\method{lsm_c_core_mn}{RasterLayer}(landscape)

\method{lsm_c_core_mn}{RasterStack}(landscape)

\method{lsm_c_core_mn}{RasterBrick}(landscape)

\method{lsm_c_core_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of core area (Core area metric)
}
\details{
\deqn{CORE_{MN} = mean(CORE[patch_{ij}])}
where \eqn{CORE[patch_{ij}]} is the core area in square meters of each patch.

CORE_MN is a 'Core area metric' and equals the mean of core areas of all patches
belonging to class i. The core area is defined as all cells that have no
neighbour with a different value than themselves (rook's case).

\subsection{Units}{Hectares}
\subsection{Range}{CORE_MN >= 0}
\subsection{Behaviour}{Equals CORE_MN = 0 if CORE = 0 for all patches. Increases,
without limit, as the corea area indices increase.}
}
\examples{
lsm_c_core_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_core}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_sd}},
\code{\link{lsm_c_cv}}, \cr
\code{\link{lsm_l_mn}},
\code{\link{lsm_l_sd}},
\code{\link{lsm_l_cv}}
}
