% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core_sd.R
\name{lsm_c_core_sd}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.RasterLayer}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.RasterStack}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.RasterBrick}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.list}
\title{Core area distribution (class level)}
\usage{
lsm_c_core_sd(landscape)

\method{lsm_c_core_sd}{RasterLayer}(landscape)

\method{lsm_c_core_sd}{RasterStack}(landscape)

\method{lsm_c_core_sd}{RasterBrick}(landscape)

\method{lsm_c_core_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of patch core area (class level)
}
\details{
Equals the standard deviation of the patch core area of class i
\deqn{CORE_SD = sd(core[patch_i])}
\subsection{Units}{Square meter (assuming that the input cellsize is in meter)}
\subsection{Range}{???}
}
\examples{
lsm_c_core_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
