% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cpland.R
\name{lsm_c_cpland}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.RasterLayer}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.RasterStack}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.RasterBrick}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.list}
\title{Core area percentage of landscape (class level)}
\usage{
lsm_c_cpland(landscape, directions)

\method{lsm_c_cpland}{RasterLayer}(landscape, directions = 4)

\method{lsm_c_cpland}{RasterStack}(landscape, directions = 4)

\method{lsm_c_cpland}{RasterBrick}(landscape, directions = 4)

\method{lsm_c_cpland}{list}(landscape, directions = 4)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{...}
}
\value{
tibble
}
\description{
Core area of class divided by landscape area (class level)
}
\examples{
lsm_c_cpland(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
