% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_dcad.R
\name{lsm_c_dcad}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterLayer}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterStack}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterBrick}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.list}
\title{Disjunct core area density (class level)}
\usage{
lsm_c_dcad(landscape)

\method{lsm_c_dcad}{RasterLayer}(landscape)

\method{lsm_c_dcad}{RasterStack}(landscape)

\method{lsm_c_dcad}{RasterBrick}(landscape)

\method{lsm_c_dcad}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Disjunct core area density (class level)
}
\examples{
lsm_c_dcad(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
