% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_dcore_mn.R
\name{lsm_c_dcore_mn}
\alias{lsm_c_dcore_mn}
\alias{lsm_c_dcore_mn}
\alias{lsm_c_dcore_mn.RasterLayer}
\alias{lsm_c_dcore_mn}
\alias{lsm_c_dcore_mn.RasterStack}
\alias{lsm_c_dcore_mn}
\alias{lsm_c_dcore_mn.RasterBrick}
\alias{lsm_c_dcore_mn}
\alias{lsm_c_dcore_mn.list}
\title{DCORE_MN (class level)}
\usage{
lsm_c_dcore_mn(landscape)

\method{lsm_c_dcore_mn}{RasterLayer}(landscape)

\method{lsm_c_dcore_mn}{RasterStack}(landscape)

\method{lsm_c_dcore_mn}{RasterBrick}(landscape)

\method{lsm_c_dcore_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean number of disjunct core areas (Core area metric)
}
\details{
\deqn{DCORE_{MN} = mean(NCORE[patch_{ij}])}
where \eqn{NCORE[patch_{ij}]} is the number of core areas.

DCORE_MN is an 'Core area metric'. It summarises each class as the mean of all
patch areas belonging to class i. A cell is defined as core if the cell
has no neighbour with a different value than itself (rook's case). NCORE counts the disjunct
core areas, whereby a core area is a 'patch within the patch' containing only core cells.

\subsection{Units}{None}
\subsection{Range}{DCORE_MN > 0}
\subsection{Behaviour}{Equals DCORE_MN = 0 if NCORE = 0 for all patches. Increases,
without limit, as the number of disjunct corea areas increases.}
}
\examples{
lsm_c_dcore_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_ncore}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_dcore_sd}},
\code{\link{lsm_c_dcore_cv}}, \cr
\code{\link{lsm_l_dcore_mn}},
\code{\link{lsm_l_dcore_sd}},
\code{\link{lsm_l_dcore_cv}}
}
