% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ed.R
\name{lsm_c_ed}
\alias{lsm_c_ed}
\title{Edge Density (ED, class scale)}
\usage{
lsm_c_ed(landscape, class)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{class}{Which class}
}
\value{
tibble
}
\description{
Edge Density
}
\examples{
lsm_c_ed(landscape, 1)
lsm_c_ed(landscape_stack, 1)
}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
