% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_frac_cv.R
\name{lsm_c_frac_cv}
\alias{lsm_c_frac_cv}
\alias{lsm_c_frac_cv}
\alias{lsm_c_frac_cv.RasterLayer}
\alias{lsm_c_frac_cv}
\alias{lsm_c_frac_cv.RasterStack}
\alias{lsm_c_frac_cv}
\alias{lsm_c_frac_cv.RasterBrick}
\alias{lsm_c_frac_cv}
\alias{lsm_c_frac_cv.list}
\title{FRAC_CV (class level)}
\usage{
lsm_c_frac_cv(landscape)

\method{lsm_c_frac_cv}{RasterLayer}(landscape)

\method{lsm_c_frac_cv}{RasterStack}(landscape)

\method{lsm_c_frac_cv}{RasterBrick}(landscape)

\method{lsm_c_frac_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation fractal dimension index (Shape metric)
}
\details{
\deqn{FRAC_{CV} = cv(FRAC[patch_{ij}])}
where \eqn{FRAC[patch_{ij}]} equals the fractal dimension index of each patch.

FRAC_CV is a 'Shape metric'. The metric summarises each class
as the coeffiecent of variation of the fractal dimension index of all patches
belonging to class i. The fractal dimenstion index is based on the patch perimeter and
the patch area and describes the patch complexity. The coeffiecent of variation is
scaled to the mean and comparable among different landscapes.

\subsection{Units}{None}
\subsection{Range}{FRAC_CV >= 0 }
\subsection{Behaviour}{Equals FRAC_CV = 0 if the fractal dimension index is identical
for all patches. Increases, without limit, as the variation of the fractal dimension
indices increases.}
}
\examples{
lsm_c_frac_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_frac}},
\code{\link{cv}}, \cre
\code{\link{lsm_c_frac_mn}},
\code{\link{lsm_c_frac_sd}}, \cr
\code{\link{lsm_l_frac_mn}},
\code{\link{lsm_l_frac_sd}},
\code{\link{lsm_l_frac_cv}}
}
