% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_gyrate_mn.R
\name{lsm_c_gyrate_mn}
\alias{lsm_c_gyrate_mn}
\alias{lsm_c_gyrate_mn}
\alias{lsm_c_gyrate_mn.RasterLayer}
\alias{lsm_c_gyrate_mn}
\alias{lsm_c_gyrate_mn.RasterStack}
\alias{lsm_c_gyrate_mn}
\alias{lsm_c_gyrate_mn.RasterBrick}
\alias{lsm_c_gyrate_mn}
\alias{lsm_c_gyrate_mn.list}
\title{GYRATE_MN (class level)}
\usage{
lsm_c_gyrate_mn(landscape, directions)

\method{lsm_c_gyrate_mn}{RasterLayer}(landscape, directions = 8)

\method{lsm_c_gyrate_mn}{RasterStack}(landscape, directions = 8)

\method{lsm_c_gyrate_mn}{RasterBrick}(landscape, directions = 8)

\method{lsm_c_gyrate_mn}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Mean radius of gyration (Area and edge metric)
}
\details{
\deqn{GYRATE_{MN} = mean(GYRATE[patch_{ij}])}
where \eqn{GYRATE[patch_{ij}]} equals the radius of gyration of each patch.

GYRATE_MN is an 'Area and edge metric'. The metric summarises each class
as the mean of the radius of gyration of all patches belonging to class i.
GYRATE measures the distance from each cell to the patch centroid and is based on
cell center-to-cell center distances. The metrics characterises
both the patch area and compactness. The Coefficient of variation is
scaled to the mean and comparable among different landscapes.

\subsection{Units}{Meters}
\subsection{Range}{GYRATE_MN >= 0 }
\subsection{Behaviour}{Approaches GYRATE_MN = 0 if every patch is a single cell. Increases,
without limit, when only one patch is present.}
}
\examples{
lsm_c_gyrate_mn(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Keitt, T. H., Urban, D. L., & Milne, B. T. 1997. Detecting critical scales
in fragmented landscapes. Conservation ecology, 1(1).
}
\seealso{
\code{\link{lsm_p_gyrate}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_gyrate_sd}},
\code{\link{lsm_c_gyrate_cv}}, \cr
\code{\link{lsm_l_gyrate_mn}},
\code{\link{lsm_l_gyrate_sd}},
\code{\link{lsm_l_gyrate_cv}}
}
