% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_gyrate_sd.R
\name{lsm_c_gyrate_sd}
\alias{lsm_c_gyrate_sd}
\alias{lsm_c_gyrate_sd}
\alias{lsm_c_gyrate_sd.RasterLayer}
\alias{lsm_c_gyrate_sd}
\alias{lsm_c_gyrate_sd.RasterStack}
\alias{lsm_c_gyrate_sd}
\alias{lsm_c_gyrate_sd.RasterBrick}
\alias{lsm_c_gyrate_sd}
\alias{lsm_c_gyrate_sd.list}
\title{Radius of Gyration (GYRATE)}
\usage{
lsm_c_gyrate_sd(landscape)

\method{lsm_c_gyrate_sd}{RasterLayer}(landscape)

\method{lsm_c_gyrate_sd}{RasterStack}(landscape)

\method{lsm_c_gyrate_sd}{RasterBrick}(landscape)

\method{lsm_c_gyrate_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Radius of Gyration (class)
}
\details{
Equals the mean distance of each cell centroid in a patch to the centroid of the whole patch (mean location of all cell centroids).
\subsection{Units}{Meters}
\subsection{Range}{GYRATE >= 0, without limit}
\subsection{Behaviour}{0 if single cell, maximum if patch occupies the entire landscape.}
}
\examples{
lsm_c_gyrate_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
