% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_iji.R
\name{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterLayer}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterStack}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterBrick}
\alias{lsm_c_iji}
\alias{lsm_c_iji.list}
\title{Interspersion and Juxtaposition index (class level)}
\usage{
lsm_c_iji(landscape, ...)

\method{lsm_c_iji}{RasterLayer}(landscape, ...)

\method{lsm_c_iji}{RasterStack}(landscape, ...)

\method{lsm_c_iji}{RasterBrick}(landscape, ...)

\method{lsm_c_iji}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Interspersion and Juxtaposition index (class level)
}
\details{
???
\deqn{???}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_iji(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
