% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_lpi.R
\name{lsm_c_lpi}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterLayer}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterStack}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterBrick}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.list}
\title{Largest patch index (class level)}
\usage{
lsm_c_lpi(landscape)

\method{lsm_c_lpi}{RasterLayer}(landscape)

\method{lsm_c_lpi}{RasterStack}(landscape)

\method{lsm_c_lpi}{RasterBrick}(landscape)

\method{lsm_c_lpi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Largest patch index of class i (class level)
}
\details{
The largest patch index equals the area of the larges patch of class i
divided by the total area. It is a simple measure of dominance
\deqn{LPI = (max(area[patch_i]) / total area) * 100}
\subsection{Units}{Percentage}
\subsection{Range}{0 < LPI <= 100}
\subsection{Behaviour}{LPI approaches LPI = 0 when the largest patch of the corresponding class is becoming small
and equals LPI = 100 when only one class and patch is present}
}
\examples{
lsm_c_lpi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
