% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_lsi.R
\name{lsm_c_lsi}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterLayer}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterStack}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterBrick}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.list}
\title{LSI (class level)}
\usage{
lsm_c_lsi(landscape)

\method{lsm_c_lsi}{RasterLayer}(landscape)

\method{lsm_c_lsi}{RasterStack}(landscape)

\method{lsm_c_lsi}{RasterBrick}(landscape)

\method{lsm_c_lsi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Landscape shape index (Aggregation index)
}
\details{
\deqn{LSI = \frac{e_{i}}{\min e_{i}}}
where

The landscape shape index equals a quarter of the sum of all edges of class i
divided by the square root of the total area.
\subsection{Units}{none}
\subsection{Ranges}{LSI >= 1}
\subsection{Behaviour}{Equals LSI = 1 when only one class and patch is present and
increases when the length of edges increases, i.e. the patches of class i become more complex}
}
\examples{
lsm_c_lsi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
