% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_mesh.R
\name{lsm_c_mesh}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterLayer}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterStack}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterBrick}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.list}
\title{Effective mesh size (class level)}
\usage{
lsm_c_mesh(landscape)

\method{lsm_c_mesh}{RasterLayer}(landscape)

\method{lsm_c_mesh}{RasterStack}(landscape)

\method{lsm_c_mesh}{RasterBrick}(landscape)

\method{lsm_c_mesh}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Effective mesh size (class level)
}
\examples{
lsm_c_mesh(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
