% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ncore_cv.R
\name{lsm_c_ncore_cv}
\alias{lsm_c_ncore_cv}
\alias{lsm_c_ncore_cv}
\alias{lsm_c_ncore_cv.RasterLayer}
\alias{lsm_c_ncore_cv}
\alias{lsm_c_ncore_cv.RasterStack}
\alias{lsm_c_ncore_cv}
\alias{lsm_c_ncore_cv.RasterBrick}
\alias{lsm_c_ncore_cv}
\alias{lsm_c_ncore_cv.list}
\title{Number of core areas distribution (class level)}
\usage{
lsm_c_ncore_cv(landscape)

\method{lsm_c_ncore_cv}{RasterLayer}(landscape)

\method{lsm_c_ncore_cv}{RasterStack}(landscape)

\method{lsm_c_ncore_cv}{RasterBrick}(landscape)

\method{lsm_c_ncore_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation of number of core areas (class level)
}
\details{
Equals the coeffiecent of variation of number of core area of class i.
A core area is a 'patch within the patch' without any edge cells. In other words,
the number of patches within the patch that only have neighbouring cells of the same type
\deqn{NCORE_CV = cv(NCORE[patch_i])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_ncore_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
