% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ncore_mn.R
\name{lsm_c_ncore_mn}
\alias{lsm_c_ncore_mn}
\alias{lsm_c_ncore_mn}
\alias{lsm_c_ncore_mn.RasterLayer}
\alias{lsm_c_ncore_mn}
\alias{lsm_c_ncore_mn.RasterStack}
\alias{lsm_c_ncore_mn}
\alias{lsm_c_ncore_mn.RasterBrick}
\alias{lsm_c_ncore_mn}
\alias{lsm_c_ncore_mn.list}
\title{Number of core areas distribution (class level)}
\usage{
lsm_c_ncore_mn(landscape, directions)

\method{lsm_c_ncore_mn}{RasterLayer}(landscape)

\method{lsm_c_ncore_mn}{RasterStack}(landscape)

\method{lsm_c_ncore_mn}{RasterBrick}(landscape)

\method{lsm_c_ncore_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of number of core areas (class level)
}
\details{
Equals the mean of number of core area of class i.
A core area is a 'patch within the patch' without any edge cells. In other words,
the number of patches within the patch that only have neighbouring cells of the same type
\deqn{NCORE_MN = mean(NCORE[patch_i])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_ncore_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
