% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ncore_sd.R
\name{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.RasterLayer}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.RasterStack}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.RasterBrick}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.list}
\title{Number of core areas distribution (class level)}
\usage{
lsm_c_ncore_sd(landscape, directions)

\method{lsm_c_ncore_sd}{RasterLayer}(landscape, directions = 8)

\method{lsm_c_ncore_sd}{RasterStack}(landscape, directions = 8)

\method{lsm_c_ncore_sd}{RasterBrick}(landscape, directions = 8)

\method{lsm_c_ncore_sd}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{???}
}
\value{
tibble
}
\description{
Standart deviation of number of core areas (class level)
}
\details{
Equals the standard deviation of number of core area of class i.
A core area is a 'patch within the patch' without any edge cells. In other words,
the number of patches within the patch that only have neighbouring cells of the same type
\deqn{NCORE_SD = sd(NCORE[patch_i])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_ncore_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
