% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ndca.R
\name{lsm_c_ndca}
\alias{lsm_c_ndca}
\alias{lsm_c_ndca}
\alias{lsm_c_ndca.RasterLayer}
\alias{lsm_c_ndca}
\alias{lsm_c_ndca.RasterStack}
\alias{lsm_c_ndca}
\alias{lsm_c_ndca.RasterBrick}
\alias{lsm_c_ndca}
\alias{lsm_c_ndca.list}
\title{Number of core areas (class level)}
\usage{
lsm_c_ndca(landscape)

\method{lsm_c_ndca}{RasterLayer}(landscape)

\method{lsm_c_ndca}{RasterStack}(landscape)

\method{lsm_c_ndca}{RasterBrick}(landscape)

\method{lsm_c_ndca}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of core areas (class level)
}
\details{
Number of core areas equals the sum of number of core areas of class i.
Called number of disjunct core areas in FRAGSTATS. A core area is a 'patch within
the patch' without any edge cells. In other words, the number of patches within
the patch that only have neighbouring cells of the same type
\deqn{DCORE = sum(NCORE[patch_i])}
\subsection{Units}{None}
\subsection{Range}{DCAD >= 0}
\subsection{Behaviour}{DCAD = 0 when CORE = 0, i.e. every cell in patches of class i is
an edge. DCAD increases without limit as core areas become more present, i.e. patches
becoming larger and less complex}
}
\examples{
lsm_c_ndca(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
