% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_para_cv.R
\name{lsm_c_para_cv}
\alias{lsm_c_para_cv}
\alias{lsm_c_para_cv}
\alias{lsm_c_para_cv.RasterLayer}
\alias{lsm_c_para_cv}
\alias{lsm_c_para_cv.RasterStack}
\alias{lsm_c_para_cv}
\alias{lsm_c_para_cv.RasterBrick}
\alias{lsm_c_para_cv}
\alias{lsm_c_para_cv.list}
\title{Perimeter-area ratio distribution (class level)}
\usage{
lsm_c_para_cv(landscape)

\method{lsm_c_para_cv}{RasterLayer}(landscape)

\method{lsm_c_para_cv}{RasterStack}(landscape)

\method{lsm_c_para_cv}{RasterBrick}(landscape)

\method{lsm_c_para_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{details}{Coeffiecent of variation of the perimeter-area ratio. PARA equals the ration of
patch perimeter and patch area. It is a simple measure of complexity
\deqn{PARA_CV = cv(PARA[patch_i]}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}}
}
\value{
tibble
}
\description{
Coeffiecent of variation of perimeter-area ratio (class level)
}
\examples{
lsm_c_para_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
