% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pd.R
\name{lsm_c_pd}
\alias{lsm_c_pd}
\alias{lsm_c_pd}
\alias{lsm_c_pd.RasterLayer}
\alias{lsm_c_pd}
\alias{lsm_c_pd.RasterStack}
\alias{lsm_c_pd}
\alias{lsm_c_pd.RasterBrick}
\alias{lsm_c_pd}
\alias{lsm_c_pd.list}
\title{Patch density (class level)}
\usage{
lsm_c_pd(landscape)

\method{lsm_c_pd}{RasterLayer}(landscape)

\method{lsm_c_pd}{RasterStack}(landscape)

\method{lsm_c_pd}{RasterBrick}(landscape)

\method{lsm_c_pd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Patch density (class level)
}
\examples{
lsm_c_pd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
