% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pland.R
\name{lsm_c_pland}
\alias{lsm_c_pland}
\alias{lsm_c_pland}
\alias{lsm_c_pland.RasterLayer}
\alias{lsm_c_pland}
\alias{lsm_c_pland.RasterStack}
\alias{lsm_c_pland}
\alias{lsm_c_pland.RasterBrick}
\alias{lsm_c_pland}
\alias{lsm_c_pland.list}
\title{Percentage of landscape (class level)}
\usage{
lsm_c_pland(landscape)

\method{lsm_c_pland}{RasterLayer}(landscape)

\method{lsm_c_pland}{RasterStack}(landscape)

\method{lsm_c_pland}{RasterBrick}(landscape)

\method{lsm_c_pland}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Percentage of landscape of class (class level)
}
\details{
Percentage of landscape equals the sum of the patch area of class i divided
by the total area. In other words, PLAND is the percentage of the landscape
belonging to class i. It is a measure of compostion and because of the relative
character directly comparable among landscapes with different total areas
\deqn{PLAND = (sum(area[patch_i]) / total area) * 100}
\subsection{Units}{Percentage}
\subsection{Range}{0 < PLAND <= 100}
\subsection{Behaviour}{PLAND approaches PLAND = 0 when the class area
is decreasing. PLAND = 100 when only one class is present}
}
\examples{
lsm_c_pland(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
