% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_shape_cv.R
\name{lsm_c_shape_cv}
\alias{lsm_c_shape_cv}
\alias{lsm_c_shape_cv}
\alias{lsm_c_shape_cv.RasterLayer}
\alias{lsm_c_shape_cv}
\alias{lsm_c_shape_cv.RasterStack}
\alias{lsm_c_shape_cv}
\alias{lsm_c_shape_cv.RasterBrick}
\alias{lsm_c_shape_cv}
\alias{lsm_c_shape_cv.list}
\title{Shape index distribution  (class level)}
\usage{
lsm_c_shape_cv(landscape)

\method{lsm_c_shape_cv}{RasterLayer}(landscape)

\method{lsm_c_shape_cv}{RasterStack}(landscape)

\method{lsm_c_shape_cv}{RasterBrick}(landscape)

\method{lsm_c_shape_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation of shape index (class level)
}
\details{
Equals the coeffiecent of variation shape index of class i. SHAPE equals a quater
of the patch perimeter divided by the square root of the patch area
\deqn{SHAPE_CV = cv(SHAPE[patch_i])}
\subsection{Units}{None}
\subsection{Range}{???}
}
\examples{
lsm_c_shape_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
