% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ta.R
\name{lsm_c_ta}
\alias{lsm_c_ta}
\alias{lsm_c_ta}
\alias{lsm_c_ta.RasterLayer}
\alias{lsm_c_ta}
\alias{lsm_c_ta.RasterStack}
\alias{lsm_c_ta}
\alias{lsm_c_ta.RasterBrick}
\alias{lsm_c_ta}
\alias{lsm_c_ta.list}
\title{Total class area (class level)}
\usage{
lsm_c_ta(landscape)

\method{lsm_c_ta}{RasterLayer}(landscape)

\method{lsm_c_ta}{RasterStack}(landscape)

\method{lsm_c_ta}{RasterBrick}(landscape)

\method{lsm_c_ta}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total area of class (class level)
}
\examples{
lsm_c_ta(landscape)
lsm_c_ta(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
