% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_te.R
\name{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterLayer}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterStack}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterBrick}
\alias{lsm_c_te}
\alias{lsm_c_te.list}
\title{TE  (class level)}
\usage{
lsm_c_te(landscape, count_boundary)

\method{lsm_c_te}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_c_te}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_c_te}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_c_te}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{Include landscape boundary in edge length}
}
\value{
tibble
}
\description{
Total (class) edge (Area and Edge metric)
}
\details{
\deqn{TE = \sum_{k=1}^{m} e_{ik}}
where \eqn{e_{ik}} is the edge lengths in meters
TE is an 'Area and edge metric'. Total (class) edge includes all edges between class i and
all other classes k. It measures the configuration of the landscape because a highly
fragmentated landscape will have many edges. However, total edge is an absolute measure,
making comparisons among landscapes with different total areas difficult. If
\code{cound_boundary = TRUE} also edges to the landscape boundary are included.

\subsection{Units}{Meters}
\subsection{Range}{TE >= 0}
\subsection{Behaviour}{Equals TE = 0 if all cells are edge cells. Increases, without limit,
as landscape becomes more fragmentated}
}
\examples{
lsm_c_te(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_perim}}
\code{\link{lsm_l_te}}
}
