% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_calculate.R
\name{lsm_calculate}
\alias{lsm_calculate}
\alias{lsm_calculate}
\alias{lsm_calculate.RasterLayer}
\alias{lsm_calculate}
\alias{lsm_calculate.RasterStack}
\alias{lsm_calculate}
\alias{lsm_calculate.RasterBrick}
\alias{lsm_calculate}
\alias{lsm_calculate.list}
\title{Calculate metrics}
\usage{
lsm_calculate(landscape, what, ...)

\method{lsm_calculate}{RasterLayer}(landscape, what = "all", ...)

\method{lsm_calculate}{RasterStack}(landscape, what = "all",
  classes_max = NULL, ...)

\method{lsm_calculate}{RasterBrick}(landscape, what = "all",
  classes_max = NULL, ...)

\method{lsm_calculate}{list}(landscape, what = "all", classes_max = NULL,
  ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{what}{Selected level of metrics: either "all", "patch", "class", "landscape".
The default is "all". It is also possible to specifiy functions as a vector of strings,
e.g. \code{what = c("lsm_c_ca", "lsm_l_ta")}.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Calculate a selected group of metrics
}
\examples{
\dontrun{
lsm_calculate(landscape)
lsm_calculate(landscape, what = "patch")
}

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
