% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_aggr.R
\name{lsm_l_aggr}
\alias{lsm_l_aggr}
\alias{lsm_l_aggr}
\alias{lsm_l_aggr.RasterLayer}
\alias{lsm_l_aggr}
\alias{lsm_l_aggr.RasterStack}
\alias{lsm_l_aggr}
\alias{lsm_l_aggr.RasterBrick}
\alias{lsm_l_aggr}
\alias{lsm_l_aggr.list}
\title{Aggregation (landscape level)}
\usage{
lsm_l_aggr(landscape, directions, ordered, base)

\method{lsm_l_aggr}{RasterLayer}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_aggr}{RasterStack}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_aggr}{RasterBrick}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_aggr}{list}(landscape, directions = 4, ordered = TRUE,
  base = "log2")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Aggregation
}
\details{
Details.
}
\examples{
lsm_l_aggr(landscape)

}
\references{
Nowosad J., TF Stepinski. 2018. Information-theoretical approach to measure
landscape complexity. DOI:
}
\seealso{
\code{\link{lsm_l_comp}},
\code{\link{lsm_l_conf}},
\code{\link{lsm_l_cplx}},
}
