% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ai.R
\name{lsm_l_ai}
\alias{lsm_l_ai}
\title{AI (landscape level)}
\usage{
lsm_l_ai(landscape, directions = 8)
}
\arguments{
\item{landscape}{A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Aggregation index (Aggregation metric)
}
\details{
\deqn{AI = \Bigg[\sum\limits_{i=1}^m \Big( \frac{g_{ii}}{max-g_{ii}} \Big) P_{i} \Bigg](100) }

where \eqn{g_{ii}} is the number of like adjacencies based on the single-count method and
\eqn{max-g_{ii}} is the classwise maximum number of like adjacencies of class i and \eqn{P_{i}}
the proportion of landscape compromised of class i.

AI is an 'Aggregation metric'. It equals the number of like adjacencies divided
by the theoretical maximum possible number of like adjacencies for that class summed over each
class for the entire landscape.
The metric is based on he adjacency matrix and the single-count method.

\subsection{Units}{Percent}
\subsection{Range}{0 <= AI <= 100}
\subsection{Behaviour}{Equals 0 for maximally disaggregated and 100
for maximally aggregated classes.}
}
\examples{
landscape <- terra::rast(landscapemetrics::landscape)
lsm_l_ai(landscape)

}
\references{
McGarigal K., SA Cushman, and E Ene. 2023. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical Maps. Computer software program produced by the authors;
available at the following web site: https://www.fragstats.org

He, H. S., DeZonia, B. E., & Mladenoff, D. J. 2000. An aggregation index (AI)
to quantify spatial patterns of landscapes. Landscape ecology, 15(7), 591-601.
}
\seealso{
\code{\link{lsm_c_ai}}
}
