% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_area_cv.R
\name{lsm_l_area_cv}
\alias{lsm_l_area_cv}
\alias{lsm_l_area_cv}
\alias{lsm_l_area_cv.RasterLayer}
\alias{lsm_l_area_cv}
\alias{lsm_l_area_cv.RasterStack}
\alias{lsm_l_area_cv}
\alias{lsm_l_area_cv.RasterBrick}
\alias{lsm_l_area_cv}
\alias{lsm_l_area_cv.list}
\title{Patch area distribution (landscape level)}
\usage{
lsm_l_area_cv(landscape)

\method{lsm_l_area_cv}{RasterLayer}(landscape)

\method{lsm_l_area_cv}{RasterStack}(landscape)

\method{lsm_l_area_cv}{RasterBrick}(landscape)

\method{lsm_l_area_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coefficient of variation patch size (landscape level)
}
\examples{
lsm_l_area_cv(landscape)
lsm_l_area_cv(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
