% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_area_mn.R
\name{lsm_l_area_mn}
\alias{lsm_l_area_mn}
\alias{lsm_l_area_mn}
\alias{lsm_l_area_mn.RasterLayer}
\alias{lsm_l_area_mn}
\alias{lsm_l_area_mn.RasterStack}
\alias{lsm_l_area_mn}
\alias{lsm_l_area_mn.RasterBrick}
\alias{lsm_l_area_mn}
\alias{lsm_l_area_mn.list}
\title{Patch area distribution (landscape level)}
\usage{
lsm_l_area_mn(landscape)

\method{lsm_l_area_mn}{RasterLayer}(landscape)

\method{lsm_l_area_mn}{RasterStack}(landscape)

\method{lsm_l_area_mn}{RasterBrick}(landscape)

\method{lsm_l_area_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean patch size (landscape level)
}
\examples{
lsm_l_area_mn(landscape)
lsm_l_area_mn(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
