% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_cai_sd.R
\name{lsm_l_cai_sd}
\alias{lsm_l_cai_sd}
\alias{lsm_l_cai_sd}
\alias{lsm_l_cai_sd.RasterLayer}
\alias{lsm_l_cai_sd}
\alias{lsm_l_cai_sd.RasterStack}
\alias{lsm_l_cai_sd}
\alias{lsm_l_cai_sd.RasterBrick}
\alias{lsm_l_cai_sd}
\alias{lsm_l_cai_sd.list}
\title{Core area index distribution (landscape level)}
\usage{
lsm_l_cai_sd(landscape)

\method{lsm_l_cai_sd}{RasterLayer}(landscape)

\method{lsm_l_cai_sd}{RasterStack}(landscape)

\method{lsm_l_cai_sd}{RasterBrick}(landscape)

\method{lsm_l_cai_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of core area index (landscape level)
}
\details{
Equals the standard deviation of the core area index of all patches in the landscape.
The core area index equals the percentage of a patch that is core area
\deqn{CAI_SD = sd(CAI[patch]}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_cai_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
