% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_circle_cv.R
\name{lsm_l_circle_cv}
\alias{lsm_l_circle_cv}
\alias{lsm_l_circle_cv.RasterLayer}
\alias{lsm_l_circle_cv.RasterStack}
\alias{lsm_l_circle_cv.RasterBrick}
\alias{lsm_l_circle_cv.list}
\title{CIRCLE_CV (landscape level)}
\usage{
lsm_l_circle_cv(landscape, directions)

\method{lsm_l_circle_cv}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_circle_cv}{RasterStack}(landscape, directions = 8)

\method{lsm_l_circle_cv}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_circle_cv}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Coeffiecient of variation of related circumscribing circle (Shape metric)
}
\details{
\deqn{CIRCLE_{CV} = cv(CIRCLE[patch_{ij}])}
where \eqn{CIRCLE[patch_{ij}]} is the related circumscribing circle of each patch.

CIRCLE_CV is a 'Shape metric' and summarises the landscape as the coeffiecent of variation
of the related circumscribing circle of all patches in the landscape. CIRCLE describes
the ratio between the patch area and the smallest circumscribing circle of the patch
and characterises the compactness of the patch. CIRCLE_CV describes the differences among
all patches in the landscape. Because it is scaled to the mean, it is easily comparable.

\subsection{Units}{None}
\subsection{Range}{CIRCLE_CV >= 0}
\subsection{Behaviour}{Equals CIRCLE_CV if the related circumscribing circle is identical
for all patches. Increases, without limit, as the variation of related circumscribing
circles increases.}
}
\examples{
lsm_l_circle_cv(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_circle}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_circle_mn}},
\code{\link{lsm_c_circle_sd}},
\code{\link{lsm_c_circle_cv}}, \cr
\code{\link{lsm_l_circle_mn}},
\code{\link{lsm_l_circle_sd}}
}
