% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_comp.R
\name{lsm_l_comp}
\alias{lsm_l_comp}
\alias{lsm_l_comp}
\alias{lsm_l_comp.RasterLayer}
\alias{lsm_l_comp}
\alias{lsm_l_comp.RasterStack}
\alias{lsm_l_comp}
\alias{lsm_l_comp.RasterBrick}
\alias{lsm_l_comp}
\alias{lsm_l_comp.list}
\title{Composition (landscape level)}
\usage{
lsm_l_comp(landscape, base)

\method{lsm_l_comp}{RasterLayer}(landscape, base = "log2")

\method{lsm_l_comp}{RasterStack}(landscape, base = "log2")

\method{lsm_l_comp}{RasterBrick}(landscape, base = "log2")

\method{lsm_l_comp}{list}(landscape, base = "log2")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Composition
}
\details{
Details.
}
\examples{
lsm_l_comp(landscape)

}
\references{
Nowosad J., TF Stepinski. 2018. Information-theoretical approach to measure
landscape complexity. DOI:
}
\seealso{
\code{\link{lsm_l_conf}},
\code{\link{lsm_l_aggr}},
\code{\link{lsm_l_cplx}},
}
