% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_conf.R
\name{lsm_l_conf}
\alias{lsm_l_conf}
\alias{lsm_l_conf}
\alias{lsm_l_conf.RasterLayer}
\alias{lsm_l_conf}
\alias{lsm_l_conf.RasterStack}
\alias{lsm_l_conf}
\alias{lsm_l_conf.RasterBrick}
\alias{lsm_l_conf}
\alias{lsm_l_conf.list}
\title{Configuration (landscape level)}
\usage{
lsm_l_conf(landscape, directions, ordered, base)

\method{lsm_l_conf}{RasterLayer}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_conf}{RasterStack}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_conf}{RasterBrick}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_conf}{list}(landscape, directions = 4, ordered = TRUE,
  base = "log2")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Configuration
}
\details{
Details.
}
\examples{
lsm_l_conf(landscape)

}
\references{
Nowosad J., TF Stepinski. 2018. Information-theoretical approach to measure
landscape complexity. DOI:
}
\seealso{
\code{\link{lsm_l_comp}},
\code{\link{lsm_l_aggr}},
\code{\link{lsm_l_cplx}},
}
