% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_contag.R
\name{lsm_l_contag}
\alias{lsm_l_contag}
\alias{lsm_l_contag}
\alias{lsm_l_contag.RasterLayer}
\alias{lsm_l_contag}
\alias{lsm_l_contag.RasterStack}
\alias{lsm_l_contag}
\alias{lsm_l_contag.RasterBrick}
\alias{lsm_l_contag}
\alias{lsm_l_contag.list}
\title{CONTAG (landscape level)}
\usage{
lsm_l_contag(landscape)

\method{lsm_l_contag}{RasterLayer}(landscape)

\method{lsm_l_contag}{RasterStack}(landscape)

\method{lsm_l_contag}{RasterBrick}(landscape)

\method{lsm_l_contag}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Contagion (Contagion and Interspersion metrics)
}
\details{
\deqn{CONTAG = 1 + \frac{\sum \limits_{q = 1}^{n_{a}} p_{q} ln(p_{q})}{2ln(t)}}

where \eqn{p_{q}} the adjacency table for all classes divided by the sum of that table and
\eqn{t} the number of classes in the landscape.

CONTAG is an 'Aggregation metric'. It is based on cell adjacencies and describes
the probability of two random cells belonging to the same class. \eqn{p_{q}} is
the cell adjacency table, where the order is preserved and pairs of adjacent cells
are counted twice. Contagion is affected by both the dispersion and interspersion
of classes. E.g., low class dispersion (= high proportion of like adjacencies) and
low interspersion (= uneven distribution of pairwise adjacencies) lead to a high
contagion value.

The number of classes to calculate CONTAG must be >= than 2.

\subsection{Units}{Percent}
\subsection{Range}{0 < Contag <=100}
\subsection{Behaviour}{Approaches CONTAG = 0 if all cells are unevenly distributed
and 100 indicates that all cells are equally adjacent to all other classes.}
}
\examples{
lsm_l_contag(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Riitters, K.H., O’Neill, R.V., Wickham, J.D. & Jones, K.B. (1996). A note on contagion indices for landscape analysis. Landscape ecology, 11, 197–202.
}
\seealso{
\code{\link{lsm_p_perim}}
\code{\link{lsm_l_contag}}
}
