% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_contig_sd.R
\name{lsm_l_contig_sd}
\alias{lsm_l_contig_sd}
\alias{lsm_l_contig_sd}
\alias{lsm_l_contig_sd.RasterLayer}
\alias{lsm_l_contig_sd}
\alias{lsm_l_contig_sd.RasterStack}
\alias{lsm_l_contig_sd}
\alias{lsm_l_contig_sd.RasterBrick}
\alias{lsm_l_contig_sd}
\alias{lsm_l_contig_sd.list}
\title{CONTIG_SD (landscape level)}
\usage{
lsm_l_contig_sd(landscape)

\method{lsm_l_contig_sd}{RasterLayer}(landscape)

\method{lsm_l_contig_sd}{RasterStack}(landscape)

\method{lsm_l_contig_sd}{RasterBrick}(landscape)

\method{lsm_l_contig_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of Contiguity index (Shape metric)
}
\details{
\deqn{CONTIG_{SD} =  sd(CONTIG[patch_{ij}])}

where \eqn{CONTIG[patch_{ij}]} is the contiguity of each patch.

CONTIG_SD is a 'Shape metric'. It summarises the landscape as the standard deviation of all patches
in the landscape. CONTIG_SD asses the spatial connectedness (contiguity) of
cells in patches. The metric coerces patch values to a value of 1 and the background
to NA. A nine cell focal filter matrix:\preformatted{filter_matrix <- matrix(c(1, 2, 1,
                          2, 1, 2,
                          1, 2, 1), 3, 3, byrow = T)
}

... is then used to weight orthogonally contiguous pixels more heavily than
diagonally contiguous pixels. Therefore, larger and more connections between
patch cells in the rookie case result in larger contiguity index values.

\subsection{Units}{None}
\subsection{Range}{CONTIG_SD >= 0}
\subsection{Behaviour}{CONTIG_SD = 0 if the contiguity index is
identical for all patches. Increases, without limit, as the variation of
CONTIG increases.}
}
\examples{
lsm_l_contig_sd(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_contig}},
\code{\link{lsm_c_contig_sd}},
\code{\link{lsm_c_contig_cv}},
\code{\link{lsm_c_contig_mn}}, \cr
\code{\link{lsm_l_contig_cv}},
\code{\link{lsm_l_contig_mn}}
}
