% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_core.R
\name{lsm_l_core}
\alias{lsm_l_core}
\alias{lsm_l_core}
\alias{lsm_l_core.RasterLayer}
\alias{lsm_l_core}
\alias{lsm_l_core.RasterStack}
\alias{lsm_l_core}
\alias{lsm_l_core.RasterBrick}
\alias{lsm_l_core}
\alias{lsm_l_core.list}
\title{Total area of core areas (landscape level)}
\usage{
lsm_l_core(landscape, directions)

\method{lsm_l_core}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_core}{RasterStack}(landscape, directions = 8)

\method{lsm_l_core}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_core}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{???}
}
\value{
tibble
}
\description{
Area of corea area (landscape level)
}
\details{
Equals the area within a patch that is not on the edge of all patches in the landscape.
In other words, the area of a patch that has only neighbouring cells of the same type
of class i
\subsection{Units}{Hectares}
\subsection{Range}{CORE >= 0}
\subsection{Behaviour}{CORE increases without limit as patch areas increase
and patch shapes simplify. CORE = 0 when every cell in every patch of class i
is an edge}
}
\examples{
lsm_l_core(landscape, 8)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
