% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_core_cv.R
\name{lsm_l_core_cv}
\alias{lsm_l_core_cv}
\alias{lsm_l_core_cv}
\alias{lsm_l_core_cv.RasterLayer}
\alias{lsm_l_core_cv}
\alias{lsm_l_core_cv.RasterStack}
\alias{lsm_l_core_cv}
\alias{lsm_l_core_cv.RasterBrick}
\alias{lsm_l_core_cv}
\alias{lsm_l_core_cv.list}
\title{Core area distribution (landscape level)}
\usage{
lsm_l_core_cv(landscape, ...)

\method{lsm_l_core_cv}{RasterLayer}(landscape, ...)

\method{lsm_l_core_cv}{RasterStack}(landscape, ...)

\method{lsm_l_core_cv}{RasterBrick}(landscape, ...)

\method{lsm_l_core_cv}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Coeffiecent of variation of patch core area (landscape level)
}
\details{
Equals the coeffiecent of variation of the patch core area of all patches in the landscape.
The core area is the area within a patch that is not on the edge of the patch of class i.
In other words, the area of a patch that has only neighbouring cells of the same type
of class i
\deqn{CORE_CV = cv(core[patch])}
\subsection{Units}{Hectares}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_core_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
