% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_core_mn.R
\name{lsm_l_core_mn}
\alias{lsm_l_core_mn}
\alias{lsm_l_core_mn}
\alias{lsm_l_core_mn.RasterLayer}
\alias{lsm_l_core_mn}
\alias{lsm_l_core_mn.RasterStack}
\alias{lsm_l_core_mn}
\alias{lsm_l_core_mn.RasterBrick}
\alias{lsm_l_core_mn}
\alias{lsm_l_core_mn.list}
\title{Core area distribution (landscape level)}
\usage{
lsm_l_core_mn(landscape)

\method{lsm_l_core_mn}{RasterLayer}(landscape)

\method{lsm_l_core_mn}{RasterStack}(landscape)

\method{lsm_l_core_mn}{RasterBrick}(landscape)

\method{lsm_l_core_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of patch core area (landscape level)
}
\details{
Equals the mean of the patch core area of all patches in the landscape.
The core area is the area within a patch that is not on the edge of the patch of class i.
In other words, the area of a patch that has only neighbouring cells of the same type
of class i
\deqn{CORE_MN = mean(core[patch])}
\subsection{Units}{Hectares}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_core_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
