% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_cplx.R
\name{lsm_l_cplx}
\alias{lsm_l_cplx}
\alias{lsm_l_cplx}
\alias{lsm_l_cplx.RasterLayer}
\alias{lsm_l_cplx}
\alias{lsm_l_cplx.RasterStack}
\alias{lsm_l_cplx}
\alias{lsm_l_cplx.RasterBrick}
\alias{lsm_l_cplx}
\alias{lsm_l_cplx.list}
\title{Complexity (landscape level)}
\usage{
lsm_l_cplx(landscape, directions, ordered, base)

\method{lsm_l_cplx}{RasterLayer}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_cplx}{RasterStack}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_cplx}{RasterBrick}(landscape, directions = 4, ordered = TRUE,
  base = "log2")

\method{lsm_l_cplx}{list}(landscape, directions = 4, ordered = TRUE,
  base = "log2")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Complexity
}
\details{
Details.
}
\examples{
lsm_l_cplx(landscape)

}
\references{
Nowosad J., TF Stepinski. 2018. Information-theoretical approach to measure
landscape complexity. DOI:
}
\seealso{
\code{\link{lsm_l_comp}},
\code{\link{lsm_l_conf}},
\code{\link{lsm_l_aggr}}
}
