% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_dcad.R
\name{lsm_l_dcad}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.RasterLayer}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.RasterStack}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.RasterBrick}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.list}
\title{Disjunct core area density (landscape level)}
\usage{
lsm_l_dcad(landscape, ...)

\method{lsm_l_dcad}{RasterLayer}(landscape, ...)

\method{lsm_l_dcad}{RasterStack}(landscape, ...)

\method{lsm_l_dcad}{RasterBrick}(landscape, ...)

\method{lsm_l_dcad}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Disjunct core area density (landscape level)
}
\details{
Disjunct core area density equals the sum of number of core areas of the
landscape divided by the total area. In other words, it is the number of core areas
relative to the total area. The measure is relative and therefore comparable among
landscapes with different total areas
\deqn{DCAD = sum(ncore[patch]) / total area}
\subsection{Units}{Number per hectares (FRAGSTATS has per 100 ha)}
\subsection{Range}{DCAD >= 0}
\subsection{Behaviour}{DCAD = 0 when CORE = 0, i.e. every cell in patches
of class i is an edge. DCAD increases without limit as core areas become more
present, i.e. patches becoming larger and less complex}
}
\examples{
lsm_l_dcad(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
