% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_dcore_cv.R
\name{lsm_l_dcore_cv}
\alias{lsm_l_dcore_cv}
\alias{lsm_l_dcore_cv}
\alias{lsm_l_dcore_cv.RasterLayer}
\alias{lsm_l_dcore_cv}
\alias{lsm_l_dcore_cv.RasterStack}
\alias{lsm_l_dcore_cv}
\alias{lsm_l_dcore_cv.RasterBrick}
\alias{lsm_l_dcore_cv}
\alias{lsm_l_dcore_cv.list}
\title{Number of core areas distribution (landscape level)}
\usage{
lsm_l_dcore_cv(landscape)

\method{lsm_l_dcore_cv}{RasterLayer}(landscape)

\method{lsm_l_dcore_cv}{RasterStack}(landscape)

\method{lsm_l_dcore_cv}{RasterBrick}(landscape)

\method{lsm_l_dcore_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation of number of core areas (landscape level)
}
\details{
Equals the coeffiecent of variation of number of core area of all patches in the landscape.
A core area is a 'patch within the patch' without any edge cells. In other words,
the number of patches within the patch that only have neighbouring cells of the same type
\deqn{dcore_CV = cv(dcore[patch])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_dcore_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
