% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_dcore_mn.R
\name{lsm_l_dcore_mn}
\alias{lsm_l_dcore_mn}
\alias{lsm_l_dcore_mn.RasterLayer}
\alias{lsm_l_dcore_mn.RasterStack}
\alias{lsm_l_dcore_mn.RasterBrick}
\alias{lsm_l_dcore_mn.stars}
\alias{lsm_l_dcore_mn.list}
\title{DCORE_MN (landscape level)}
\usage{
lsm_l_dcore_mn(landscape, directions)

\method{lsm_l_dcore_mn}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_dcore_mn}{RasterStack}(landscape, directions = 8)

\method{lsm_l_dcore_mn}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_dcore_mn}{stars}(landscape, directions = 8)

\method{lsm_l_dcore_mn}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Mean number of disjunct core areas (Core area metric)
}
\details{
\deqn{DCORE_{MN} = mean(NCORE[patch_{ij}])}
where \eqn{NCORE[patch_{ij}]} is the number of core areas.

DCORE_MN is an 'Core area metric'. It summarises the landscape as the mean of all
patches in the landscape. A cell is defined as core if the cell has no neighbour with
a different value than itself (rook's case). NCORE counts the disjunct core areas,
whereby a core area is a 'patch within the patch' containing only core cells.

\subsection{Units}{None}
\subsection{Range}{DCORE_MN > 0}
\subsection{Behaviour}{Equals DCORE_MN = 0 if NCORE = 0 for all patches. Increases,
without limit, as the number of disjunct core areas increases.}
}
\examples{
lsm_l_dcore_mn(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_ncore}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_dcore_mn}},
\code{\link{lsm_c_dcore_sd}},
\code{\link{lsm_c_dcore_cv}}, \cr
\code{\link{lsm_l_dcore_sd}},
\code{\link{lsm_l_dcore_cv}}
}
