% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ed.R
\name{lsm_l_ed}
\alias{lsm_l_ed}
\alias{lsm_l_ed}
\alias{lsm_l_ed.RasterLayer}
\alias{lsm_l_ed}
\alias{lsm_l_ed.RasterStack}
\alias{lsm_l_ed}
\alias{lsm_l_ed.RasterBrick}
\alias{lsm_l_ed}
\alias{lsm_l_ed.list}
\title{Edge density (landscape level)}
\usage{
lsm_l_ed(landscape, count_boundary)

\method{lsm_l_ed}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_l_ed}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_l_ed}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_l_ed}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{Include landscape boundary in edge length}
}
\value{
tibble
}
\description{
Edge density (landscape level)
}
\details{
The edge density equals the sum of all edges in the landscape divided by the landscape area.
ED is scaled to 1 hectare and therefore compareable among landscapes with different total areas
\deqn{ED = total edge / total area}
\subsection{Units}{Meters per hectar}
\subsection{Range}{ED >= 0}
\subsection{Behaviour}{ED = 0 when only one class and patch is present and increases
without limit as edges increase, i.e. number of patches and classes}
}
\examples{
lsm_l_ed(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
