% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_frac_cv.R
\name{lsm_l_frac_cv}
\alias{lsm_l_frac_cv}
\alias{lsm_l_frac_cv}
\alias{lsm_l_frac_cv.RasterLayer}
\alias{lsm_l_frac_cv}
\alias{lsm_l_frac_cv.RasterStack}
\alias{lsm_l_frac_cv}
\alias{lsm_l_frac_cv.RasterBrick}
\alias{lsm_l_frac_cv}
\alias{lsm_l_frac_cv.list}
\title{Fractal dimension index distribution (landscape level)}
\usage{
lsm_l_frac_cv(landscape)

\method{lsm_l_frac_cv}{RasterLayer}(landscape)

\method{lsm_l_frac_cv}{RasterStack}(landscape)

\method{lsm_l_frac_cv}{RasterBrick}(landscape)

\method{lsm_l_frac_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation fractal dimension index (landscape level)
}
\details{
Coeffiecent of variation of the fractal dimension index of all patches in the landscape.
FRAC equals two times the natural logarithm of a quarter of the patch area divided by the
natural logarithm of the patch area
\deqn{FRAC_CV = cv(FRAC[patch])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_frac_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
