% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_frac_sd.R
\name{lsm_l_frac_sd}
\alias{lsm_l_frac_sd}
\alias{lsm_l_frac_sd}
\alias{lsm_l_frac_sd.RasterLayer}
\alias{lsm_l_frac_sd}
\alias{lsm_l_frac_sd.RasterStack}
\alias{lsm_l_frac_sd}
\alias{lsm_l_frac_sd.RasterBrick}
\alias{lsm_l_frac_sd}
\alias{lsm_l_frac_sd.list}
\title{Fractal dimension index distribution (landscape level)}
\usage{
lsm_l_frac_sd(landscape)

\method{lsm_l_frac_sd}{RasterLayer}(landscape)

\method{lsm_l_frac_sd}{RasterStack}(landscape)

\method{lsm_l_frac_sd}{RasterBrick}(landscape)

\method{lsm_l_frac_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation fractal dimension index (landscape level)
}
\details{
Standard deviation of the fractal dimension index of all patches in the landscape.
FRAC equals two times the natural logarithm of a quarter of the patch area divided by the
natural logarithm of the patch area
\deqn{FRAC_SD = sd(FRAC[patch])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_frac_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
