% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_gyrate_cv.R
\name{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.RasterLayer}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.RasterStack}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.RasterBrick}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.list}
\title{Radius of Gyration (GYRATE)}
\usage{
lsm_l_gyrate_cv(landscape)

\method{lsm_l_gyrate_cv}{RasterLayer}(landscape)

\method{lsm_l_gyrate_cv}{RasterStack}(landscape)

\method{lsm_l_gyrate_cv}{RasterBrick}(landscape)

\method{lsm_l_gyrate_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Radius of Gyration (landscape)
}
\details{
Equals the mean distance of each cell centroid in a patch to the centroid of the whole patch (mean location of all cell centroids).
\subsection{Units}{Meters}
\subsection{Range}{GYRATE >= 0, without limit}
\subsection{Behaviour}{0 if single cell, maximum if patch occupies the entire landscape.}
}
\examples{
lsm_l_gyrate_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
