% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_gyrate_mn.R
\name{lsm_l_gyrate_mn}
\alias{lsm_l_gyrate_mn}
\title{GYRATE_MN (landscape level)}
\usage{
lsm_l_gyrate_mn(landscape, directions = 8, cell_center = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}

\item{cell_center}{If true, the coordinates of the centroid are forced to be
a cell center within the patch.}
}
\value{
tibble
}
\description{
Mean radius of gyration (Area and edge metric)
}
\details{
\deqn{GYRATE_{MN} = mean(GYRATE[patch_{ij}])}
where \eqn{GYRATE[patch_{ij}]} equals the radius of gyration of each patch.

GYRATE_MN is an 'Area and edge metric'. The metric summarises the landscape
as the mean of the radius of gyration of all patches in the landscape.
GYRATE measures the distance from each cell to the patch centroid and is based on
cell center-to-cell center distances. The metrics characterises
both the patch area and compactness.

If \code{cell_center = TRUE} some patches might have several possible cell-center
centroids. In this case, the gyrate index is based on the mean distance of all
cells to all possible cell-center centroids.

\subsection{Units}{Meters}
\subsection{Range}{GYRATE_MN >= 0 }
\subsection{Behaviour}{Approaches GYRATE_MN = 0 if every patch is a single cell. Increases,
without limit, when only one patch is present.}
}
\examples{
lsm_l_gyrate_mn(terra::unwrap(landscape))

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: https://www.umass.edu/landeco/

Keitt, T. H., Urban, D. L., & Milne, B. T. 1997. Detecting critical scales
in fragmented landscapes. Conservation ecology, 1(1).
}
\seealso{
\code{\link{lsm_p_gyrate}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_gyrate_mn}},
\code{\link{lsm_c_gyrate_sd}},
\code{\link{lsm_c_gyrate_cv}}, \cr
\code{\link{lsm_l_gyrate_sd}},
\code{\link{lsm_l_gyrate_cv}}
}
