% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_iji.R
\name{lsm_l_iji}
\alias{lsm_l_iji}
\alias{lsm_l_iji}
\alias{lsm_l_iji.RasterLayer}
\alias{lsm_l_iji}
\alias{lsm_l_iji.RasterStack}
\alias{lsm_l_iji}
\alias{lsm_l_iji.RasterBrick}
\alias{lsm_l_iji}
\alias{lsm_l_iji.list}
\title{Interspersion and Juxtaposition index (landscape level)}
\usage{
lsm_l_iji(landscape)

\method{lsm_l_iji}{RasterLayer}(landscape)

\method{lsm_l_iji}{RasterStack}(landscape)

\method{lsm_l_iji}{RasterBrick}(landscape)

\method{lsm_l_iji}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Interspersion and Juxtaposition index (landscape level)
}
\details{
???
\deqn{???}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_iji(landscape)
}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
