% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_msidi.R
\name{lsm_l_msidi}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterLayer}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterStack}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterBrick}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.list}
\title{Modified Simpson's diversity index (landscape level)}
\usage{
lsm_l_msidi(landscape)

\method{lsm_l_msidi}{RasterLayer}(landscape)

\method{lsm_l_msidi}{RasterStack}(landscape)

\method{lsm_l_msidi}{RasterBrick}(landscape)

\method{lsm_l_msidi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Modified Simpson's diversity index (landscape level)
}
\examples{
lsm_l_msidi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
