% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_msiei.R
\name{lsm_l_msiei}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterLayer}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterStack}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterBrick}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.list}
\title{Modified Simpson's evenness index (landscape level)}
\usage{
lsm_l_msiei(landscape)

\method{lsm_l_msiei}{RasterLayer}(landscape)

\method{lsm_l_msiei}{RasterStack}(landscape)

\method{lsm_l_msiei}{RasterBrick}(landscape)

\method{lsm_l_msiei}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Modified Simpson's evenness index (landscape level)
}
\details{
The modified Simpson's evenness index equals the negative natural logarithm of
the sum of the squared proportions of all patches in the landscape grouped by class divided
by the natural logarithm of patch richness
\deqn{MSIDI = -ln(sum(proportion[patch_i] ^ 2)) / ln(patch richness)}
\subsection{Units}{None}
\subsection{Range}{0 <= MSIEI < 1}
\subsection{Behaviour}{MSIEI = 0 when only one class and patch is present and approaches
MSIEI = 1 as the proportional distribution of patches becomes more even}
}
\examples{
lsm_l_msiei(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
