% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_mutinf.R
\name{lsm_l_mutinf}
\alias{lsm_l_mutinf}
\alias{lsm_l_mutinf}
\alias{lsm_l_mutinf.RasterLayer}
\alias{lsm_l_mutinf}
\alias{lsm_l_mutinf.RasterStack}
\alias{lsm_l_mutinf}
\alias{lsm_l_mutinf.RasterBrick}
\alias{lsm_l_mutinf}
\alias{lsm_l_mutinf.list}
\title{Mutual information (landscape level)}
\usage{
lsm_l_mutinf(landscape, directions = 4, ordered = TRUE, base = "log2")

\method{lsm_l_mutinf}{RasterLayer}(landscape, directions = 4,
  ordered = TRUE, base = "log2")

\method{lsm_l_mutinf}{RasterStack}(landscape, directions = 4,
  ordered = TRUE, base = "log2")

\method{lsm_l_mutinf}{RasterBrick}(landscape, directions = 4,
  ordered = TRUE, base = "log2")

\method{lsm_l_mutinf}{list}(landscape, directions = 4, ordered = TRUE,
  base = "log2")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which cells should be connected:
4 (rook's case) or 8 (queen's case).
The default is 4.}

\item{ordered}{The type of pairs considered.
Either ordered (TRUE) or unordered (FALSE).
The default is TRUE.}

\item{base}{The unit in which entropy is measured.
The default is "log2", which compute entropy in "bits".
"log" and "log10" can be also used.}
}
\value{
tibble
}
\description{
It disambiguates landscape pattern types characterize
by the same value of an overall complexity (\code{\link{lsm_l_joinent}}).
}
\examples{
lsm_l_mutinf(landscape)

}
\references{
Nowosad J., TF Stepinski. 2018. Information-theoretical approach to measure
landscape complexity. DOI:
}
\seealso{
\code{\link{lsm_l_ent}},
\code{\link{lsm_l_condent}},
\code{\link{lsm_l_joinent}},
}
