% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ncore.R
\name{lsm_l_ncore}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.RasterLayer}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.RasterStack}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.RasterBrick}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.list}
\title{Number of core areas (landscape level)}
\usage{
lsm_l_ncore(landscape)

\method{lsm_l_ncore}{RasterLayer}(landscape)

\method{lsm_l_ncore}{RasterStack}(landscape)

\method{lsm_l_ncore}{RasterBrick}(landscape)

\method{lsm_l_ncore}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of disjunct core areas (landscape level)
}
\examples{
lsm_l_ncore(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
