% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ncore_mn.R
\name{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.RasterLayer}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.RasterStack}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.RasterBrick}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.list}
\title{Distribution number of core areas (landscape level)}
\usage{
lsm_l_ncore_mn(landscape)

\method{lsm_l_ncore_mn}{RasterLayer}(landscape)

\method{lsm_l_ncore_mn}{RasterStack}(landscape)

\method{lsm_l_ncore_mn}{RasterBrick}(landscape)

\method{lsm_l_ncore_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean (landscape level)
}
\examples{
lsm_l_ncore_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
