% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ndca.R
\name{lsm_l_ndca}
\alias{lsm_l_ndca}
\alias{lsm_l_ndca}
\alias{lsm_l_ndca.RasterLayer}
\alias{lsm_l_ndca}
\alias{lsm_l_ndca.RasterStack}
\alias{lsm_l_ndca}
\alias{lsm_l_ndca.RasterBrick}
\alias{lsm_l_ndca}
\alias{lsm_l_ndca.list}
\title{Number of core areas (landscape level)}
\usage{
lsm_l_ndca(landscape)

\method{lsm_l_ndca}{RasterLayer}(landscape)

\method{lsm_l_ndca}{RasterStack}(landscape)

\method{lsm_l_ndca}{RasterBrick}(landscape)

\method{lsm_l_ndca}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of core areas (landscape level)
}
\details{
Number of core areas equals the sum of number of core areas of all patches in the
landscape. Called number of disjunct core areas in FRAGSTATS. A core area is a
'patch within the patch' without any edge cells. In other words, the number of
patches within the patch that only have neighbouring cells of the same type
\deqn{ndca = sum(ndca[patch])}
\subsection{Units}{None}
\subsection{Range}{ndca >= 0}
\subsection{Behaviour}{ndca = 0 when CORE = 0, i.e. every cell in patches of class i is
an edge. ndca increases without limit as core areas become more present, i.e. patches
becoming larger and less complex}
}
\examples{
lsm_l_ndca(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
