% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_np.R
\name{lsm_l_np}
\alias{lsm_l_np}
\title{Number of patches (class level)}
\usage{
lsm_l_np(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of patches (class level)
}
\examples{
lsm_np(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
