% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_para_cv.R
\name{lsm_l_para_cv}
\alias{lsm_l_para_cv}
\alias{lsm_l_para_cv}
\alias{lsm_l_para_cv.RasterLayer}
\alias{lsm_l_para_cv}
\alias{lsm_l_para_cv.RasterStack}
\alias{lsm_l_para_cv}
\alias{lsm_l_para_cv.RasterBrick}
\alias{lsm_l_para_cv}
\alias{lsm_l_para_cv.list}
\title{Perimeter area ratio  (patch level)}
\usage{
lsm_l_para_cv(landscape)

\method{lsm_l_para_cv}{RasterLayer}(landscape)

\method{lsm_l_para_cv}{RasterStack}(landscape)

\method{lsm_l_para_cv}{RasterBrick}(landscape)

\method{lsm_l_para_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Ratio of patch perimeter and area
}
\examples{
lsm_l_para_cv(landscape)
lsm_l_para_cv(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
