% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_pd.R
\name{lsm_l_pd}
\alias{lsm_l_pd}
\alias{lsm_l_pd}
\alias{lsm_l_pd.RasterLayer}
\alias{lsm_l_pd}
\alias{lsm_l_pd.RasterStack}
\alias{lsm_l_pd}
\alias{lsm_l_pd.RasterBrick}
\alias{lsm_l_pd}
\alias{lsm_l_pd.list}
\title{Patch density (landscape level)}
\usage{
lsm_l_pd(landscape)

\method{lsm_l_pd}{RasterLayer}(landscape)

\method{lsm_l_pd}{RasterStack}(landscape)

\method{lsm_l_pd}{RasterBrick}(landscape)

\method{lsm_l_pd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Patch density (landscape level)
}
\details{
Patch density equals the number of all patches in the landscape divided by the total
area. Patch density is a relative measure and compareable among landscapes with
different total areas
\deqn{PD = number of patches / total area}
\subsection{Units}{Number per hectares (Number per 100 ha in FRAGSTATS???)}
\subsection{Ranges}{PD > 0}
\subsection{Behaviour}{Increases as the landscape gets more patchy}
}
\examples{
lsm_l_pd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
