% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_pr.R
\name{lsm_l_pr}
\alias{lsm_l_pr}
\alias{lsm_l_pr}
\alias{lsm_l_pr.RasterLayer}
\alias{lsm_l_pr}
\alias{lsm_l_pr.RasterStack}
\alias{lsm_l_pr}
\alias{lsm_l_pr.RasterBrick}
\alias{lsm_l_pr}
\alias{lsm_l_pr.list}
\title{Patch richness (lanscape level)}
\usage{
lsm_l_pr(landscape)

\method{lsm_l_pr}{RasterLayer}(landscape)

\method{lsm_l_pr}{RasterStack}(landscape)

\method{lsm_l_pr}{RasterBrick}(landscape)

\method{lsm_l_pr}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Patch richness (lanscape level)
}
\details{
Number of classes in the landscape. This is a very simple measure of composition.
\subsection{Units}{None}
\subsection{Range}{PR >= 1}
\subsection{Behaviour}{PR = 1 when only 1 class is present and increases without
limit as the number of classes increases}
}
\examples{
lsm_l_pr(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
