% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_rpr.R
\name{lsm_l_rpr}
\alias{lsm_l_rpr}
\alias{lsm_l_rpr}
\alias{lsm_l_rpr.RasterLayer}
\alias{lsm_l_rpr}
\alias{lsm_l_rpr.RasterStack}
\alias{lsm_l_rpr}
\alias{lsm_l_rpr.RasterBrick}
\title{Relative patch richness (landscape level)}
\usage{
lsm_l_rpr(landscape, classes_max)

\method{lsm_l_rpr}{RasterLayer}(landscape, classes_max = NULL)

\method{lsm_l_rpr}{RasterStack}(landscape, classes_max = NULL)

\method{lsm_l_rpr}{RasterBrick}(landscape, classes_max = NULL)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Relative patch richness (landscape level)
}
\details{
Relative patch richness equals the number of classes divided by the (potential)
maximum number of classes. The maximum number of classes needs to be specified
\deqn{RPR = (number of classes / potential number of classes) * 100}
\subsection{Units}{Percentage}
\subsection{Ranges}{0 < RPR < 100}
\subsection{Behaviour}{RPR approaches RPR = 0 when only one class type is present,
but the maximum number of classes is large. RPR approaches 100 when the number of
present class types approaches the maximum number of classes}
}
\examples{
lsm_l_rpr(landscape, classes_max = 5)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
