% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shape_mn.R
\name{lsm_l_shape_mn}
\alias{lsm_l_shape_mn}
\alias{lsm_l_shape_mn}
\alias{lsm_l_shape_mn.RasterLayer}
\alias{lsm_l_shape_mn}
\alias{lsm_l_shape_mn.RasterStack}
\alias{lsm_l_shape_mn}
\alias{lsm_l_shape_mn.RasterBrick}
\alias{lsm_l_shape_mn}
\alias{lsm_l_shape_mn.list}
\title{Shape index distribution  (landscape. level)}
\usage{
lsm_l_shape_mn(landscape)

\method{lsm_l_shape_mn}{RasterLayer}(landscape)

\method{lsm_l_shape_mn}{RasterStack}(landscape)

\method{lsm_l_shape_mn}{RasterBrick}(landscape)

\method{lsm_l_shape_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of shape index (landscape. level)
}
\details{
Equals the mean shape index of all patches in the landscape. SHAPE equals a quater
of the patch perimeter divided by the square root of the patch area
\deqn{SHAPE_MN = mean(SHAPE[patch])}
\subsection{Units}{None}
\subsection{Range}{???}
}
\examples{
lsm_l_shape_mn(landscape)
lsm_l_shape_mn(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
