% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shape_sd.R
\name{lsm_l_shape_sd}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.RasterLayer}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.RasterStack}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.RasterBrick}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.list}
\title{Shape index distribution  (landscape level)}
\usage{
lsm_l_shape_sd(landscape)

\method{lsm_l_shape_sd}{RasterLayer}(landscape)

\method{lsm_l_shape_sd}{RasterStack}(landscape)

\method{lsm_l_shape_sd}{RasterBrick}(landscape)

\method{lsm_l_shape_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of shape index (landscape level)
}
\details{
Equals the standard deviation of the shape index of all patches in the landscape.
SHAPE equals a quater of the patch perimeter divided by the square root of the patch area
\deqn{SHAPE_SD = sd(SHAPE[patch])}
\subsection{Units}{None}
\subsection{Range}{???}
}
\examples{
lsm_l_shape_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
