% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shdi.R
\name{lsm_l_shdi}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.RasterLayer}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.RasterStack}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.RasterBrick}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.list}
\title{Shannon's Diversity Index}
\usage{
lsm_l_shdi(landscape)

\method{lsm_l_shdi}{RasterLayer}(landscape)

\method{lsm_l_shdi}{RasterStack}(landscape)

\method{lsm_l_shdi}{RasterBrick}(landscape)

\method{lsm_l_shdi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Diversity of patch types (classes) in the landscape
}
\examples{
lsm_l_shdi(landscape)
lsm_l_shdi(landscape_stack)
lsm_l_shdi(raster::as.list(landscape_stack))

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
