% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shei.R
\name{lsm_l_shei}
\alias{lsm_l_shei}
\alias{lsm_l_shei}
\alias{lsm_l_shei.RasterLayer}
\alias{lsm_l_shei}
\alias{lsm_l_shei.RasterStack}
\alias{lsm_l_shei}
\alias{lsm_l_shei.RasterBrick}
\alias{lsm_l_shei}
\alias{lsm_l_shei.list}
\title{Shannons's evenness index (landscape level)}
\usage{
lsm_l_shei(landscape)

\method{lsm_l_shei}{RasterLayer}(landscape)

\method{lsm_l_shei}{RasterStack}(landscape)

\method{lsm_l_shei}{RasterBrick}(landscape)

\method{lsm_l_shei}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Shannons's evenness index (landscape level)
}
\details{
Shannon's evenness index equals minus the sum of class proportions multipled by the natural
logarithm of class proportions divided by the natural logarithm of class types
\deqn{SHEI = sum(proportion[patch_i] * log(proportion[patch_i])) / log(number of classes)}
\subsection{Units}{None}
\subsection{Range}{0 <= SHEI < 1}
\subsection{Behaviour}{Shannons's evenness index approaches SHEI = 0 when only one patch and
class is present and approaches SHEI = 1 when the number of class types increases while the
proportions are equally distributed}
}
\examples{
lsm_l_shei(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
