% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_siei.R
\name{lsm_l_siei}
\alias{lsm_l_siei}
\alias{lsm_l_siei}
\alias{lsm_l_siei.RasterLayer}
\alias{lsm_l_siei}
\alias{lsm_l_siei.RasterStack}
\alias{lsm_l_siei}
\alias{lsm_l_siei.RasterBrick}
\alias{lsm_l_siei}
\alias{lsm_l_siei.list}
\title{Simpson's evenness index (landscape level)}
\usage{
lsm_l_siei(landscape)

\method{lsm_l_siei}{RasterLayer}(landscape)

\method{lsm_l_siei}{RasterStack}(landscape)

\method{lsm_l_siei}{RasterBrick}(landscape)

\method{lsm_l_siei}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Simpson's evenness index (landscape level)
}
\details{
Simpson's evenness index equals 1 minus the sum of squared class proportions
divided by 1 minus 1 divided by the number of classes
\deqn{SIEI = sum(proportion[patch_i] ^ 2) / 1 - (1 / number of classes)}
\subsection{Units}{None}
\subsection{Range}{0 < SIEI <= 1}
\subsection{Behaviour}{Simpson's evenness index approaches SIEI = 0 when only one patch and
class is present and approaches SIDI = 1 when the number of class types increases while the
proportions are equally distributed}
}
\examples{
lsm_l_siei(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
