% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_split.R
\name{lsm_l_split}
\alias{lsm_l_split}
\alias{lsm_l_split}
\alias{lsm_l_split.RasterLayer}
\alias{lsm_l_split}
\alias{lsm_l_split.RasterStack}
\alias{lsm_l_split}
\alias{lsm_l_split.RasterBrick}
\alias{lsm_l_split}
\alias{lsm_l_split.list}
\title{Splitting index (landscape level)}
\usage{
lsm_l_split(landscape)

\method{lsm_l_split}{RasterLayer}(landscape)

\method{lsm_l_split}{RasterStack}(landscape)

\method{lsm_l_split}{RasterBrick}(landscape)

\method{lsm_l_split}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Splitting index (landscape level)
}
\examples{
lsm_l_split(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
