% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_tca.R
\name{lsm_l_tca}
\alias{lsm_l_tca}
\alias{lsm_l_tca}
\alias{lsm_l_tca.RasterLayer}
\alias{lsm_l_tca}
\alias{lsm_l_tca.RasterStack}
\alias{lsm_l_tca}
\alias{lsm_l_tca.RasterBrick}
\alias{lsm_l_tca}
\alias{lsm_l_tca.list}
\title{Total area of core areas (landscape level)}
\usage{
lsm_l_tca(landscape)

\method{lsm_l_tca}{RasterLayer}(landscape)

\method{lsm_l_tca}{RasterStack}(landscape)

\method{lsm_l_tca}{RasterBrick}(landscape)

\method{lsm_l_tca}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Area of corea area (landscape level)
}
\details{
Equals the area within a patch that is not on the edge of all patches in the landscape.
In other words, the area of a patch that has only neighbouring cells of the same type
of class i
\subsection{Units}{Hectares}
\subsection{Range}{CORE >= 0}
\subsection{Behaviour}{CORE increases without limit as patch areas increase
and patch shapes simplify. CORE = 0 when every cell in every patch of class i
is an edge}
}
\examples{
lsm_l_tca(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
