% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_te.R
\name{lsm_l_te}
\alias{lsm_l_te}
\alias{lsm_l_te}
\alias{lsm_l_te.RasterLayer}
\alias{lsm_l_te}
\alias{lsm_l_te.RasterStack}
\alias{lsm_l_te}
\alias{lsm_l_te.RasterBrick}
\alias{lsm_l_te}
\alias{lsm_l_te.list}
\title{Total Edge (TE, landscape scale)}
\usage{
lsm_l_te(landscape, count_boundary)

\method{lsm_l_te}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_l_te}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_l_te}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_l_te}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{Include landscape boundary in edge length}
}
\value{
tibble
}
\description{
Total Edge
}
\examples{
lsm_l_te(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
