% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_cai.R
\name{lsm_p_cai}
\alias{lsm_p_cai}
\alias{lsm_p_cai}
\alias{lsm_p_cai.RasterLayer}
\alias{lsm_p_cai}
\alias{lsm_p_cai.RasterStack}
\alias{lsm_p_cai}
\alias{lsm_p_cai.RasterBrick}
\alias{lsm_p_cai}
\alias{lsm_p_cai.list}
\title{Core area index  (patch level)}
\usage{
lsm_p_cai(landscape, directions)

\method{lsm_p_cai}{RasterLayer}(landscape, directions = 8)

\method{lsm_p_cai}{RasterStack}(landscape, directions = 8)

\method{lsm_p_cai}{RasterBrick}(landscape, directions = 8)

\method{lsm_p_cai}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{???}
}
\value{
tibble
}
\description{
Core area index of patch (patch level)
}
\details{
The core area index equals the percentage of a patch that is core area
\deqn{CAI = (core[patch] / area[patch]) * 100}
\subsection{Units}{Percentage}
\subsection{Range}{0 <= CAI <= 100}
\subsection{Behaviour}{CAI = 0 when the patch has no core area and
approaches CAI = 100 with increasing percentage of core area within a patch}
}
\examples{
lsm_p_cai(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
