% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_circle.R
\name{lsm_p_circle}
\alias{lsm_p_circle}
\alias{lsm_p_circle}
\alias{lsm_p_circle.RasterLayer}
\alias{lsm_p_circle}
\alias{lsm_p_circle.RasterStack}
\alias{lsm_p_circle}
\alias{lsm_p_circle.RasterBrick}
\alias{lsm_p_circle}
\alias{lsm_p_circle.list}
\title{Related Circumscribing Circle (CIRCLE)}
\usage{
lsm_p_circle(landscape)

\method{lsm_p_circle}{RasterLayer}(landscape)

\method{lsm_p_circle}{RasterStack}(landscape)

\method{lsm_p_circle}{RasterBrick}(landscape)

\method{lsm_p_circle}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Related Circumscribing Circle (patch)
}
\details{
Equals the 1 - the patch area (m^2) divided by the area (m^2) of the smallest circumscribing circle.
\subsection{Units}{None}
\subsection{Range}{0 <= CIRCLE < 1}
}
\examples{
lsm_p_circle(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
