% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_core.R
\name{lsm_p_core}
\alias{lsm_p_core}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterLayer}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterStack}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterBrick}
\alias{lsm_p_core}
\alias{lsm_p_core.list}
\title{CORE (patch level)}
\usage{
lsm_p_core(landscape)

\method{lsm_p_core}{RasterLayer}(landscape)

\method{lsm_p_core}{RasterStack}(landscape)

\method{lsm_p_core}{RasterBrick}(landscape)

\method{lsm_p_core}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Core area (Core area metric)
}
\details{
\deqn{CORE = a_{ij}^{core}}
where \eqn{a_{ij}^{core}} is the core area in square meters

CORE is a 'Core area metric' and equals the area within a patch that is not
on the edge of it. A cell is defined as core area if the cell has no
neighbour with a different value than itself (rook's case). It describes patch area
and shape simultaneously (more core area when the patch is large and the shape is
rather compact, i.e. a square).

\subsection{Units}{Hectares}
\subsection{Range}{CORE >= 0}
\subsection{Behaviour}{Increases, without limit, as the patch area increases
and the patch shape simplifies (more core area). CORE = 0 when every cell in
the patch is an edge.}
}
\examples{
lsm_p_core(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_c_core_mn}},
\code{\link{lsm_c_core_sd}},
\code{\link{lsm_c_core_cv}},
\code{\link{lsm_c_tca}}, \cr
\code{\link{lsm_l_core_mn}},
\code{\link{lsm_l_core_sd}},
\code{\link{lsm_l_core_cv}},
\code{\link{lsm_l_tca}}
}
