% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_frac.R
\name{lsm_p_frac}
\alias{lsm_p_frac}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterLayer}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterStack}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterBrick}
\alias{lsm_p_frac}
\alias{lsm_p_frac.list}
\title{FRAC (patch level)}
\usage{
lsm_p_frac(landscape)

\method{lsm_p_frac}{RasterLayer}(landscape)

\method{lsm_p_frac}{RasterStack}(landscape)

\method{lsm_p_frac}{RasterBrick}(landscape)

\method{lsm_p_frac}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Fractacl dimension index (Shape metric)
}
\details{
\deqn{FRAC = \frac{2 * \ln * (0.25 * p_{ij})}{\ln a_{ij}}}
where \eqn{p_{ij}} is the perimeter in meters and \eqn{a_{ij}} is the
area in square meters

FRAC is a 'Shape metric'. The index is based on the patch perimeter and
the patch area and describes the patch complexity. Because it is standardized,
it is independent of the patch area.

\subsection{Units}{None}
\subsection{Range}{1 <= FRAC <= 2 }
\subsection{Behaviour}{Approaches FRAC = 1 a squared patch shape form and FRAC = 2
for a irregular patch shape.}
}
\examples{
lsm_p_frac(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{lsm_p_perim}}, \cr
\code{\link{lsm_c_frac_mn}},
\code{\link{lsm_c_frac_sd}},
\code{\link{lsm_c_frac_cv}}, \cr
\code{\link{lsm_l_frac_mn}},
\code{\link{lsm_l_frac_sd}},
\code{\link{lsm_l_frac_cv}}
}
