% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_shape.R
\name{lsm_p_shape}
\alias{lsm_p_shape}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterLayer}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterStack}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterBrick}
\alias{lsm_p_shape}
\alias{lsm_p_shape.list}
\title{Shape index  (patch level)}
\usage{
lsm_p_shape(landscape)

\method{lsm_p_shape}{RasterLayer}(landscape)

\method{lsm_p_shape}{RasterStack}(landscape)

\method{lsm_p_shape}{RasterBrick}(landscape)

\method{lsm_p_shape}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Perimeter divided by squareroot of area
}
\examples{
lsm_p_shape(landscape)
lsm_p_shape(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
