% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_pland.R
\name{lsm_pland}
\alias{lsm_pland}
\title{Percentage of landscape (class level)}
\usage{
lsm_pland(landscape, scale = "class")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Percentage of landscape (class level)
}
\examples{
library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))
lsm_pland(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
