% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_prd.R
\name{lsm_prd}
\alias{lsm_prd}
\title{Patch richness density}
\usage{
lsm_prd(landscape, scale = "landscape")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{scale}{xxx}
}
\value{
Value >= 1
}
\description{
Number of patch types (classes) in the landscape divided by the total
landscape area
}
\examples{
lsm_prd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
