% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MedPS}
\alias{MedPS}
\title{Median Patch Size}
\usage{
MedPS(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the patch size in the median in hectares (ha).
}
\description{
Get the fragment size found in the median in hectares (ha).
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
