% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_area.R
\name{metric_area}
\alias{metric_area}
\title{metric_area}
\usage{
metric_area(landscape, poi = NULL)
}
\arguments{
\item{landscape}{[\code{Raster* object}]}

\item{poi}{[\code{numerical}]\cr  Vector of numeric values indicating classes to be reported.}
}
\value{
List of tibbles
}
\description{
Report area for each of the constituent values making up a matrix
}
\details{
Function reports the number of cells and the proportion of total cells made up by each unique class value
}
\examples{
library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))

metric_area(landscape)

}
