% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_pr.R, R/metric_shdi.R
\name{metric_pr}
\alias{metric_pr}
\alias{metric_pr}
\title{Patch richness}
\usage{
metric_pr(landscape)

metric_pr(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1

Value >= 1
}
\description{
Number of patch types (classes) in the landscape

Number of patch types (classes) in the landscape
}
\examples{
library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))
metric_pr(landscape)

library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))
metric_pr(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
 PNW-351.

McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
 PNW-351.
}
