% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MNN}
\alias{MNN}
\title{Mean Nearest-Neighbor Distance}
\usage{
MNN(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the mean nearest neighbor distance in meters (m) between patchs.
}
\description{
Calculates the mean nearest neighbor distance in meters (m) between patchs.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
