% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MPAR}
\alias{MPAR}
\title{Mean Perimeter-Area Ratio}
\usage{
MPAR(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the mean perimeter area ratio in hectares (ha).
}
\description{
Calculates de mean perimeter area ratio in hectares (ha).
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
