% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MPFD}
\alias{MPFD}
\title{Mean Patch Fractal Dimension}
\usage{
MPFD(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the mean patch fractal dimension.
}
\description{
Calculates the mean patch fractal dimension.
}
\details{
The MPFD value varies between 1 and 2.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
