% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MSI}
\alias{MSI}
\title{Mean Shape Index}
\usage{
MSI(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the mean shape index.
}
\description{
Calculates the mean shape index.
}
\details{
The value must be greater than or equal to 1.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
