% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{NCA_100}
\alias{NCA_100}
\title{Number of Core Area with edge of 100 meters}
\usage{
NCA_100(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the total number of core area, with subtracted edge of 100 meters.
}
\description{
Calculates the total number of core area when excluded the edge from 100 meters.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
