% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{NP}
\alias{NP}
\title{Number of Patches}
\usage{
NP(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the number of patches of the respective class.
}
\description{
Get the number of patches of the corresponding patch type (class).
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
