% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding.R
\name{padding}
\alias{padding}
\alias{padding}
\alias{padding.RasterLayer}
\alias{padding}
\alias{padding.RasterStack}
\alias{padding}
\alias{padding.RasterBrick}
\alias{padding}
\alias{padding.list}
\title{Padding}
\usage{
padding(landscape, padding_value, padding_cells)

\method{padding}{RasterLayer}(landscape, padding_value = -999,
  padding_cells = 1)

\method{padding}{RasterStack}(landscape, padding_value = -999,
  padding_cells = 1)

\method{padding}{RasterBrick}(landscape, padding_value = -999,
  padding_cells = 1)

\method{padding}{list}(landscape, padding_value = -999, padding_cells = 1)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{padding_value}{value of cells added}

\item{padding_cells}{Number of rows and columns added}
}
\value{
raster
}
\description{
Adding padding to raster
}
\examples{
padding(landscape)
padding(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
