% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{podlasie_ccilc}
\alias{podlasie_ccilc}
\title{Podlasie ESA CCI LC}
\format{
A raster layer object.
}
\source{
http://maps.elie.ucl.ac.be/CCI/viewer/
}
\usage{
podlasie_ccilc
}
\description{
A real landscape of the Podlasie region in Poland from the ESA CCI Land
Cover
}
\keyword{datasets}
