% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_as_mat.R
\name{points_as_mat}
\alias{points_as_mat}
\title{points_as_mat}
\usage{
points_as_mat(pts)
}
\arguments{
\item{pts}{Point geometry as SpatVector or sf object.}
}
\value{
matrix
}
\description{
Points as matrix
}
\details{
Converts sf points to coordinates matrix
}
\keyword{internal}
