% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_info.R
\name{proj_info}
\alias{proj_info}
\title{proj_info}
\usage{
proj_info(landscape)
}
\arguments{
\item{landscape}{RasterLayer}
}
\value{
raster
}
\description{
Projection info
}
\details{
Check projection of RasterLayer
}
\examples{

augusta_nlcd <- terra::rast(augusta_nlcd)
proj_info(augusta_nlcd)
podlasie_ccilc <- terra::rast(podlasie_ccilc)
proj_info(podlasie_ccilc)
landscape <- terra::rast(landscape)
proj_info(landscape)

}
\keyword{internal}
