% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{PSCoV}
\alias{PSCoV}
\title{Patch Size Coefficient of Variation}
\usage{
PSCoV(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The value of the patch size coefficient of variation in percentage.
}
\description{
Calculates of the patch size coefficient of variation in percentage.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
