% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cell_from_xy2}
\alias{rcpp_cell_from_xy2}
\title{Get cell number}
\usage{
rcpp_cell_from_xy2(x, y)
}
\arguments{
\item{x}{A matrix}

\item{y}{A matrix with two columns (row and column numbers)}
}
\description{
Get cell number(s) of a matrix from row and column numbers.
Cell numbers start at 0 in the upper left corner,
and increase from top to bottom, and then from left to right.
}
\keyword{internal}
