% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_nearest_neighbor}
\alias{rcpp_get_nearest_neighbor}
\title{First nearest neighbor distance}
\usage{
rcpp_get_nearest_neighbor(points)
}
\arguments{
\item{points}{A two numeric matrix, where the first two columns are x and y}
}
\value{
Vector
}
\description{
Efficiently calculate the distance to the first nearest neighbor.
Quasi linear runtime.
}
\author{
Florian Privé \email{florian.prive.21@gmail.com}
}
\keyword{internal}
