% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_pairs}
\alias{rcpp_get_pairs}
\title{Adjacent cells values}
\usage{
rcpp_get_pairs(x, directions)
}
\arguments{
\item{x}{A matrix}

\item{directions}{The number of directions in which cells should be connected:
4 (rook's case), 8 (queen's case), or a neigborhood matrix.
The neigborhood matrix should have one cell with value 0 (the focal cell),
and at least one cell with value 1 (the adjacent cells).
Cells with other values (e.g. NA) are ignored.}
}
\description{
Extract values of adjacent cells.
An output is a two columns matrix, where the first column contains a
main cell value and the second column contains an adjacent cell value.
}
\keyword{internal}
