% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_xy_from_matrix}
\alias{rcpp_xy_from_matrix}
\title{Coordinates from a matrix}
\usage{
rcpp_xy_from_matrix(x)
}
\arguments{
\item{x}{A matrix}
}
\description{
These functions get coordinates (row and column numbers) of the matrix cells.
}
