% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_lsm.R
\name{scale_lsm}
\alias{scale_lsm}
\alias{scale_lsm.RasterLayer}
\alias{scale_lsm.RasterStack}
\alias{scale_lsm.RasterBrick}
\alias{scale_lsm.stars}
\alias{scale_lsm.list}
\title{scale_lsm}
\usage{
scale_lsm(landscape, y, buffer_width, max_width, verbose, progress, ...)

\method{scale_lsm}{RasterLayer}(landscape, y, buffer_width, max_width,
  verbose = TRUE, progress = FALSE, ...)

\method{scale_lsm}{RasterStack}(landscape, y, buffer_width, max_width,
  verbose = TRUE, progress = FALSE, ...)

\method{scale_lsm}{RasterBrick}(landscape, y, buffer_width, max_width,
  verbose = TRUE, progress = FALSE, ...)

\method{scale_lsm}{stars}(landscape, y, buffer_width, max_width,
  verbose = TRUE, progress = FALSE, ...)

\method{scale_lsm}{list}(landscape, y, buffer_width, max_width,
  verbose = TRUE, progress = FALSE, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{y}{2-column matrix with coordinates or SpatialPoints.}

\item{buffer_width}{Buffer width in which a landscape metric is measured}

\item{max_width}{Max distance to which buffer_width is summed up}

\item{verbose}{Print warning messages.}

\item{progress}{Print progress report.}

\item{...}{Arguments passed on to \code{calculate_lsm()}.}
}
\value{
tibble
}
\description{
Metrics on changing scale
}
\details{
This function calculates the selected metricsin subsequential buffers around
a/multiple point(s) of interest.

The size of the actual
sampled landscape can be different to the provided size due to two reasons.
Firstly, because clipping raster cells using a circle or a sample plot not directly
at a cell center lead to inaccuracies. Secondly, sample plots can exceed the
landscape boundary. Therefore, we report the actual clipped sample plot area relative
in relation to the theoretical, maximum sample plot area e.g. a sample plot only half
within the landscape will have a \code{percentage_inside = 50}. Please be aware that the
output is sligthly different to all other \code{lsm}-function of \code{landscapemetrics}.

The metrics can be specified by the arguments \code{what}, \code{level}, \code{metric}, \code{name}
and/or \code{type} (combinations of different arguments are possible (e.g.
\code{level = "class", type = "aggregation metric"}). If an argument is not provided,
automatically all possibilities are selected. Therefore, to get \strong{all}
available metrics, don't specify any of the above arguments.
}
\examples{
my_points = matrix(c(1265000, 1250000, 1255000, 1257000),
                  ncol = 2, byrow = TRUE)

scale_lsm(augusta_nlcd, my_points, buffer_width = 500, max_width = 5000,
 what = c("lsm_l_ent", "lsm_l_mutinf"))

}
\seealso{
\code{\link{list_lsm}} \cr
\code{\link{calculate_lsm}} \cr
\code{\link{sample_lsm}} \cr
\code{\link{construct_buffer}}
}
