% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{show_info}
\alias{show_info}
\alias{show_info.RasterLayer}
\alias{show_info.RasterStack}
\alias{show_info.RasterBrick}
\alias{show_info.stars}
\alias{show_info.list}
\title{Show info}
\usage{
show_info(landscape)

\method{show_info}{RasterLayer}(landscape)

\method{show_info}{RasterStack}(landscape)

\method{show_info}{RasterBrick}(landscape)

\method{show_info}{stars}(landscape)

\method{show_info}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers}
}
\value{
tibble
}
\description{
Show info
}
\details{
This function extracts basic information about
the input landscape.
It includes a type of coordinate reference system (crs) -
either "geographic", "projected", or NA,
units of the coordinate reference system,
and a class of the input landscape's values.
}
\examples{
show_info(augusta_nlcd)
show_info(podlasie_ccilc)
show_info(landscape)
landscape_stack = raster::stack(landscape, landscape)
show_info(landscape_stack)

}
