% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_lsm.R
\name{show_lsm}
\alias{show_lsm}
\alias{show_lsm.RasterLayer}
\alias{show_lsm.RasterStack}
\alias{show_lsm.RasterBrick}
\alias{show_lsm.stars}
\alias{show_lsm.list}
\title{Show landscape metrics}
\usage{
show_lsm(landscape, what, class, directions, labels, nrow, ncol)

\method{show_lsm}{RasterLayer}(landscape, what, class = "global",
  directions = 8, labels = TRUE, nrow = NULL, ncol = NULL)

\method{show_lsm}{RasterStack}(landscape, what, class = "global",
  directions = 8, labels = TRUE, nrow = NULL, ncol = NULL)

\method{show_lsm}{RasterBrick}(landscape, what, class = "global",
  directions = 8, labels = TRUE, nrow = NULL, ncol = NULL)

\method{show_lsm}{stars}(landscape, what, class = "global",
  directions = 8, labels = TRUE, nrow = NULL, ncol = NULL)

\method{show_lsm}{list}(landscape, what, class = "global",
  directions = 8, labels = TRUE, nrow = NULL, ncol = NULL)
}
\arguments{
\item{landscape}{*Raster object}

\item{what}{Patch level what to plot}

\item{class}{How to show the labeled patches: "global" (single map), "all" (every class as facet),
or a vector with the specific classes one wants to show (every selected class as facet).}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}

\item{labels}{Logical flag indicating whether to print or not to print patch labels.}

\item{nrow, ncol}{Number of rows and columns for the facet.}
}
\value{
ggplot
}
\description{
Show patches
}
\details{
The function plots all patches with a fill corresponding to the value of the chosen what
}
\examples{
show_lsm(landscape, what = "lsm_p_area")
show_lsm(landscape, what = "lsm_p_shape", class = c(1, 2), labels = FALSE)
show_lsm(landscape, what = "lsm_p_circle", class = 3, labels = TRUE)

}
