% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_patches.R
\name{show_patches}
\alias{show_patches}
\alias{show_patches}
\alias{show_patches.RasterLayer}
\alias{show_patches}
\alias{show_patches.RasterStack}
\alias{show_patches}
\alias{show_patches.RasterBrick}
\alias{show_patches}
\alias{show_patches.list}
\title{Show patches}
\usage{
show_patches(landscape)

\method{show_patches}{RasterLayer}(landscape)

\method{show_patches}{RasterStack}(landscape)

\method{show_patches}{RasterBrick}(landscape)

\method{show_patches}{list}(landscape)
}
\arguments{
\item{landscape}{*Raster object}
}
\value{
ggplot
}
\description{
Show patches
}
\details{
The functions plots the landscape with the patches labelled with the corresponding
patch id.
}
\examples{
show_patches(landscape)

}
