% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{TE}
\alias{TE}
\title{Total Edge}
\usage{
TE(x)
}
\arguments{
\item{x}{Raster of data with patches identified as classes in RasterLayer format.}
}
\value{
The sum of the lengths, in meters (m), of all edge segments involving the corresponding patch type.
}
\description{
Calculates the lengths, in meters (m), of all edge segm ents involving the corresponding patch type.
}
\references{
MCGARIGAL. K.; MARKS, B. J. Fragstats: Spatial pattern analysis program for quantifying landscape structure. Reference manual. Forest Science Department Oregon State University. Corvallis Oregon 1994. 60 p.
}
