% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cclabel.R
\name{cclabel}
\alias{cclabel}
\title{Connected compoment labeling for classes in a raster}
\usage{
cclabel(landscape, what)
}
\arguments{
\item{landscape}{xxx}

\item{what}{Either "all" for every class in the raster, or specify class value. See Details.}
}
\value{
RasterLayer/RasterStack
}
\description{
xxx
}
\details{
\code{what} returns a RasterStack if the argument \code{"all"} is used. Each layer
represents a unique class found in the raster and contains the labelled patches.
}
\examples{
cclabeled_raster <-  cclabel(landscape, 0.5)
raster::plot(cclabeled_raster)

# count patches
length(raster::unique(cclabeled_raster))
}
