% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cclabel.R
\name{cclabel}
\alias{cclabel}
\alias{cclabel}
\alias{cclabel.RasterLayer}
\alias{cclabel}
\alias{cclabel.RasterStack}
\alias{cclabel}
\alias{cclabel.RasterBrick}
\alias{cclabel}
\alias{cclabel.list}
\title{Connected compoment labeling for classes in a raster}
\usage{
cclabel(landscape, what = "all")

\method{cclabel}{RasterLayer}(landscape, what = "all")

\method{cclabel}{RasterStack}(landscape, what = "all")

\method{cclabel}{RasterBrick}(landscape, what = "all")

\method{cclabel}{list}(landscape, what = "all")
}
\arguments{
\item{landscape}{xxx}

\item{what}{Either "all" (default) for every class in the raster, or specify
class value. See Details.}
}
\value{
List with RasterLayer/RasterBrick
}
\description{
xxx
}
\details{
\code{what} returns a RasterBrick if the argument \code{"all"} is used. Each layer
represents a unique class found in the raster and contains the labelled patches.
}
\examples{
# check for patches of class 1
cclabeled_raster  <-  cclabel(landscape, 1)

# count patches
length(raster::unique(cclabeled_raster[[1]]))

# check for patches of every class
cclabeled_raster <-  cclabel(landscape)

#label a rasterstack
cclabeled_rasterstack <-  cclabel(landscape_stack)

}
