% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cclabel.R
\name{cclabel}
\alias{cclabel}
\title{Connected compoment labeling for classes in a raster}
\usage{
cclabel(landscape, what = "all")
}
\arguments{
\item{landscape}{xxx}

\item{what}{Either "all" (default) for every class in the raster, or specify
class value. See Details.}
}
\value{
RasterLayer/RasterStack
}
\description{
xxx
}
\details{
\code{what} returns a RasterBrick if the argument \code{"all"} is used. Each layer
represents a unique class found in the raster and contains the labelled patches.
}
\examples{
# check for patches of class 1
cclabeled_raster <-  cclabel(landscape, 1)
raster::plot(cclabeled_raster)

# count patches
length(raster::unique(cclabeled_raster))

# check for patches of every class
cclabeled_raster <-  cclabel(landscape)
raster::plot(cclabeled_raster)

#label a rasterstack
cclabeled_rasterstack <-  cclabel(landscape_stack, 1)
raster::plot(cclabeled_rasterstack)

}
