% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{data_info}
\alias{data_info}
\title{data_info}
\usage{
data_info(landscape)
}
\arguments{
\item{landscape}{RasterLayer}
}
\value{
raster
}
\description{
Adding padding to raster
}
\details{
Check class of raster values and number of classes
}
\examples{
data_info(landscape)
data_info(augusta_nlcd)
data_info(podlasie_ccilc)

}
\keyword{internal}
