% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{data_info}
\alias{data_info}
\title{data_info}
\usage{
data_info(landscape)
}
\arguments{
\item{landscape}{RasterLayer}
}
\value{
raster
}
\description{
Data info
}
\details{
Check class of raster values and number of classes
}
\examples{
data_info(terra::unwrap(landscape))
data_info(terra::unwrap(augusta_nlcd))
data_info(terra::unwrap(podlasie_ccilc))

}
\keyword{internal}
