% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{get_boundaries}
\usage{
get_boundaries(landscape, rooks_case = TRUE, as_NA = FALSE,
  return_raster = TRUE)
}
\arguments{
\item{landscape}{RasterLayer or matrix.}

\item{rooks_case}{Logical if rook's case (4 neighbours) or queen's
case (8 neighbours) should be used as neighbourhood rule}

\item{as_NA}{If true, non-boundary cells area labeld NA}

\item{return_raster}{If false, matrix is returned}
}
\value{
RasterLayer or matrix
}
\description{
Get boundary cells of patches
}
\details{
All boundary/edge cells are labeled 1, all non-boundary cells 0. NA values are
not changed. Boundary cells are defined as cells that neighbour either a NA
cell or a cell with a different value than itself. Non-boundary cells only
neighbour cells with the same value than themself.
}
\examples{
class_1 <- get_patches(landscape, class = 1)[[1]]

get_boundaries(class_1)
get_boundaries(class_1, return_raster = FALSE)

class_1_matrix <- raster::as.matrix(class_1)
get_boundaries(class_1_matrix, return_raster = FALSE)

}
