% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsf_pr.R
\name{lsf_pr}
\alias{lsf_pr}
\alias{landscape_pr}
\title{Patch richness}
\usage{
lsf_pr(landscape, scale = "landscape")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{scale}{xxx}
}
\value{
Value >= 1
}
\description{
Number of patch types (classes) in the landscape
}
\examples{
lsf_pr(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
