% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsf_pland.R, R/lsf_pr.R, R/lsf_ta.R
\name{lsf_pland}
\alias{lsf_pland}
\alias{landscape_pr}
\alias{lsf_pr}
\alias{landscape_pr}
\alias{lsf_ta}
\alias{landscape_pr}
\title{Percentage of landscape (class level)}
\usage{
lsf_pland(landscape, scale = "class")

lsf_pr(landscape, scale = "landscape")

lsf_ta(landscape, scale = "class")
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{scale}{xxx}

\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble

Value >= 1

tibble
}
\description{
Percentage of landscape (class level)

Number of patch types (classes) in the landscape

Total area of class (class level)
}
\examples{
library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))
landscape_pr(landscape)

lsf_pr(landscape)

library(NLMR)
landscape <- nlm_randomcluster(ncol = 30, nrow = 30,
                               p = 0.4, ai = c(0.25, 0.25, 0.5))
landscape_pr(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.

McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.

McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
