% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape_stack}
\alias{landscape_stack}
\title{Example maps as raster stack (random cluster neutral landscape model).}
\format{A raster stack object.}
\source{
Simulated neutral landscape model with R. https://github.com/ropensci/NLMR/
}
\usage{
landscape_stack
}
\description{
An example rasterstack to show landscapetools functionality
generated with \code{nlm_mpd()} and \code{nlm_random()}.
}
\keyword{datasets}
