% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapemetrics-package.R
\docType{package}
\name{landscapemetrics}
\alias{landscapemetrics}
\alias{landscapemetrics-package}
\alias{landscapemetrics-package}
\title{landscapemetrics}
\description{
Calculates landscape metrics for categorical landscape patterns in
a tidy workflow. 'landscapemetrics' reimplements the most common metrics from
FRAGSTATS and new ones from the current literature on landscape metrics.
This package supports raster spatial objects and takes
RasterLayer, RasterStacks, RasterBricks or lists of RasterLayer from the
'raster' package as input arguments. It further provides utility functions
to visualize patches, select metrics and building blocks to develop new
metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://landscapeecology.github.io/landscapemetrics/}
  \item Report bugs at \url{https://github.com/landscapeecology/landscapemetrics/issues}
}

}
\author{
\strong{Maintainer}: Maximillian, H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

Authors:
\itemize{
  \item Marco Sciaini \email{sciaini.marco@gmail.com} (0000-0002-3042-5435)
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)
}

Other contributors:
\itemize{
  \item Laura J. Graham (0000-0002-1057-3721) [contributor]
  \item Kimberly A. With (0000-0001-5570-1515) [contributor]
  \item Jeffrey Hollister (0000-0002-3611-7281) [contributor]
  \item Jeremy VanDerWal [contributor]
  \item Florian Privé [contributor]
}

}
