% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapemetrics-package.R
\docType{package}
\name{landscapemetrics}
\alias{landscapemetrics}
\alias{landscapemetrics-package}
\alias{landscapemetrics-package}
\title{landscapemetrics}
\description{
xxx
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/landscapemetrics}
  \item Report bugs at \url{https://github.com/marcosci/landscapemetrics/issues}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{sciaini.marco@gmail.com} (0000-0002-3042-5435)

Authors:
\itemize{
  \item Maximillian Hesselbarth
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)
  \item Laura Graham
  \item Jeffrey Hollister
  \item Kimberly A, With \email{kwith@ksu.edu} (0000-0001-5570-1515)
}

}
