% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapemetrics-package.R
\docType{package}
\name{landscapemetrics}
\alias{landscapemetrics}
\alias{landscapemetrics-package}
\title{landscapemetrics}
\description{
Calculates landscape metrics for categorical landscape patterns in
a tidy workflow. 'landscapemetrics' reimplements the most common metrics from
FRAGSTATS and new ones from the current literature on landscape metrics.
This package supports raster spatial objects and takes
RasterLayer, RasterStacks, RasterBricks or lists of RasterLayer from the
'raster' package as input arguments. It further provides utility functions
to visualize patches, select metrics and building blocks to develop new
metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatialecology.github.io/landscapemetrics/}
  \item Report bugs at \url{https://github.com/r-spatialecology/landscapemetrics/issues}
}

}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Authors:
\itemize{
  \item Marco Sciaini \email{sciaini.marco@gmail.com} (\href{https://orcid.org/0000-0002-3042-5435}{ORCID})
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
  \item Sebastian Hanss (\href{https://orcid.org/0000-0002-3990-4897}{ORCID})
}

Other contributors:
\itemize{
  \item Laura J. Graham (Input on package structure) [contributor]
  \item Jeffrey Hollister (Input on package structure) [contributor]
  \item Kimberly A. With (Input on package structure) [contributor]
  \item Florian Privé (Original author of underlying C++ code for get_nearestneighbour() function) [contributor]
  \item Project Nayuki (Original author of underlying C++ code for get_circumscribingcircle and lsm_p_circle) [contributor]
  \item Matt Strimas-Mackey (Bugfix in sample_metrics()) [contributor]
}

}
