% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsm_abbreviations_names}
\alias{lsm_abbreviations_names}
\title{Tibble of abbreviations coming from FRAGSTATS}
\format{A tibble object.}
\usage{
lsm_abbreviations_names
}
\description{
A single tibble for every abbreviation of every metric that is
reimplemented in landscapemetrics and its corrosponding full name
in the literature.
}
\details{
Can be used ofter calculating the metric(s) with a join to have
a more readable results tibble or for visualizing your results.
}
\examples{
aggregation_index <- lsm_p_area(landscape)
aggregation_index <- dplyr::right_join(aggregation_index,
                                       lsm_abbreviations_names,
                                       by = 'metric')
aggregation_index


}
\keyword{datasets}
