% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_area_cv.R
\name{lsm_c_area_cv}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.RasterLayer}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.RasterStack}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.RasterBrick}
\alias{lsm_c_area_cv}
\alias{lsm_c_area_cv.list}
\title{Patch area distribution (class level)}
\usage{
lsm_c_area_cv(landscape)

\method{lsm_c_area_cv}{RasterLayer}(landscape)

\method{lsm_c_area_cv}{RasterStack}(landscape)

\method{lsm_c_area_cv}{RasterBrick}(landscape)

\method{lsm_c_area_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation of patch area (class level)
}
\details{
Equals the coeffiecent of variation of the patch area of class i
\deqn{AREA_CV = cv(AREA[patch_i])}
\subsection{Units}{Hectares}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_area_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
