% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_area_mn.R
\name{lsm_c_area_mn}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.RasterLayer}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.RasterStack}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.RasterBrick}
\alias{lsm_c_area_mn}
\alias{lsm_c_area_mn.list}
\title{Patch area distribution (class level)}
\usage{
lsm_c_area_mn(landscape)

\method{lsm_c_area_mn}{RasterLayer}(landscape)

\method{lsm_c_area_mn}{RasterStack}(landscape)

\method{lsm_c_area_mn}{RasterBrick}(landscape)

\method{lsm_c_area_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean patch area (class level)
}
\details{
Equals the mean patch area of class i.
\deqn{AREA_MN = mean(area[patch_i])}
\subsection{Units}{Square meter (assuming that the input cellsize is in meter)}
\subsection{Range}{AREA_MN > 0 and increases without limit as the amount of the
class increases. AREA_MN = TA if only one class is present}
}
\examples{
lsm_c_area_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
