% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_area_sd.R
\name{lsm_c_area_sd}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.RasterLayer}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.RasterStack}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.RasterBrick}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.list}
\title{AREA_SD (class level)}
\usage{
lsm_c_area_sd(landscape)

\method{lsm_c_area_sd}{RasterLayer}(landscape)

\method{lsm_c_area_sd}{RasterStack}(landscape)

\method{lsm_c_area_sd}{RasterBrick}(landscape)

\method{lsm_c_area_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers}
}
\value{
tibble
}
\description{
Standard deviation of patch area (Area and edge metric)
}
\details{
\deqn{AREA_{SD} = sd(AREA[patch_{ij}])}
where \eqn{AREA[patch_{ij}]} is the area of each patch in hectares.

AREA_SD is an 'Area and Edge metric'. The metric summarises each class
as the standard deviation of all patch areas belonging to class i. The metric describes the differences among patches of the same class i in
the landscape.

\subsection{Units}{Hectares}
\subsection{Range}{AREA_SD >= 0}
\subsection{Behaviour}{Equals AREA_SD = 0 if all patches are identical in size.
Increases, without limit, as the variation of patch areas increases.}
}
\examples{
lsm_c_area_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{sd}}, \cr
\code{\link{lsm_c_area_mn}},
\code{\link{lsm_c_area_cv}}, \cr
\code{\link{lsm_l_area_mn}},
\code{\link{lsm_l_area_sd}},
\code{\link{lsm_l_area_cv}}
}
