% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_area_sd.R
\name{lsm_c_area_sd}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.RasterLayer}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.RasterStack}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.RasterBrick}
\alias{lsm_c_area_sd}
\alias{lsm_c_area_sd.list}
\title{Patch area distribution (class level)}
\usage{
lsm_c_area_sd(landscape)

\method{lsm_c_area_sd}{RasterLayer}(landscape)

\method{lsm_c_area_sd}{RasterStack}(landscape)

\method{lsm_c_area_sd}{RasterBrick}(landscape)

\method{lsm_c_area_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of patch area (class level)
}
\details{
Equals the standard deviation of the patch area of class i
\deqn{AREA_SD = sd(AREA[patch_i])}
\subsection{Units}{Hectares}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_area_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
