% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ca.R
\name{lsm_c_ca}
\alias{lsm_c_ca}
\alias{lsm_c_ca}
\alias{lsm_c_ca.RasterLayer}
\alias{lsm_c_ca}
\alias{lsm_c_ca.RasterStack}
\alias{lsm_c_ca}
\alias{lsm_c_ca.RasterBrick}
\alias{lsm_c_ca}
\alias{lsm_c_ca.list}
\title{Total class area (class level)}
\usage{
lsm_c_ca(landscape)

\method{lsm_c_ca}{RasterLayer}(landscape)

\method{lsm_c_ca}{RasterStack}(landscape)

\method{lsm_c_ca}{RasterBrick}(landscape)

\method{lsm_c_ca}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total area of class (class level)
}
\details{
Total class area equals the sum of the area of all patches of class i. Total class area is
an absolute measure, making comparisons among landscapes with different
total areas difficult. It is a measure of landscape composition.
\deqn{CA = sum(area[patch_i])}
\subsection{Units}{Hectares}
\subsection{Range}{CA > 0}
\subsection{Behaviour}{CA increases without limit as the amount of the class increases.
CA = TA if only one class is present}
}
\examples{
lsm_c_ca(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
