% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cai_cv.R
\name{lsm_c_cai_cv}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.RasterLayer}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.RasterStack}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.RasterBrick}
\alias{lsm_c_cai_cv}
\alias{lsm_c_cai_cv.list}
\title{Core area index distribution (class level)}
\usage{
lsm_c_cai_cv(landscape)

\method{lsm_c_cai_cv}{RasterLayer}(landscape)

\method{lsm_c_cai_cv}{RasterStack}(landscape)

\method{lsm_c_cai_cv}{RasterBrick}(landscape)

\method{lsm_c_cai_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coefficient of variation of core area index (class level)
}
\details{
Equals the coeffiecent of variation of the core area index of class i.
The core area index equals the percentage of a patch that is core area
\deqn{CAI_CV = cv(CAI[patch_i]}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_cai_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
