% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cai_mn.R
\name{lsm_c_cai_mn}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterLayer}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterStack}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterBrick}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.list}
\title{Core area index distribution (class level)}
\usage{
lsm_c_cai_mn(landscape)

\method{lsm_c_cai_mn}{RasterLayer}(landscape)

\method{lsm_c_cai_mn}{RasterStack}(landscape)

\method{lsm_c_cai_mn}{RasterBrick}(landscape)

\method{lsm_c_cai_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.

#' Equals the mean of the core area index of class i.
The core area index equals the percentage of a patch that is core area
\deqn{CAI_MN = mean(CAI[patch_i]}
\subsection{Units}{Percentage}
\subsection{Range}{???}
\subsection{Behaviour}{???}}
}
\value{
tibble
}
\description{
Mean of core area index (class level)
}
\examples{
lsm_c_cai_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
