% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cai_mn.R
\name{lsm_c_cai_mn}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterLayer}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterStack}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.RasterBrick}
\alias{lsm_c_cai_mn}
\alias{lsm_c_cai_mn.list}
\title{CAI_MN (class level)}
\usage{
lsm_c_cai_mn(landscape)

\method{lsm_c_cai_mn}{RasterLayer}(landscape)

\method{lsm_c_cai_mn}{RasterStack}(landscape)

\method{lsm_c_cai_mn}{RasterBrick}(landscape)

\method{lsm_c_cai_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of core area index (Core area metric)
}
\details{
\deqn{CAI_{MN} = mean(CAI[patch_{ij}]}
where \eqn{CAI[patch_{ij}]} is the core area index of each patch.

CAI_MN is a 'Core area metric'. The metric summarises each class
as the mean of the core area index of all patches belonging to class i.
The core area index is the percentag of core area in relation to patch area.
A cell is defined as core area if the cell has no neighbour with a different
value than itself (rook's case).

\subsection{Units}{Percent}
\subsection{Range}{CAI_MN >= 0}
\subsection{Behaviour}{Equals CAI_MN = 0 if CAI = 0 for all patches. Increases,
without limit, as the core area indices increase.}
}
\examples{
lsm_c_cai_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_cai}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_cai_sd}},
\code{\link{lsm_c_cai_cv}}, \cr
\code{\link{lsm_l_cai_mn}},
\code{\link{lsm_l_cai_sd}},
\code{\link{lsm_l_cai_cv}}
}
