% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_circle_cv.R
\name{lsm_c_circle_cv}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.RasterLayer}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.RasterStack}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.RasterBrick}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.list}
\title{Related Circumscribing Circle (CIRCLE)}
\usage{
lsm_c_circle_cv(landscape)

\method{lsm_c_circle_cv}{RasterLayer}(landscape)

\method{lsm_c_circle_cv}{RasterStack}(landscape)

\method{lsm_c_circle_cv}{RasterBrick}(landscape)

\method{lsm_c_circle_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Related Circumscribing Circle (class)
}
\details{
Equals the 1 - the patch area (m^2) divided by the area (m^2) of the smallest circumscribing circle.
\subsection{Units}{None}
\subsection{Range}{0 <= CIRCLE < 1}
}
\examples{
lsm_c_circle_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
