% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_circle_cv.R
\name{lsm_c_circle_cv}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.RasterLayer}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.RasterStack}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.RasterBrick}
\alias{lsm_c_circle_cv}
\alias{lsm_c_circle_cv.list}
\title{CIRCLE_CV (Class level)}
\usage{
lsm_c_circle_cv(landscape)

\method{lsm_c_circle_cv}{RasterLayer}(landscape)

\method{lsm_c_circle_cv}{RasterStack}(landscape)

\method{lsm_c_circle_cv}{RasterBrick}(landscape)

\method{lsm_c_circle_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecient of variation of related circumscribing circle (Shape metric)
}
\details{
\deqn{CIRCLE_{CV} = cv(CIRCLE[patch_{ij}])}
where \eqn{CIRCLE[patch_{ij}]} is the related circumscribing circle of each patch.

CIRCLE_CV is a 'Shape metric' and summarises each class as the coeffiecent of variation of
the related circumscribing circle of all patches belonging to class i. CIRCLE describes
the ratio between the patch area and the smallest circumscribing circle of the patch
and characterises the compactness of the patch. CIRCLE_CV describes the differences among
patches of the same class i in the landscape. Because it is scaled to the mean,
it is easily comparable.

\subsection{Units}{None}
\subsection{Range}{CIRCLE_CV >= 0}
\subsection{Behaviour}{Equals CIRCLE_CV if the related circumscribing circle is identical
for all patches. Increases, without limit, as the variation of related circumscribing
circles increases.}
}
\examples{
lsm_c_circle_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_circle}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_circle_mn}},
\code{\link{lsm_c_circle_sd}}, \cr
\code{\link{lsm_l_circle_mn}},
\code{\link{lsm_l_circle_sd}},
\code{\link{lsm_l_circle_cv}}
}
