% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_circle_mn.R
\name{lsm_c_circle_mn}
\alias{lsm_c_circle_mn}
\alias{lsm_c_circle_mn}
\alias{lsm_c_circle_mn.RasterLayer}
\alias{lsm_c_circle_mn}
\alias{lsm_c_circle_mn.RasterStack}
\alias{lsm_c_circle_mn}
\alias{lsm_c_circle_mn.RasterBrick}
\alias{lsm_c_circle_mn}
\alias{lsm_c_circle_mn.list}
\title{CIRCLE_MN (Class level)}
\usage{
lsm_c_circle_mn(landscape)

\method{lsm_c_circle_mn}{RasterLayer}(landscape)

\method{lsm_c_circle_mn}{RasterStack}(landscape)

\method{lsm_c_circle_mn}{RasterBrick}(landscape)

\method{lsm_c_circle_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of related circumscribing circle (Shape metric)
}
\details{
\deqn{CIRCLE_{MN} = mean(CIRCLE[patch_{ij}])}
where \eqn{CIRCLE[patch_{ij}]} is the related circumscribing circle of each patch.

CIRCLE_MN is a 'Shape metric' and summarises each class as the mean of the related
circumscribing circle of all patches belonging to class i. CIRCLE describes
the ratio between the patch area and the smallest circumscribing circle of the patch
and characterises the compactness of the patch.

\subsection{Units}{None}
\subsection{Range}{CIRCLE_MN > 0}
\subsection{Behaviour}{Approaches CIRCLE_MN = 0 if the related circumscribing circle
of all patches is small. Increases, without limit, as the related circumscribing circles
increase.}
}
\examples{
lsm_c_circle_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_circle}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_circle_sd}},
\code{\link{lsm_c_circle_cv}}, \cr
\code{\link{lsm_l_circle_mn}},
\code{\link{lsm_l_circle_sd}},
\code{\link{lsm_l_circle_cv}}
}
