% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core.R
\name{lsm_c_core}
\alias{lsm_c_core}
\alias{lsm_c_core}
\alias{lsm_c_core.RasterLayer}
\alias{lsm_c_core}
\alias{lsm_c_core.RasterStack}
\alias{lsm_c_core}
\alias{lsm_c_core.RasterBrick}
\alias{lsm_c_core}
\alias{lsm_c_core.list}
\title{Total core area (class level)}
\usage{
lsm_c_core(landscape)

\method{lsm_c_core}{RasterLayer}(landscape)

\method{lsm_c_core}{RasterStack}(landscape)

\method{lsm_c_core}{RasterBrick}(landscape)

\method{lsm_c_core}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total core area (class level)
}
\details{
Equals the area within a patch that is not on the edge of the patch of class i.
In other words, the area of a patch that has only neighbouring cells of the same type
of class i
\subsection{Units}{Hectares}
\subsection{Range}{CORE >= 0}
\subsection{Behaviour}{CORE increases without limit as patch areas increase
and patch shapes simplify. CORE = 0 when every cell in every patch of class i
is an edge}
}
\examples{
lsm_c_core(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
