% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core_mn.R
\name{lsm_c_core_mn}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterLayer}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterStack}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterBrick}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.list}
\title{CORE_MN (class level)}
\usage{
lsm_c_core_mn(landscape, directions)

\method{lsm_c_core_mn}{RasterLayer}(landscape, directions = 8)

\method{lsm_c_core_mn}{RasterStack}(landscape, directions = 8)

\method{lsm_c_core_mn}{RasterBrick}(landscape, directions = 8)

\method{lsm_c_core_mn}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Mean of core area (Core area metric)
}
\details{
\deqn{CORE_{MN} = mean(CORE[patch_{ij}])}
where \eqn{CORE[patch_{ij}]} is the core area in square meters of each patch.

CORE_MN is a 'Core area metric' and equals the mean of core areas of all patches
belonging to class i. The core area is defined as all cells that have no
neighbour with a different value than themselves (rook's case).

\subsection{Units}{Hectares}
\subsection{Range}{CORE_MN >= 0}
\subsection{Behaviour}{Equals CORE_MN = 0 if CORE = 0 for all patches. Increases,
without limit, as the core area indices increase.}
}
\examples{
lsm_c_core_mn(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_core}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_core_sd}},
\code{\link{lsm_c_core_cv}}, \cr
\code{\link{lsm_l_core_mn}},
\code{\link{lsm_l_core_sd}},
\code{\link{lsm_l_core_cv}}
}
