% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core_mn.R
\name{lsm_c_core_mn}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterLayer}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterStack}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.RasterBrick}
\alias{lsm_c_core_mn}
\alias{lsm_c_core_mn.list}
\title{Core area distribution (class level)}
\usage{
lsm_c_core_mn(landscape, ...)

\method{lsm_c_core_mn}{RasterLayer}(landscape, ...)

\method{lsm_c_core_mn}{RasterStack}(landscape, ...)

\method{lsm_c_core_mn}{RasterBrick}(landscape, ...)

\method{lsm_c_core_mn}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Mean of patch core area (class level)
}
\details{
Equals the mean of the patch core area of class i. The core area is
the area within a patch that is not on the edge of the patch of class i.
In other words, the area of a patch that has only neighbouring cells of the same type
of class i
\deqn{CORE_MN = mean(core[patch_i])}
\subsection{Units}{Hectares}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_core_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
