% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_core_sd.R
\name{lsm_c_core_sd}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.RasterLayer}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.RasterStack}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.RasterBrick}
\alias{lsm_c_core_sd}
\alias{lsm_c_core_sd.list}
\title{Core area distribution (class level)}
\usage{
lsm_c_core_sd(landscape)

\method{lsm_c_core_sd}{RasterLayer}(landscape)

\method{lsm_c_core_sd}{RasterStack}(landscape)

\method{lsm_c_core_sd}{RasterBrick}(landscape)

\method{lsm_c_core_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation of patch core area (class level)
}
\details{
\deqn{CORE_{SD} = sd(CORE[patch_{ij}])}
where \eqn{CORE[patch_{ij}]} is the core area in square meters of each patch.

CORE_SD is a 'Core area metric'. It equals the standard deviation of the core area
of each patch belonging to class i. The core area is defined as all cells that have no
neighbour with a different value than themselves (rook's case). The metric describes the
differences among patches of the same class i in the landscape.

\subsection{Units}{Hectares}
\subsection{Range}{CORE_SD >= 0}
\subsection{Behaviour}{Equals CORE_SD = 0 if all patches have the same core area.
Increases, without limit, as the variation of patch core areas increases.}
}
\examples{
lsm_c_core_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_core}},
\code{\link{sd}}, \cr
\code{\link{lsm_c_core_mn}},
\code{\link{lsm_c_core_cv}}, \cr
\code{\link{lsm_l_core_mn}},
\code{\link{lsm_l_core_sd}},
\code{\link{lsm_l_core_cv}}
}
