% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_cpland.R
\name{lsm_c_cpland}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.RasterLayer}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.RasterStack}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.RasterBrick}
\alias{lsm_c_cpland}
\alias{lsm_c_cpland.list}
\title{Core area percentage of landscape (class level)}
\usage{
lsm_c_cpland(landscape, ...)

\method{lsm_c_cpland}{RasterLayer}(landscape, ...)

\method{lsm_c_cpland}{RasterStack}(landscape, ...)

\method{lsm_c_cpland}{RasterBrick}(landscape, ...)

\method{lsm_c_cpland}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{???}
}
\value{
tibble
}
\description{
Core area percentage of landscape (class level)
}
\details{
Core area percenage of landscape equals the sum of core area of class i
divided by the total area. In other words, CPLAND equals the percentage of
the landscape belonging to class i only including core area. Because CPLAND is
a relative measure, it is comparable among landscapes with different total areas
\deqn{CPLAND = (sum(core[patch_i]) / total area) * 100}
\subsection{Units}{Percentage}
\subsection{Range}{0 <= CPLAND < 100}
\subsection{Behaviour}{CPLAND approaches CPLAND =  0 when patches
do not contain large core areas, i.e. are small and complex. Approaches CPLAND = 100
when only one class and patch is present}
}
\examples{
lsm_c_cpland(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
