% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_dcad.R
\name{lsm_c_dcad}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterLayer}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterStack}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterBrick}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.list}
\title{DCAD (class level)}
\usage{
lsm_c_dcad(landscape)

\method{lsm_c_dcad}{RasterLayer}(landscape)

\method{lsm_c_dcad}{RasterStack}(landscape)

\method{lsm_c_dcad}{RasterBrick}(landscape)

\method{lsm_c_dcad}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Disjunct core area density (Corea area metric)
}
\details{
\deqn{DCAD = (\frac{\sum_{j=1}^{n} n_{ij}^{core}}{A}) * 10000 * 100}
where \eqn{n_{ij}^{core}} is the number of disjunct core areas and \eqn{A}
is the total landscape area in square meters.

DCAD is a 'Core area metric'. It equals the number of disjunct core areas per
100 ha. relative to the total area. A disjunct core area is a 'patch within
the patch' containing only core cells. A cell is defined as core area if the cell has no
neighbour with a different value than itself (rook's case). The metric is relative and
therefore comparable among landscapes with different total areas.

\subsection{Units}{Number per 100 hectares}
\subsection{Range}{DCAD >= 0}
\subsection{Behaviour}{Equals DCAD = 0 when DCORE = 0, i.e. no patch of class i contains
a disjunct core area. Increases, without limit, as disjunct core areas become more
present, i.e. patches becoming larger and less complex.}
}
\examples{
lsm_c_dcad(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_c_ndca}},
\code{\link{lsm_l_ta}}, \cr
\code{\link{lsm_l_dcad}}
}
