% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_dcad.R
\name{lsm_c_dcad}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterLayer}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterStack}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.RasterBrick}
\alias{lsm_c_dcad}
\alias{lsm_c_dcad.list}
\title{Disjunct core area density (class level)}
\usage{
lsm_c_dcad(landscape)

\method{lsm_c_dcad}{RasterLayer}(landscape)

\method{lsm_c_dcad}{RasterStack}(landscape)

\method{lsm_c_dcad}{RasterBrick}(landscape)

\method{lsm_c_dcad}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Disjunct core area density (class level)
}
\details{
Disjunct core area density equals the sum of number of core areas  of
class i divided by the total area multiplied by 100 and 10 000. In other words,
it is the number of core areas relative to the total area scaled to 100 hectares (???).
The measure is relative and scaled to 100 hecatres (???) it is comparable among
landscapes with different total areas
\deqn{DCAD = (sum(ncore[patch_i]) / total area) * 100 * 10 000}
\subsection{Units}{Number per hectares (FRAGSTATS has per 100 ha)}
\subsection{Range}{DCAD >= 0}
\subsection{Behaviour}{DCAD = 0 when TCA = 0, i.e. every cell in patches
of class i is an edge. DCAD increases without limit as core areas become more
present, i.e. patches becoming larger and less complex}
}
\examples{
lsm_c_dcad(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
