% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_division.R
\name{lsm_c_division}
\alias{lsm_c_division}
\alias{lsm_c_division}
\alias{lsm_c_division.RasterLayer}
\alias{lsm_c_division}
\alias{lsm_c_division.RasterStack}
\alias{lsm_c_division}
\alias{lsm_c_division.RasterBrick}
\alias{lsm_c_division}
\alias{lsm_c_division.list}
\title{Division index (class level)}
\usage{
lsm_c_division(landscape)

\method{lsm_c_division}{RasterLayer}(landscape)

\method{lsm_c_division}{RasterStack}(landscape)

\method{lsm_c_division}{RasterBrick}(landscape)

\method{lsm_c_division}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Division index (class level)
}
\examples{
lsm_c_division(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
