% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ed.R
\name{lsm_c_ed}
\alias{lsm_c_ed}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterLayer}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterStack}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterBrick}
\alias{lsm_c_ed}
\alias{lsm_c_ed.list}
\title{Edge Density (class level)}
\usage{
lsm_c_ed(landscape, count_boundary)

\method{lsm_c_ed}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_c_ed}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_c_ed}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_c_ed}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Edge Density (class level)
}
\details{
Edge density equals the sum of the length of all edges of class i divided by the total
area multiplied 10 000. Because edge density is relative, comparison among landcapes with different total areas
are possible
\deqn{ED = (sum(edges[class_i]) / total area) * 10 000}
\subsection{Units}{Meters per hectare}
\subsection{Range}{ED >= 0}
\subsection{Behaviour}{ED increases as the landcapes becomes more patchy}
}
\examples{
lsm_c_ed(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
