% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ed.R
\name{lsm_c_ed}
\alias{lsm_c_ed}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterLayer}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterStack}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterBrick}
\alias{lsm_c_ed}
\alias{lsm_c_ed.list}
\title{Edge Density}
\usage{
lsm_c_ed(landscape)

\method{lsm_c_ed}{RasterLayer}(landscape)

\method{lsm_c_ed}{RasterStack}(landscape)

\method{lsm_c_ed}{RasterBrick}(landscape)

\method{lsm_c_ed}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Edge Density
}
\examples{
lsm_c_ed(landscape)
lsm_c_ed(landscape_stack)
lsm_c_ed(raster::as.list(landscape_stack))

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
