% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ed.R
\name{lsm_c_ed}
\alias{lsm_c_ed}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterLayer}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterStack}
\alias{lsm_c_ed}
\alias{lsm_c_ed.RasterBrick}
\alias{lsm_c_ed}
\alias{lsm_c_ed.list}
\title{ED (class level)}
\usage{
lsm_c_ed(landscape, count_boundary)

\method{lsm_c_ed}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_c_ed}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_c_ed}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_c_ed}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{}
}
\value{
tibble
}
\description{
Edge Density (Area and Edge metric)
}
\details{
\deqn{ED = \frac{\sum_{k = 1} ^ {m} e_{ik}}{A} * 10000}
where \eqn{e_{ik}} is the total edge length in meters and \eqn{A} is the total
landcape area in square meters.

ED is an 'Area and Edge metric'. The edge density equals the sum of all edges of class i
in relation to the landcape area. The boundary of the landscape is only included in the
corresponding total class edge length if \code{count_boundary} is set to \code{TRUE}.
The metric describes the configuration of the landscape, e.g. because  an  aggregation
of the same class will result in a low edge density. The metric is standarized to the
total landscape area, and therefore comparisons among landscapes with different total
areas are possible.
\subsection{Units}{Meters per hectare}
\subsection{Range}{ED >= 0}
\subsection{Behaviour}{Equals ED = 0 if only one patch is present (and the landcape
boundary is not included) and increases, without limit, as the landcapes becomes more
patchy}
}
\examples{
lsm_c_ed(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_c_te}},
\code{\link{lsm_l_ta}}, \cr
\code{\link{lsm_l_ed}}
}
