% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_enn_cv.R
\name{lsm_c_enn_cv}
\alias{lsm_c_enn_cv}
\alias{lsm_c_enn_cv}
\alias{lsm_c_enn_cv.RasterLayer}
\alias{lsm_c_enn_cv}
\alias{lsm_c_enn_cv.RasterStack}
\alias{lsm_c_enn_cv}
\alias{lsm_c_enn_cv.RasterBrick}
\alias{lsm_c_enn_cv}
\alias{lsm_c_enn_cv.list}
\title{Euclidean Nearest Neighbor Distance Distribution (class level)}
\usage{
lsm_c_enn_cv(landscape)

\method{lsm_c_enn_cv}{RasterLayer}(landscape)

\method{lsm_c_enn_cv}{RasterStack}(landscape)

\method{lsm_c_enn_cv}{RasterBrick}(landscape)

\method{lsm_c_enn_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation Euclidean Nearest Neighbor Distance (class level)
}
\details{
Equals the coeffiecent of variation of the eclidean nearest neighbor distance of class i.
ENN equals the distance the the nearest neigbouring patch of the same patch type
(shortest edge-to-edge distance)
\deqn{ENN_CV = cv(ENN[patch_i])}
\subsection{Units}{Meters}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_enn_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
