% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_enn_mn.R
\name{lsm_c_enn_mn}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.RasterLayer}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.RasterStack}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.RasterBrick}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.list}
\title{Euclidean Nearest Neighbor Distance (ENN_MN)}
\usage{
lsm_c_enn_mn(landscape)

\method{lsm_c_enn_mn}{RasterLayer}(landscape)

\method{lsm_c_enn_mn}{RasterStack}(landscape)

\method{lsm_c_enn_mn}{RasterBrick}(landscape)

\method{lsm_c_enn_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean Euclidean Nearest Neighbor Distance (class level)
}
\details{
Equals the mean of the eclidean nearest neighbor distance of class i. ENN equals
the distance the the nearest neigbouring patch of the same patch type
(shortest edge-to-edge distance)
\deqn{ENN_MN = mean(ENN[class_i])}
\subsection{Units}{Meters}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_enn_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
