% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_enn_mn.R
\name{lsm_c_enn_mn}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.RasterLayer}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.RasterStack}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.RasterBrick}
\alias{lsm_c_enn_mn}
\alias{lsm_c_enn_mn.list}
\title{ENN_MN (patch level)}
\usage{
lsm_c_enn_mn(landscape)

\method{lsm_c_enn_mn}{RasterLayer}(landscape)

\method{lsm_c_enn_mn}{RasterStack}(landscape)

\method{lsm_c_enn_mn}{RasterBrick}(landscape)

\method{lsm_c_enn_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean of euclidean nearest-neighbor distance (Aggregation metric)
}
\details{
\deqn{ENN_{MN} = cv(mean[patch_{ij}])}
where \eqn{ENN[patch_{ij}]} is the euclidean nearest-neighbor distance
of each patch.

ENN_CV is an 'Aggregation metric'. It summarises each class as the mean of each patch
belonging to class i. ENN measures the distance to the  nearest neighbouring patch
of the same class i. The distance is measured from edge-to-edge. The range is limited
by the cell resolution on the lower limit and the landscape extent on the upper limit.

\subsection{Units}{Meters}
\subsection{Range}{ENN_MN > 0}
\subsection{Behaviour}{Approaches ENN_MN = 0 as the distance to the nearest neighbour
decreases, i.e. patches of the same class i are more aggregated. Increases, without limit,
as the distance between neighbouring patches of the same class i increases, i.e. patches are
more isolated.}
}
\examples{
lsm_c_enn_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_enn}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_enn_sd}},
\code{\link{lsm_c_enn_cv}}, \cr
\code{\link{lsm_l_enn_mn}},
\code{\link{lsm_l_enn_sd}},
\code{\link{lsm_l_enn_cv}}
}
