% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_enn_mn.R
\name{lsm_c_enn_mn}
\alias{lsm_c_enn_mn}
\title{Euclidean Nearest Neighbor Distance Distribution (ENN_MN)}
\usage{
lsm_c_enn_mn(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Euclidean Nearest Neighbor Distance Distribution (class level)
}
\examples{
lsm_c_enn_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
