% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_enn_sd.R
\name{lsm_c_enn_sd}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.RasterLayer}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.RasterStack}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.RasterBrick}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.list}
\title{Euclidean Nearest Neighbor Distance Distribution (class level)}
\usage{
lsm_c_enn_sd(landscape)

\method{lsm_c_enn_sd}{RasterLayer}(landscape)

\method{lsm_c_enn_sd}{RasterStack}(landscape)

\method{lsm_c_enn_sd}{RasterBrick}(landscape)

\method{lsm_c_enn_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation Euclidean Nearest Neighbor Distance (class level)
}
\details{
Equals the standard deviation of the eclidean nearest neighbor distance of class i.
ENN equals the distance the the nearest neigbouring patch of the same patch type
(shortest edge-to-edge distance)
\deqn{ENN_SD = sd(ENN[patch_i])}
\subsection{Units}{Meters}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_enn_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
