% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_enn_sd.R
\name{lsm_c_enn_sd}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.RasterLayer}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.RasterStack}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.RasterBrick}
\alias{lsm_c_enn_sd}
\alias{lsm_c_enn_sd.list}
\title{ENN_SD (class level)}
\usage{
lsm_c_enn_sd(landscape, directions)

\method{lsm_c_enn_sd}{RasterLayer}(landscape, directions = 8)

\method{lsm_c_enn_sd}{RasterStack}(landscape, directions = 8)

\method{lsm_c_enn_sd}{RasterBrick}(landscape, directions = 8)

\method{lsm_c_enn_sd}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Standard deviation of euclidean nearest-neighbor distance (Aggregation metric)
}
\details{
\deqn{ENN_{SD} = sd(ENN[patch_{ij}])}
where \eqn{ENN[patch_{ij}]} is the euclidean nearest-neighbor distance
of each patch.

ENN_CV is an 'Aggregation metric'. It summarises each class as the standard
deviation of each patch belonging to class i. ENN measures the distance to the  nearest
neighbouring patch of the same class i. The distance is measured from edge-to-edge.
The range is limited by the cell resolution on the lower limit and the landscape extent
on the upper limit. The metric is a simple way to describe patch isolation. Because it is
scaled to the mean, it is easily comparable among different landscapes.

\subsection{Units}{Meters}
\subsection{Range}{ENN_SD >= 0}
\subsection{Behaviour}{Equals ENN_SD = 0 if the euclidean nearest-neighbor distance is
identical for all patches. Increases, without limit, as the variation of ENN increases.}
}
\examples{
lsm_c_enn_sd(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_enn}},
\code{\link{sd}}, \cr
\code{\link{lsm_c_enn_mn}},
\code{\link{lsm_c_enn_cv}}, \cr
\code{\link{lsm_l_enn_mn}},
\code{\link{lsm_l_enn_sd}},
\code{\link{lsm_l_enn_cv}}
}
