% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_frac_mn.R
\name{lsm_c_frac_mn}
\alias{lsm_c_frac_mn}
\alias{lsm_c_frac_mn}
\alias{lsm_c_frac_mn.RasterLayer}
\alias{lsm_c_frac_mn}
\alias{lsm_c_frac_mn.RasterStack}
\alias{lsm_c_frac_mn}
\alias{lsm_c_frac_mn.RasterBrick}
\alias{lsm_c_frac_mn}
\alias{lsm_c_frac_mn.list}
\title{Perimeter area ratio  (patch level)}
\usage{
lsm_c_frac_mn(landscape)

\method{lsm_c_frac_mn}{RasterLayer}(landscape)

\method{lsm_c_frac_mn}{RasterStack}(landscape)

\method{lsm_c_frac_mn}{RasterBrick}(landscape)

\method{lsm_c_frac_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Ratio of patch perimeter and area
}
\examples{
lsm_c_frac_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
