% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_frac_sd.R
\name{lsm_c_frac_sd}
\alias{lsm_c_frac_sd}
\alias{lsm_c_frac_sd}
\alias{lsm_c_frac_sd.RasterLayer}
\alias{lsm_c_frac_sd}
\alias{lsm_c_frac_sd.RasterStack}
\alias{lsm_c_frac_sd}
\alias{lsm_c_frac_sd.RasterBrick}
\alias{lsm_c_frac_sd}
\alias{lsm_c_frac_sd.list}
\title{FRAC_SD (class level)}
\usage{
lsm_c_frac_sd(landscape)

\method{lsm_c_frac_sd}{RasterLayer}(landscape)

\method{lsm_c_frac_sd}{RasterStack}(landscape)

\method{lsm_c_frac_sd}{RasterBrick}(landscape)

\method{lsm_c_frac_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation fractal dimension index (Shape metric)
}
\details{
\deqn{FRAC_{SD} = sd(FRAC[patch_{ij}])}
where \eqn{FRAC[patch_{ij}]} equals the fractal dimension index of each patch.

FRAC_SD is a 'Shape metric'. The metric summarises each class
as the standard deviation of the fractal dimension index of all patches
belonging to class i. The fractal dimenstion index is based on the patch perimeter and
the patch area and describes the patch complexity.

\subsection{Units}{None}
\subsection{Range}{FRAC_SD>= 0 }
\subsection{Behaviour}{Equals FRAC_SD = 0 if the fractal dimension index is identical
for all patches. Increases, without limit, as the variation of the fractal dimension
indices increases.}
}
\examples{
lsm_c_frac_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_frac}},
\code{\link{sd}}, \cre
\code{\link{lsm_c_frac_mn}},
\code{\link{lsm_c_frac_cv}}, \cr
\code{\link{lsm_l_frac_mn}},
\code{\link{lsm_l_frac_sd}},
\code{\link{lsm_l_frac_cv}}
}
