% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_gyrate_cv.R
\name{lsm_c_gyrate_cv}
\alias{lsm_c_gyrate_cv}
\alias{lsm_c_gyrate_cv}
\alias{lsm_c_gyrate_cv.RasterLayer}
\alias{lsm_c_gyrate_cv}
\alias{lsm_c_gyrate_cv.RasterStack}
\alias{lsm_c_gyrate_cv}
\alias{lsm_c_gyrate_cv.RasterBrick}
\alias{lsm_c_gyrate_cv}
\alias{lsm_c_gyrate_cv.list}
\title{GYRATE_CV (class level)}
\usage{
lsm_c_gyrate_cv(landscape)

\method{lsm_c_gyrate_cv}{RasterLayer}(landscape)

\method{lsm_c_gyrate_cv}{RasterStack}(landscape)

\method{lsm_c_gyrate_cv}{RasterBrick}(landscape)

\method{lsm_c_gyrate_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation radius of gyration (Area and edge metric)
}
\details{
\deqn{GYRATE_{CV} = cv(GYRATE[patch_{ij}])}
where \eqn{GYRATE[patch_{ij}]} equals the radius of gyration of each patch.

GYRATE_CV is an 'Area and edge metric'. The metric summarises each class
as the coeffiecent of variation of the radius of gyration of all patches
belonging to class i. GYRATE measures the distance from each cell to the patch
centroid and is based on cell center-to-cell center distances. The metrics characterises
both the patch area and compactness. The coeffiecent of variation is
scaled to the mean and comparable among different landscapes.

\subsection{Units}{Meters}
\subsection{Range}{GYRATE_CV >= 0 }
\subsection{Behaviour}{Equals GYRATE_CV = 0 if the radius of gyration is identical
for all patches. Increases, without limit, as the variation of the radius of gyration
increases.}
}
\examples{
lsm_c_gyrate_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_gyrate}},
\code{\link{cv}}, \cr
\code{\link{lsm_c_gyrate_mn}},
\code{\link{lsm_c_gyrate_sd}}, \cr
\code{\link{lsm_l_gyrate_mn}},
\code{\link{lsm_l_gyrate_sd}},
\code{\link{lsm_l_gyrate_cv}}
}
