% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_iji.R
\name{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterLayer}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterStack}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterBrick}
\alias{lsm_c_iji}
\alias{lsm_c_iji.list}
\title{Interspersion and Juxtaposition index (class level)}
\usage{
lsm_c_iji(landscape)

\method{lsm_c_iji}{RasterLayer}(landscape)

\method{lsm_c_iji}{RasterStack}(landscape)

\method{lsm_c_iji}{RasterBrick}(landscape)

\method{lsm_c_iji}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Interspersion and Juxtaposition index (class level)
}
\details{
\deqn{IJI = \frac{- \sum \limits_{k = 1}^{m} \Bigg[ \Bigg( \frac{e_{ik}}{\sum \limits_{k = 1}^{m} e_{ik}} \Bigg) ln \Bigg( \frac{e_{ik}}{\sum \limits_{k = 1}^{m} e_{ik}} \Bigg) \Bigg]}{ln(m - 1)}  * 100}

"salt and pepper" metric.
\deqn{???}
\subsection{Units}{Percent}
\subsection{Range}{0 < IJI ≤ 100}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_iji(landscape)
}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
