% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_iji.R
\name{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterLayer}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterStack}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterBrick}
\alias{lsm_c_iji}
\alias{lsm_c_iji.list}
\title{Interspersion and Juxtaposition index (class level)}
\usage{
lsm_c_iji(landscape, count_boundary)

\method{lsm_c_iji}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_c_iji}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_c_iji}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_c_iji}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{???}
}
\value{
tibble
}
\description{
Interspersion and Juxtaposition index (class level)
}
\details{
???
\deqn{???}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_iji(landscape)
}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
