% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_iji.R
\name{lsm_c_iji}
\alias{lsm_c_iji}
\alias{lsm_c_iji.RasterLayer}
\alias{lsm_c_iji.RasterStack}
\alias{lsm_c_iji.RasterBrick}
\alias{lsm_c_iji.list}
\title{Interspersion and Juxtaposition index (class level)}
\usage{
lsm_c_iji(landscape, verbose)

\method{lsm_c_iji}{RasterLayer}(landscape, verbose = TRUE)

\method{lsm_c_iji}{RasterStack}(landscape, verbose = TRUE)

\method{lsm_c_iji}{RasterBrick}(landscape, verbose = TRUE)

\method{lsm_c_iji}{list}(landscape, verbose = TRUE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{verbose}{Print warning message if not sufficient patches are present}
}
\value{
tibble
}
\description{
Interspersion and Juxtaposition index (class level)
}
\details{
\deqn{IJI = \frac{- \sum \limits_{k = 1}^{m} \Bigg[ \Bigg( \frac{e_{ik}}{\sum \limits_{k = 1}^{m} e_{ik}} \Bigg) ln \Bigg( \frac{e_{ik}}{\sum \limits_{k = 1}^{m} e_{ik}} \Bigg) \Bigg]}{ln(m - 1)}  * 100}

where \eqn{e_{ik}} are the unique adjacencies of all classes (lower/upper triangle of
the adjacency table - without the diagonal) and \eqn{m} is the number of classes.

IJI is an 'Aggregation metric'. It is a so called "salt and pepper" metric and
describes the intermixing of classes (i.e. without considering like adjacencies - the
diagonal of the adjacency table). The number of classes to calculate IJI must be >= than 3.

\subsection{Units}{Percent}
\subsection{Range}{0 < IJI ≤ 100}
\subsection{Behaviour}{Approaches 0 if a class is only adjacent to a single other class
and equals 100 when a class is equally adjacent to all other classes.}
}
\examples{
lsm_c_iji(landscape)
}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
