% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_lpi.R
\name{lsm_c_lpi}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterLayer}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterStack}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterBrick}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.list}
\title{LPI (class level)}
\usage{
lsm_c_lpi(landscape)

\method{lsm_c_lpi}{RasterLayer}(landscape)

\method{lsm_c_lpi}{RasterStack}(landscape)

\method{lsm_c_lpi}{RasterBrick}(landscape)

\method{lsm_c_lpi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Largest patch index (Area and Edge metric)
}
\details{
\deqn{LPI = \frac{\max\limits_{j = 1} ^ {n} (a_{ij})}{A} * 100}
where \eqn{max(a_{ij})} is the area of the patch in square meters and \eqn{A}
is the total landscape area in square meters.

The largest patch index is an 'Area and edge metric'. It is the percentage of the
landscape covered by the corresponding largest patch of each class i. It is a simple
measure of dominance.

\subsection{Units}{Percentage}
\subsection{Range}{0 < LPI <= 100}
\subsection{Behaviour}{Approaches LPI = 0 when the largest patch is becoming small
and equals LPI = 100 when only one patch is present}
}
\examples{
lsm_c_lpi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{lsm_l_ta}}, \cr
\code{\link{lsm_l_lpi}}
}
