% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_lpi.R
\name{lsm_c_lpi}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterLayer}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterStack}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.RasterBrick}
\alias{lsm_c_lpi}
\alias{lsm_c_lpi.list}
\title{Largest patch index (class level)}
\usage{
lsm_c_lpi(landscape)

\method{lsm_c_lpi}{RasterLayer}(landscape)

\method{lsm_c_lpi}{RasterStack}(landscape)

\method{lsm_c_lpi}{RasterBrick}(landscape)

\method{lsm_c_lpi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Area of the larges patch divided by the total area (class level)
}
\examples{
lsm_c_lpi(landscape)
lsm_c_lpi(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
