% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_lsi.R
\name{lsm_c_lsi}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterLayer}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterStack}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterBrick}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.list}
\title{LSI (class level)}
\usage{
lsm_c_lsi(landscape)

\method{lsm_c_lsi}{RasterLayer}(landscape)

\method{lsm_c_lsi}{RasterStack}(landscape)

\method{lsm_c_lsi}{RasterBrick}(landscape)

\method{lsm_c_lsi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Landscape shape index (Aggregation metric)
}
\details{
\deqn{LSI = \frac{e_{i}}{\min e_{i}}}
where \eqn{e_{i}} is the total edge length in cell surfaces and \eqn{\min e_{i}}
is the minimum total edge length in cell surfaces

LSI is an 'Aggregation metric'. It is the ratio between the actual edge length of
class i and the hypothetical minimum edge length of class i. The minimum edge length equals
the edge length if class i would be maximally aggregated.

\subsection{Units}{None}
\subsection{Ranges}{LSI >= 1}
\subsection{Behaviour}{Equals LSI = 1 when only one squared patch is present or all
patches are maximally aggregated. Increases, without limit, as the length of the
actual edges increases, i.e. the patches become less compact.}
}
\examples{
lsm_c_lsi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_shape}}, \cr
\code{\link{lsm_l_lsi}}
}
