% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_lsi.R
\name{lsm_c_lsi}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterLayer}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterStack}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.RasterBrick}
\alias{lsm_c_lsi}
\alias{lsm_c_lsi.list}
\title{Landscape shape index (class level)}
\usage{
lsm_c_lsi(landscape)

\method{lsm_c_lsi}{RasterLayer}(landscape)

\method{lsm_c_lsi}{RasterStack}(landscape)

\method{lsm_c_lsi}{RasterBrick}(landscape)

\method{lsm_c_lsi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Landscape shape index (class level)
}
\examples{
lsm_c_lsi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
