% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_mesh.R
\name{lsm_c_mesh}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterLayer}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterStack}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterBrick}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.list}
\title{Effective Mesh Size (class level)}
\usage{
lsm_c_mesh(landscape)

\method{lsm_c_mesh}{RasterLayer}(landscape)

\method{lsm_c_mesh}{RasterStack}(landscape)

\method{lsm_c_mesh}{RasterBrick}(landscape)

\method{lsm_c_mesh}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Effective Mesh Size (class level)
}
\details{
The effective mesh size equals the sum of patch areas of class i squared
divided by the total area divided by 10 000 to convert to hectares (???)
\deqn{MESH = (sum(area[patch_i]) ^ 2) / total area}
\subsection{Units}{Hectares}
\subsection{Range}{ratio of cell size to total area (???) <= MESH <= total area}
\subsection{Behaviour}{MESH equals the lower limit when every cell is a patch and
increases when only one patch is present}
}
\examples{
lsm_c_mesh(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
