% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_mesh.R
\name{lsm_c_mesh}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterLayer}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterStack}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.RasterBrick}
\alias{lsm_c_mesh}
\alias{lsm_c_mesh.list}
\title{MESH (class level)}
\usage{
lsm_c_mesh(landscape)

\method{lsm_c_mesh}{RasterLayer}(landscape)

\method{lsm_c_mesh}{RasterStack}(landscape)

\method{lsm_c_mesh}{RasterBrick}(landscape)

\method{lsm_c_mesh}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Effective Mesh Size (Aggregation metric)
}
\details{
\deqn{MESH = \frac{\sum_{j = 1}^{n} a_{ij} ^ 2}{A} * \frac{1}{10000}}
where \eqn{a_{ij}} is the patch area in square meters and \eqn{A} is the
total landscape area in square meters.

The effective mesh size is an 'Aggregation metric'. Because each patch is squared
before the sums for each group i are calculated and the sum is standardised by the
total landscape area, MESH is a relative measure of patch structure. MESH is
perfectly, negatively correlated to \code{\link{lsm_c_division}}.

\subsection{Units}{Hectares}
\subsection{Range}{cell size / total area <= MESH <= total area}
\subsection{Behaviour}{Equals cellsize/total area if class covers only
one cell and equals total area if only one patch is present.}
}
\examples{
lsm_c_mesh(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{lsm_l_ta}}, \cr
\code{\link{lsm_l_mesh}}
}
