% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ncore.R
\name{lsm_c_ncore}
\alias{lsm_c_ncore}
\alias{lsm_c_ncore}
\alias{lsm_c_ncore.RasterLayer}
\alias{lsm_c_ncore}
\alias{lsm_c_ncore.RasterStack}
\alias{lsm_c_ncore}
\alias{lsm_c_ncore.RasterBrick}
\alias{lsm_c_ncore}
\alias{lsm_c_ncore.list}
\title{Number of core areas (class level)}
\usage{
lsm_c_ncore(landscape)

\method{lsm_c_ncore}{RasterLayer}(landscape)

\method{lsm_c_ncore}{RasterStack}(landscape)

\method{lsm_c_ncore}{RasterBrick}(landscape)

\method{lsm_c_ncore}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of core areas (class level)
}
\details{
Number of core areas equals the sum of number of core areas of class i.
Called number of disjunct core areas in FRAGSTATS
\deqn{NDCA = sum(ncore[patch_i])}
\subsection{Units}{None}
\subsection{Range}{NDCA >= 0 \cr }
}
\examples{
lsm_c_ncore(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
