% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ncore_sd.R
\name{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.RasterLayer}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.RasterStack}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.RasterBrick}
\alias{lsm_c_ncore_sd}
\alias{lsm_c_ncore_sd.list}
\title{Distrubution number of core areas (class level)}
\usage{
lsm_c_ncore_sd(landscape)

\method{lsm_c_ncore_sd}{RasterLayer}(landscape)

\method{lsm_c_ncore_sd}{RasterStack}(landscape)

\method{lsm_c_ncore_sd}{RasterBrick}(landscape)

\method{lsm_c_ncore_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation (class level)
}
\examples{
lsm_c_ncore_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
