% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_np.R
\name{lsm_c_np}
\alias{lsm_c_np}
\alias{lsm_c_np}
\alias{lsm_c_np.RasterLayer}
\alias{lsm_c_np}
\alias{lsm_c_np.RasterStack}
\alias{lsm_c_np}
\alias{lsm_c_np.RasterBrick}
\alias{lsm_c_np}
\alias{lsm_c_np.list}
\title{Number of patches (class level)}
\usage{
lsm_c_np(landscape)

\method{lsm_c_np}{RasterLayer}(landscape)

\method{lsm_c_np}{RasterStack}(landscape)

\method{lsm_c_np}{RasterBrick}(landscape)

\method{lsm_c_np}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of patches (class level)
}
\details{
Number of patches equals the number of patches of class i
\subsection{Units}{None}
\subsection{Ranges}{NP >= 1 \cr
NP equals NP = 1 when only one class and landscape is present and increases
without limit as the number of patches increases}
}
\examples{
lsm_c_np(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
