% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pafrac.R
\name{lsm_c_pafrac}
\alias{lsm_c_pafrac}
\alias{lsm_c_pafrac}
\alias{lsm_c_pafrac.RasterLayer}
\alias{lsm_c_pafrac}
\alias{lsm_c_pafrac.RasterStack}
\alias{lsm_c_pafrac}
\alias{lsm_c_pafrac.RasterBrick}
\alias{lsm_c_pafrac}
\alias{lsm_c_pafrac.list}
\title{Perimeter-area fractal dimension  (landscape level)}
\usage{
lsm_c_pafrac(landscape)

\method{lsm_c_pafrac}{RasterLayer}(landscape)

\method{lsm_c_pafrac}{RasterStack}(landscape)

\method{lsm_c_pafrac}{RasterBrick}(landscape)

\method{lsm_c_pafrac}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
2 divided by the beta coeffient of the regression log(area) ~ log(perimeter)
}
\examples{
lsm_c_pafrac(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
