% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_para_mn.R
\name{lsm_c_para_mn}
\alias{lsm_c_para_mn}
\alias{lsm_c_para_mn}
\alias{lsm_c_para_mn.RasterLayer}
\alias{lsm_c_para_mn}
\alias{lsm_c_para_mn.RasterStack}
\alias{lsm_c_para_mn}
\alias{lsm_c_para_mn.RasterBrick}
\alias{lsm_c_para_mn}
\alias{lsm_c_para_mn.list}
\title{Perimeter-area ratio distribution (class level)}
\usage{
lsm_c_para_mn(landscape)

\method{lsm_c_para_mn}{RasterLayer}(landscape)

\method{lsm_c_para_mn}{RasterStack}(landscape)

\method{lsm_c_para_mn}{RasterBrick}(landscape)

\method{lsm_c_para_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{details}{Mean of the perimeter-area ratio of all patches of class i. PARA equals the ration of
patch perimeter and patch area. It is a simple measure of complexity
\deqn{PARA_MN = mean(PARA[patch_i]}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}}
}
\value{
tibble
}
\description{
Mean of perimeter-area ratio (class level)
}
\examples{
lsm_c_para_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
