% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_para_sd.R
\name{lsm_c_para_sd}
\alias{lsm_c_para_sd}
\alias{lsm_c_para_sd}
\alias{lsm_c_para_sd.RasterLayer}
\alias{lsm_c_para_sd}
\alias{lsm_c_para_sd.RasterStack}
\alias{lsm_c_para_sd}
\alias{lsm_c_para_sd.RasterBrick}
\alias{lsm_c_para_sd}
\alias{lsm_c_para_sd.list}
\title{Perimeter-area ratio distribution (class level)}
\usage{
lsm_c_para_sd(landscape)

\method{lsm_c_para_sd}{RasterLayer}(landscape)

\method{lsm_c_para_sd}{RasterStack}(landscape)

\method{lsm_c_para_sd}{RasterBrick}(landscape)

\method{lsm_c_para_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standart deviation of perimeter-area ratio (class level)
}
\details{
Standard deviation of the perimeter-area ratio of all patches of class i.
PARA equals the ration of patch perimeter and patch area. It is a simple measure of complexity
\deqn{PARA_SD = sd(PARA[patch_i]}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_c_para_sd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
