% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_para_sd.R
\name{lsm_c_para_sd}
\alias{lsm_c_para_sd}
\title{PARA_SD (class level)}
\usage{
lsm_c_para_sd(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Standard deviation perimeter-area ratio (Shape metric)
}
\details{
\deqn{PARA_{SD} = sd(PARA[patch_{ij}]}
where \eqn{PARA[patch_{ij}]} is the perimeter area ratio of each patch.

PARA_SD is a 'Shape metric'. It summarises each class as the standard deviation of
each patch belonging to class i. The perimeter-area ratio describes the patch complexity
in a straightforward way. However, because it is not standarised to a certain shape
(e.g. a square), it is not scale independent, meaning that increasing the patch size
while not changing the patch form will change the ratio.

\subsection{Units}{None}
\subsection{Range}{PARA_SD >= 0}
\subsection{Behaviour}{Equals PARA_SD = 0 if the perimeter-area ratio is identical for
all patches. Increases, without limit, as the variation of the perimeter-area ratio
increases.}
}
\examples{
lsm_c_para_sd(terra::unwrap(landscape))

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: https://www.umass.edu/landeco/
}
\seealso{
\code{\link{lsm_p_para}},
\code{\link{sd}}, \cr
\code{\link{lsm_c_para_mn}},
\code{\link{lsm_c_para_cv}}, \cr
\code{\link{lsm_l_para_mn}},
\code{\link{lsm_l_para_sd}},
\code{\link{lsm_l_para_cv}}
}
