% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pd.R
\name{lsm_c_pd}
\alias{lsm_c_pd}
\alias{lsm_c_pd}
\alias{lsm_c_pd.RasterLayer}
\alias{lsm_c_pd}
\alias{lsm_c_pd.RasterStack}
\alias{lsm_c_pd}
\alias{lsm_c_pd.RasterBrick}
\alias{lsm_c_pd}
\alias{lsm_c_pd.list}
\title{PD (class level)}
\usage{
lsm_c_pd(landscape)

\method{lsm_c_pd}{RasterLayer}(landscape)

\method{lsm_c_pd}{RasterStack}(landscape)

\method{lsm_c_pd}{RasterBrick}(landscape)

\method{lsm_c_pd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Patch density (Aggregation metric)
}
\details{
\deqn{PD = \frac{n_{i}}{A} * 10000 * 100}
where \eqn{n_{i}} is the number of patches and \eqn{A} is the total landscape
area in square meters.

PD is an 'Aggregation metric'. It describes the fragmentation of a class, however, does not
necessarily contain information about the configuration or composition of the class. In
contrast to \code{\link{lsm_c_np}} it is standardised to the area and comparisons among
landscapes with different total area are possible.

\subsection{Units}{Number per 100 hectares}
\subsection{Ranges}{0 < PD <= 1e+06}
\subsection{Behaviour}{Increases as the landscape gets more patchy. Reaches its maximum
if every cell is a different patch.}
}
\examples{
lsm_c_pd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_c_np}},
\code{\link{lsm_l_ta}}, \cr
\code{\link{lsm_l_pd}}
}
