% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pladj.R
\name{lsm_c_pladj}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.RasterLayer}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.RasterStack}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.RasterBrick}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.list}
\title{Percentage of Like Adjacencies (class level)}
\usage{
lsm_c_pladj(landscape)

\method{lsm_c_pladj}{RasterLayer}(landscape)

\method{lsm_c_pladj}{RasterStack}(landscape)

\method{lsm_c_pladj}{RasterBrick}(landscape)

\method{lsm_c_pladj}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Percentage of Like Adjacencies (class level)
}
\details{
Percentage of like adjacencies equals the number of like adjacencies of
class i divided by the total number of like adjacencies. In other other words,
PLADJ equals the percentage of like adjacencies of class i of all like
adjacencies. It is a measure of class aggregation.

\deqn{PLADJ = Number of like adjacencies / sum(Number of like adjacencies)}
\subsection{Units}{Percentage}
\subsection{Ranges}{0 <= PLADJ <= 100 \cr
PLADJ equals PLADJ = 0 when class i is maximally dissagregated, i.e. every
cell is a different patch. PLADJ = 100 when the only one class and patch
is present}
}
\examples{
lsm_c_pladj(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
