% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pladj.R
\name{lsm_c_pladj}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.RasterLayer}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.RasterStack}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.RasterBrick}
\alias{lsm_c_pladj}
\alias{lsm_c_pladj.list}
\title{Percentage of Like Adjacencies}
\usage{
lsm_c_pladj(landscape)

\method{lsm_c_pladj}{RasterLayer}(landscape)

\method{lsm_c_pladj}{RasterStack}(landscape)

\method{lsm_c_pladj}{RasterBrick}(landscape)

\method{lsm_c_pladj}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Percentage of Like Adjacencies
}
\examples{
lsm_c_pladj(landscape)
lsm_c_pladj(landscape_stack)
lsm_c_pladj(raster::as.list(landscape_stack))

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
