% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_pland.R
\name{lsm_c_pland}
\alias{lsm_c_pland}
\alias{lsm_c_pland}
\alias{lsm_c_pland.RasterLayer}
\alias{lsm_c_pland}
\alias{lsm_c_pland.RasterStack}
\alias{lsm_c_pland}
\alias{lsm_c_pland.RasterBrick}
\alias{lsm_c_pland}
\alias{lsm_c_pland.list}
\title{PLAND (class level)}
\usage{
lsm_c_pland(landscape, directions)

\method{lsm_c_pland}{RasterLayer}(landscape, directions = 8)

\method{lsm_c_pland}{RasterStack}(landscape, directions = 8)

\method{lsm_c_pland}{RasterBrick}(landscape, directions = 8)

\method{lsm_c_pland}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which cells should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Percentage of landscape of class (Area and Edge metric)
}
\details{
\deqn{PLAND = \frac{\sum \limits_{j = 1}^{n} a_{ij}} {A} * 100}
where \eqn{a_{ij}} is the area of each patch and \eqn{A} is the total
landscape area.

PLAND is an 'Area and edge metric'. It is the percentage of the landscape
belonging to class i. It is a measure of compostion and because of the relative
character directly comparable among landscapes with different total areas.

\subsection{Units}{Percentage}
\subsection{Range}{0 < PLAND <= 100}
\subsection{Behaviour}{Approaches PLAND = 0 when the proportional class area is decreasing.
Equals PLAND = 100 when only one patch is present.}
}
\examples{
lsm_c_pland(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_c_ca}},
\code{\link{lsm_l_ta}}
}
