% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_split.R
\name{lsm_c_split}
\alias{lsm_c_split}
\alias{lsm_c_split}
\alias{lsm_c_split.RasterLayer}
\alias{lsm_c_split}
\alias{lsm_c_split.RasterStack}
\alias{lsm_c_split}
\alias{lsm_c_split.RasterBrick}
\alias{lsm_c_split}
\alias{lsm_c_split.list}
\title{Splitting index (class level)}
\usage{
lsm_c_split(landscape)

\method{lsm_c_split}{RasterLayer}(landscape)

\method{lsm_c_split}{RasterStack}(landscape)

\method{lsm_c_split}{RasterBrick}(landscape)

\method{lsm_c_split}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Splitting index (class level)
}
\details{
The splitting index equals the squared total area divided by the sum of patch
area squared of class i
\deqn{SPLIT = total area ^ 2 / sum(area[patch_i])}
\subsection{Units}{None}
\subsection{Range}{1 <= SPLIT <= Number of cells squared}
\subsection{Behaviour}{SPLIT = 1 when only one class and patch is present.
SPLIt increases as the number of patches of class i increases}
}
\examples{
lsm_c_split(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
