% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_te.R
\name{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterLayer}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterStack}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterBrick}
\alias{lsm_c_te}
\alias{lsm_c_te.list}
\title{Total Edge  (TE, class scale)}
\usage{
lsm_c_te(landscape, class)

\method{lsm_c_te}{RasterLayer}(landscape, class)

\method{lsm_c_te}{RasterStack}(landscape, class)

\method{lsm_c_te}{RasterBrick}(landscape, class)

\method{lsm_c_te}{list}(landscape, class)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{class}{Which class}
}
\value{
tibble
}
\description{
Total Edge
}
\examples{
lsm_c_te(landscape, 2)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
