% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_te.R
\name{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterLayer}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterStack}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterBrick}
\alias{lsm_c_te}
\alias{lsm_c_te.list}
\title{Total Edge  (TE, class scale)}
\usage{
lsm_c_te(landscape)

\method{lsm_c_te}{RasterLayer}(landscape)

\method{lsm_c_te}{RasterStack}(landscape)

\method{lsm_c_te}{RasterBrick}(landscape)

\method{lsm_c_te}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total Edge
}
\examples{
lsm_c_te(landscape)
lsm_c_te(landscape_stack)
lsm_c_te(raster::as.list(landscape_stack))

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
