% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_te.R
\name{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterLayer}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterStack}
\alias{lsm_c_te}
\alias{lsm_c_te.RasterBrick}
\alias{lsm_c_te}
\alias{lsm_c_te.list}
\title{Total Edge  (class level)}
\usage{
lsm_c_te(landscape, ...)

\method{lsm_c_te}{RasterLayer}(landscape, ...)

\method{lsm_c_te}{RasterStack}(landscape, ...)

\method{lsm_c_te}{RasterBrick}(landscape, ...)

\method{lsm_c_te}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{Include landscape boundary in edge length}
}
\value{
tibble
}
\description{
Total edge of class (class level)
}
\details{
Total edge equals the sum of the length of all edges of class i. Total edge is an
absolute measure making comparisons among landscapes with different total areas difficult.
??? Landscape border included or not ???
\deqn{TE = sum(edges[class_i])}
\subsection{Units}{Meters}
\subsection{Range}{TE >= 0}
\subsection{Behaviour}{TE increases without limit as landscape becomes more patchy}
}
\examples{
lsm_c_te(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
