% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_calculate.R
\name{lsm_calculate}
\alias{lsm_calculate}
\alias{lsm_calculate}
\alias{lsm_calculate.RasterLayer}
\alias{lsm_calculate}
\alias{lsm_calculate.RasterStack}
\alias{lsm_calculate}
\alias{lsm_calculate.RasterBrick}
\alias{lsm_calculate}
\alias{lsm_calculate.list}
\title{Calculate metrics}
\usage{
lsm_calculate(landscape, what, ...)

\method{lsm_calculate}{RasterLayer}(landscape, what = "all", ...)

\method{lsm_calculate}{RasterStack}(landscape, what = "all", ...)

\method{lsm_calculate}{RasterBrick}(landscape, what = "all", ...)

\method{lsm_calculate}{list}(landscape, what = "all", ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{what}{Selected level of metrics: either "all", "patch", "class", "landscape". The default is "all".}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Calculate a selected group of metrics
}
\examples{
lsm_calculate(landscape)
lsm_calculate(landscape_stack, what = "patch")

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
