% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_cai_cv.R
\name{lsm_l_cai_cv}
\alias{lsm_l_cai_cv}
\alias{lsm_l_cai_cv}
\alias{lsm_l_cai_cv.RasterLayer}
\alias{lsm_l_cai_cv}
\alias{lsm_l_cai_cv.RasterStack}
\alias{lsm_l_cai_cv}
\alias{lsm_l_cai_cv.RasterBrick}
\alias{lsm_l_cai_cv}
\alias{lsm_l_cai_cv.list}
\title{Core area index distribution (landscape level)}
\usage{
lsm_l_cai_cv(landscape, directions)

\method{lsm_l_cai_cv}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_cai_cv}{RasterStack}(landscape, directions = 8)

\method{lsm_l_cai_cv}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_cai_cv}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{???}
}
\value{
tibble
}
\description{
Coefficient of variation of core area index (landscape level)
}
\details{
Equals the coeffiecent of variation of the core area index of all patches in the landscape.
The core area index equals the percentage of a patch that is core area
\deqn{CAI_CV = cv(CAI[patch]}
\subsection{Units}{???}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_cai_cv(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
