% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_core.R
\name{lsm_l_core}
\alias{lsm_l_core}
\alias{lsm_l_core}
\alias{lsm_l_core.RasterLayer}
\alias{lsm_l_core}
\alias{lsm_l_core.RasterStack}
\alias{lsm_l_core}
\alias{lsm_l_core.RasterBrick}
\alias{lsm_l_core}
\alias{lsm_l_core.list}
\title{Total area of core areas (landscape level)}
\usage{
lsm_l_core(landscape)

\method{lsm_l_core}{RasterLayer}(landscape)

\method{lsm_l_core}{RasterStack}(landscape)

\method{lsm_l_core}{RasterBrick}(landscape)

\method{lsm_l_core}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Area of corea area (landscape level)
}
\examples{
lsm_l_core(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
