% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_core_sd.R
\name{lsm_l_core_sd}
\alias{lsm_l_core_sd}
\title{CORE_SD (landscape level)}
\usage{
lsm_l_core_sd(
  landscape,
  directions = 8,
  consider_boundary = FALSE,
  edge_depth = 1
)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick, stars, or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}

\item{consider_boundary}{Logical if cells that only neighbour the landscape
boundary should be considered as core}

\item{edge_depth}{Distance (in cells) a cell has the be away from the patch
edge to be considered as core cell}
}
\value{
tibble
}
\description{
Standard deviation of patch core area (class level)
}
\details{
\deqn{CORE_{SD} = sd(CORE[patch_{ij}])}
where \eqn{CORE[patch_{ij}]} is the core area in square meters of each patch.

CORE_SD is a 'Core area metric'. It equals the standard deviation of the core area
of all patches in the landscape. The core area is defined as all cells that have no
neighbour with a different value than themselves (rook's case). The metric describes the
differences among all patches in the landscape.

\subsection{Units}{Hectares}
\subsection{Range}{CORE_SD >= 0}
\subsection{Behaviour}{Equals CORE_SD = 0 if all patches have the same core area.
Increases, without limit, as the variation of patch core areas increases.}
}
\examples{
lsm_l_core_sd(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_core}},
\code{\link{sd}}, \cr
\code{\link{lsm_c_core_mn}},
\code{\link{lsm_c_core_sd}},
\code{\link{lsm_c_core_cv}}, \cr
\code{\link{lsm_l_core_mn}},
\code{\link{lsm_l_core_cv}}
}
