% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_dcad.R
\name{lsm_l_dcad}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.RasterLayer}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.RasterStack}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.RasterBrick}
\alias{lsm_l_dcad}
\alias{lsm_l_dcad.list}
\title{Disjunct core area density (landscape level)}
\usage{
lsm_l_dcad(landscape)

\method{lsm_l_dcad}{RasterLayer}(landscape)

\method{lsm_l_dcad}{RasterStack}(landscape)

\method{lsm_l_dcad}{RasterBrick}(landscape)

\method{lsm_l_dcad}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Core area divided by landscape area (landscape level)
}
\examples{
lsm_l_dcad(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
