% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_division.R
\name{lsm_l_division}
\alias{lsm_l_division}
\alias{lsm_l_division}
\alias{lsm_l_division.RasterLayer}
\alias{lsm_l_division}
\alias{lsm_l_division.RasterStack}
\alias{lsm_l_division}
\alias{lsm_l_division.RasterBrick}
\alias{lsm_l_division}
\alias{lsm_l_division.list}
\title{Landscape division index (landscape level)}
\usage{
lsm_l_division(landscape)

\method{lsm_l_division}{RasterLayer}(landscape)

\method{lsm_l_division}{RasterStack}(landscape)

\method{lsm_l_division}{RasterBrick}(landscape)

\method{lsm_l_division}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Landscpae division index (landscape level)
}
\details{
The landscape division index equals 1 minus the sum of all patch areas
divided by the total area squared
\deqn{DIVISON = 1 - sum((area[patch] / total area) ^ 2)}
\subsection{Units}{Proportion (???)}
\subsection{Ranges}{0 <= Division < 1}
\subsection{Behaviour}{DIVISION = 1 when only one class and patch is present. Approaches
DIVISON = 1 when all patches of class i are single cells}
}
\examples{
lsm_l_division(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
