% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_division.R
\name{lsm_l_division}
\alias{lsm_l_division}
\alias{lsm_l_division}
\alias{lsm_l_division.RasterLayer}
\alias{lsm_l_division}
\alias{lsm_l_division.RasterStack}
\alias{lsm_l_division}
\alias{lsm_l_division.RasterBrick}
\alias{lsm_l_division}
\alias{lsm_l_division.list}
\title{DIVISION (landscape level)}
\usage{
lsm_l_division(landscape)

\method{lsm_l_division}{RasterLayer}(landscape)

\method{lsm_l_division}{RasterStack}(landscape)

\method{lsm_l_division}{RasterBrick}(landscape)

\method{lsm_l_division}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Landscape division index (Aggregation metric)
}
\details{
\deqn{DIVISON = (1 - \sum \limits_{i = 1}^{m} \sum \limits_{j = 1}^{n} (\frac{a_{ij}} {A}) ^ 2) }
where \eqn{a_{ij}} is the area in square meters and \eqn{A} is the total
landscape area in square meters.

DIVISION is an 'Aggregation metric. It can be in as the probability that two
randomly selected cells are not located in the same patch. The landscape
division index is negativly correlated with the  effective mesh size (\code{\link{lsm_c_mesh}}).

\subsection{Units}{Proportion }
\subsection{Ranges}{0 <= Division < 1}
\subsection{Behaviour}{Equals DIVISION = 0 if only one patch is present. Approaches
DIVISION = 1 if all patches of class i are single cells.}
}
\examples{
lsm_l_division(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{lsm_l_ta}}, \cr
\code{\link{lsm_c_division}}
}
