% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_frac_mn.R
\name{lsm_l_frac_mn}
\alias{lsm_l_frac_mn}
\alias{lsm_l_frac_mn}
\alias{lsm_l_frac_mn.RasterLayer}
\alias{lsm_l_frac_mn}
\alias{lsm_l_frac_mn.RasterStack}
\alias{lsm_l_frac_mn}
\alias{lsm_l_frac_mn.RasterBrick}
\alias{lsm_l_frac_mn}
\alias{lsm_l_frac_mn.list}
\title{FRAC_MN (landscape level)}
\usage{
lsm_l_frac_mn(landscape, directions)

\method{lsm_l_frac_mn}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_frac_mn}{RasterStack}(landscape, directions = 8)

\method{lsm_l_frac_mn}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_frac_mn}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Mean fractal dimension index (Shape metric)
}
\details{
\deqn{FRAC_{MN} = mean(FRAC[patch_{ij}])}
where \eqn{FRAC[patch_{ij}]} equals the fractal dimension index of each patch.

FRAC_MN is a 'Shape metric'. The metric summarises the landscape
as the mean of the fractal dimension index of all patches in the landscape.
The fractal dimenstion index is based on the patch perimeter and
the patch area and describes the patch complexity. The coeffiecent of variation is
scaled to the mean and comparable among different landscapes.

\subsection{Units}{None}
\subsection{Range}{FRAC_MN > 0 }
\subsection{Behaviour}{Approaches FRAC_MN = 1 if all patches are squared and FRAC_MN = 2
if all patches are irregular.}
}
\examples{
lsm_l_frac_mn(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_frac}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_frac_mn}},
\code{\link{lsm_c_frac_sd}},
\code{\link{lsm_c_frac_cv}}, \cr
\code{\link{lsm_l_frac_sd}},
\code{\link{lsm_l_frac_cv}}
}
