% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_gyrate_cv.R
\name{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.RasterLayer}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.RasterStack}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.RasterBrick}
\alias{lsm_l_gyrate_cv}
\alias{lsm_l_gyrate_cv.list}
\title{GYRATE_CV (landscape level)}
\usage{
lsm_l_gyrate_cv(landscape)

\method{lsm_l_gyrate_cv}{RasterLayer}(landscape)

\method{lsm_l_gyrate_cv}{RasterStack}(landscape)

\method{lsm_l_gyrate_cv}{RasterBrick}(landscape)

\method{lsm_l_gyrate_cv}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Coeffiecent of variation radius of gyration (Area and edge metric)
}
\details{
\deqn{GYRATE_{CV} = cv(GYRATE[patch_{ij}])}
where \eqn{GYRATE[patch_{ij}]} equals the radius of gyration of each patch.

GYRATE_CV is an 'Area and edge metric'. The metric summarises the landscape
as the coeffiecent of variation of the radius of gyration of all patches
in the landscape. GYRATE measures the distance from each cell to the patch
centroid and is based on cell center-to-cell center distances. The metrics characterises
both the patch area and compactness. The coeffiecent of variation is
scaled to the mean and comparable among different landscapes.

\subsection{Units}{Meters}
\subsection{Range}{GYRATE_CV >= 0 }
\subsection{Behaviour}{Equals GYRATE_CV = 0 if the radius of gyration is identical
for all patches. Increases, without limit, as the variation of the radius of gyration
increases.}
}
\examples{
lsm_l_gyrate_cv(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_gyrate}},
\code{\link{cv}}, \cr
\code{\link{lsm_c_gyrate_mn}},
\code{\link{lsm_c_gyrate_sd}},
\code{\link{lsm_c_gyrate_cv}}, \cr
\code{\link{lsm_l_gyrate_mn}},
\code{\link{lsm_l_gyrate_sd}}
}
