% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_lpi.R
\name{lsm_l_lpi}
\alias{lsm_l_lpi}
\alias{lsm_l_lpi}
\alias{lsm_l_lpi.RasterLayer}
\alias{lsm_l_lpi}
\alias{lsm_l_lpi.RasterStack}
\alias{lsm_l_lpi}
\alias{lsm_l_lpi.RasterBrick}
\alias{lsm_l_lpi}
\alias{lsm_l_lpi.list}
\title{Largest patch index (landscape level)}
\usage{
lsm_l_lpi(landscape)

\method{lsm_l_lpi}{RasterLayer}(landscape)

\method{lsm_l_lpi}{RasterStack}(landscape)

\method{lsm_l_lpi}{RasterBrick}(landscape)

\method{lsm_l_lpi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Area of the larges patch divided by the total area (landscape level)
}
\examples{
lsm_l_lpi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
