% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_lsi.R
\name{lsm_l_lsi}
\alias{lsm_l_lsi}
\alias{lsm_l_lsi}
\alias{lsm_l_lsi.RasterLayer}
\alias{lsm_l_lsi}
\alias{lsm_l_lsi.RasterStack}
\alias{lsm_l_lsi}
\alias{lsm_l_lsi.RasterBrick}
\alias{lsm_l_lsi}
\alias{lsm_l_lsi.list}
\title{Landscape shape index (landscape level)}
\usage{
lsm_l_lsi(landscape, count_boundary = FALSE)

\method{lsm_l_lsi}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_l_lsi}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_l_lsi}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_l_lsi}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{xxx}
}
\value{
tibble
}
\description{
Landscape shape index (landscape level)
}
\details{
The landscape shape index equals a quarter of the sum of all edges in the landscape
divided by the square root of the total area.
\deqn{LSI = (0.25 * sum(edges[patch])) / sqrt(total area)}
\subsection{Units}{none}
\subsection{Ranges}{LSI >= 1}
\subsection{Behaviour}{Equals LSI = 1 when only one class and patch is present and
increases when the length of edges increases, i.e. the patches of class i become more complex}
}
\examples{
lsm_l_lsi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
