% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_mesh.R
\name{lsm_l_mesh}
\alias{lsm_l_mesh}
\alias{lsm_l_mesh}
\alias{lsm_l_mesh.RasterLayer}
\alias{lsm_l_mesh}
\alias{lsm_l_mesh.RasterStack}
\alias{lsm_l_mesh}
\alias{lsm_l_mesh.RasterBrick}
\alias{lsm_l_mesh}
\alias{lsm_l_mesh.list}
\title{Effective mesh size (landscape level)}
\usage{
lsm_l_mesh(landscape)

\method{lsm_l_mesh}{RasterLayer}(landscape)

\method{lsm_l_mesh}{RasterStack}(landscape)

\method{lsm_l_mesh}{RasterBrick}(landscape)

\method{lsm_l_mesh}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Effective mesh size (landscape level)
}
\examples{
lsm_l_mesh(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
