% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_msidi.R
\name{lsm_l_msidi}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterLayer}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterStack}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterBrick}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.list}
\title{MSIDI (landscape level)}
\usage{
lsm_l_msidi(landscape)

\method{lsm_l_msidi}{RasterLayer}(landscape)

\method{lsm_l_msidi}{RasterStack}(landscape)

\method{lsm_l_msidi}{RasterBrick}(landscape)

\method{lsm_l_msidi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Modified Simpson's diversity index (Diversity metric)
}
\details{
\deqn{MSIDI = -\ln \sum \limits_{i = 1}^{m} P_{i}^{2}}
where \eqn{P_{i}} is the landscape area proportion of class i.

MSIDI is a 'Diversity metric'.

\subsection{Units}{None}
\subsection{Range}{MSIDI >= 0}
\subsection{Behaviour}{MSIDI = 0 if only one patch is present and increases, without
limit, as the amount of patches with equally distributed landscape proportions increases}
}
\examples{
lsm_l_msidi(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_l_sidi}}
}
