% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_msidi.R
\name{lsm_l_msidi}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterLayer}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterStack}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.RasterBrick}
\alias{lsm_l_msidi}
\alias{lsm_l_msidi.list}
\title{Modified Simpson's diversity index (landscape level)}
\usage{
lsm_l_msidi(landscape)

\method{lsm_l_msidi}{RasterLayer}(landscape)

\method{lsm_l_msidi}{RasterStack}(landscape)

\method{lsm_l_msidi}{RasterBrick}(landscape)

\method{lsm_l_msidi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Modified Simpson's diversity index (landscape level)
}
\details{
The modified Simpson's diversity index equals the negative natural logarithm of
the sum of the squared proportions of all patches in the landscape grouped by class
\deqn{MSIDI = -ln(sum(proportion[patch_i] ^ 2))}
\subsection{Units}{None}
\subsection{Range}{MSIDI >= 1}
\subsection{Behaviour}{MSIDI = 0 when only one class and patch is present and increases
without limit as the amount of patches with equally distributed landscape proportions increases}
}
\examples{
lsm_l_msidi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
