% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_msiei.R
\name{lsm_l_msiei}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterLayer}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterStack}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterBrick}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.list}
\title{Modified Simpson's evenness index (landscape level)}
\usage{
lsm_l_msiei(landscape)

\method{lsm_l_msiei}{RasterLayer}(landscape)

\method{lsm_l_msiei}{RasterStack}(landscape)

\method{lsm_l_msiei}{RasterBrick}(landscape)

\method{lsm_l_msiei}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Modified Simpson's evenness index (landscape level)
}
\examples{
lsm_l_msiei(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
