% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_msiei.R
\name{lsm_l_msiei}
\alias{lsm_l_msiei}
\alias{lsm_l_msiei.RasterLayer}
\alias{lsm_l_msiei.RasterStack}
\alias{lsm_l_msiei.RasterBrick}
\alias{lsm_l_msiei.stars}
\alias{lsm_l_msiei.list}
\title{MSIEI (landscape level)}
\usage{
lsm_l_msiei(landscape, directions)

\method{lsm_l_msiei}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_msiei}{RasterStack}(landscape, directions = 8)

\method{lsm_l_msiei}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_msiei}{stars}(landscape, directions = 8)

\method{lsm_l_msiei}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Modified Simpson's evenness index (Diversity metric)
}
\details{
\deqn{MSIEi = \frac{- \ln \sum \limits_{i = 1}^{m} P_{i}^{2}} {\ln m}}
where \eqn{P_{i}} is the landscape area proportion of class i.
MSIEI is a 'Diversity metric'.
\subsection{Units}{None}
\subsection{Range}{0 <= MSIEI < 1}
\subsection{Behaviour}{MSIEI = 0 when only one patch is present and approaches
MSIEI = 1 as the proportional distribution of patches becomes more even}
}
\examples{
lsm_l_msiei(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Simpson, E. H. 1949. Measurement of diversity. Nature 163:688

Pielou, E. C. 1975. Ecological Diversity. Wiley-Interscience, New York.
}
\seealso{
\code{\link{lsm_l_siei}}
}
