% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ncore.R
\name{lsm_l_ncore}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.RasterLayer}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.RasterStack}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.RasterBrick}
\alias{lsm_l_ncore}
\alias{lsm_l_ncore.list}
\title{Number of core areas (landscape level)}
\usage{
lsm_l_ncore(landscape, ...)

\method{lsm_l_ncore}{RasterLayer}(landscape, ...)

\method{lsm_l_ncore}{RasterStack}(landscape, ...)

\method{lsm_l_ncore}{RasterBrick}(landscape, ...)

\method{lsm_l_ncore}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Number of core areas (landscape level)
}
\details{
Number of core areas equals the sum of number of core areas of all patches in the
landscape. Called number of disjunct core areas in FRAGSTATS. A core area is a
'patch within the patch' without any edge cells. In other words, the number of
patches within the patch that only have neighbouring cells of the same type
\deqn{NCORE = sum(ncore[patch])}
\subsection{Units}{None}
\subsection{Range}{NCORE >= 0}
\subsection{Behaviour}{NCORE = 0 when CORE = 0, i.e. every cell in patches of class i is
an edge. NCORE increases without limit as core areas become more present, i.e. patches
becoming larger and less complex}
}
\examples{
lsm_l_ncore(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
