% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ncore_mn.R
\name{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.RasterLayer}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.RasterStack}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.RasterBrick}
\alias{lsm_l_ncore_mn}
\alias{lsm_l_ncore_mn.list}
\title{Number of core areas distribution (landscape level)}
\usage{
lsm_l_ncore_mn(landscape, directions)

\method{lsm_l_ncore_mn}{RasterLayer}(landscape, directions = 8)

\method{lsm_l_ncore_mn}{RasterStack}(landscape, directions = 8)

\method{lsm_l_ncore_mn}{RasterBrick}(landscape, directions = 8)

\method{lsm_l_ncore_mn}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{direction}{Specific arguments for certain functions, if not provided they fall back to default.}
}
\value{
tibble
}
\description{
Mean of number of core areas (landscape level)
}
\details{
Equals the mean of number of core area of all patches in the landscape.
A core area is a 'patch within the patch' without any edge cells. In other words,
the number of patches within the patch that only have neighbouring cells of the same type
\deqn{NCORE_MN = mean(NCORE[patch])}
\subsection{Units}{None}
\subsection{Range}{???}
\subsection{Behaviour}{???}
}
\examples{
lsm_l_ncore_mn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
