% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_np.R
\name{lsm_l_np}
\alias{lsm_l_np}
\alias{lsm_l_np}
\alias{lsm_l_np.RasterLayer}
\alias{lsm_l_np}
\alias{lsm_l_np.RasterStack}
\alias{lsm_l_np}
\alias{lsm_l_np.RasterBrick}
\alias{lsm_l_np}
\alias{lsm_l_np.list}
\title{Number of patches}
\usage{
lsm_l_np(landscape, class)

\method{lsm_l_np}{RasterLayer}(landscape)

\method{lsm_l_np}{RasterStack}(landscape, class)

\method{lsm_l_np}{RasterBrick}(landscape, class)

\method{lsm_l_np}{list}(landscape, class)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of patches
}
\examples{
lsm_l_np(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
