% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_np.R
\name{lsm_l_np}
\alias{lsm_l_np}
\alias{lsm_l_np}
\alias{lsm_l_np.RasterLayer}
\alias{lsm_l_np}
\alias{lsm_l_np.RasterStack}
\alias{lsm_l_np}
\alias{lsm_l_np.RasterBrick}
\alias{lsm_l_np}
\alias{lsm_l_np.list}
\title{Number of patches (landscape level)}
\usage{
lsm_l_np(landscape)

\method{lsm_l_np}{RasterLayer}(landscape)

\method{lsm_l_np}{RasterStack}(landscape)

\method{lsm_l_np}{RasterBrick}(landscape)

\method{lsm_l_np}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of patches (landscape level)
}
\details{
Number of patches equals the number of patches in the landscape
\subsection{Units}{None}
\subsection{Ranges}{NP >= 1}
\subsection{Behaviour}{NP = 1 when only one class and patch is present and
increases without limit as the number of patches increases}
}
\examples{
lsm_l_np(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
