% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_pafrac.R
\name{lsm_l_pafrac}
\alias{lsm_l_pafrac}
\alias{lsm_l_pafrac}
\alias{lsm_l_pafrac.RasterLayer}
\alias{lsm_l_pafrac}
\alias{lsm_l_pafrac.RasterStack}
\alias{lsm_l_pafrac}
\alias{lsm_l_pafrac.RasterBrick}
\alias{lsm_l_pafrac}
\alias{lsm_l_pafrac.list}
\title{Perimeter-Area Fractal Dimension  (landscape level)}
\usage{
lsm_l_pafrac(landscape)

\method{lsm_l_pafrac}{RasterLayer}(landscape)

\method{lsm_l_pafrac}{RasterStack}(landscape)

\method{lsm_l_pafrac}{RasterBrick}(landscape)

\method{lsm_l_pafrac}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Perimeter-Area Fractal Dimension (landscape level)
}
\details{
The perimeter-area fractal dimension equals two divided by the slope of
the regression line of the natural logrithm of patch area against the
natural logrithm of the patch perimeter of all patches in the landscape.
The regression has equation ln(area) = beta * ln(perimeter) + intercept.
It is only meaningful if the relationship between the area and perimeter.
is linear on a logarithmic scale. If there are less than 10 patches,
the functions returns NA because of the small-sample issue
\deqn{PAFRAC = 2 / (beta(ln(area[patch]) ~ ln(perimeter[patch]))}
\subsection{Units}{None}
\subsection{Range}{1 <= PAFRAC <= 2}
\subsection{Behaviour}{If only a few patches are present the value
can exceed the range. Approaches PAFRAC = 1 for patches with simples shapes and
PAFRAC = 2 for irregular shapes}
}
\examples{
lsm_l_pafrac(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
