% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_pafrac.R
\name{lsm_l_pafrac}
\alias{lsm_l_pafrac}
\title{PAFRAC  (landscape level)}
\usage{
lsm_l_pafrac(landscape, directions = 8, verbose = TRUE)
}
\arguments{
\item{landscape}{A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.}

\item{directions}{The number of directions in which patches should be
connected: 4 (rook's case) or 8 (queen's case).}

\item{verbose}{Print warning message if not sufficient patches are present}
}
\value{
tibble
}
\description{
Perimeter-Area Fractal Dimension (Shape metric)
}
\details{
\deqn{PAFRAC = \frac{2}{\beta}}
where \eqn{\beta} is the slope of the regression of the area against the perimeter
(logarithm) \eqn{N \sum \limits_{i = 1}^{m} \sum \limits_{j = 1}^{n} \ln a_{ij} = a + \beta N \sum \limits_{i = 1}^{m} \sum \limits_{j = 1}^{n} \ln p_{ij}}

PAFRAC is a 'Shape metric'. It describes the patch complexity of the landscape while being
scale independent. This means that increasing the patch size while not changing the
patch form will not change the metric. However, it is only meaningful if the relationship
between the area and perimeter is linear on a logarithmic scale. Furthermore, if there
are less than 10 patches in the landscape, the metric returns NA because of the small-sample
issue.

\subsection{Units}{None}
\subsection{Range}{1 <= PAFRAC <= 2}
\subsection{Behaviour}{Approaches PAFRAC = 1 for patches with simple shapes and
approaches PAFRAC = 2 for irregular shapes}
}
\examples{
landscape <- terra::rast(landscapemetrics::landscape)
lsm_l_pafrac(landscape)

}
\references{
McGarigal K., SA Cushman, and E Ene. 2023. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical Maps. Computer software program produced by the authors;
available at the following web site: https://www.fragstats.org

Burrough, P. A. 1986. Principles of Geographical Information Systems for
Land Resources Assessment. Monographs on Soil and Resources Survey No. 12.
Clarendon Press, Oxford
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{lsm_p_perim}}, \cr
\code{\link{lsm_c_pafrac}}
}
