% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_para_mn.R
\name{lsm_l_para_mn}
\alias{lsm_l_para_mn}
\alias{lsm_l_para_mn}
\alias{lsm_l_para_mn.RasterLayer}
\alias{lsm_l_para_mn}
\alias{lsm_l_para_mn.RasterStack}
\alias{lsm_l_para_mn}
\alias{lsm_l_para_mn.RasterBrick}
\alias{lsm_l_para_mn}
\alias{lsm_l_para_mn.list}
\title{PARA_MN (landscape level)}
\usage{
lsm_l_para_mn(landscape)

\method{lsm_l_para_mn}{RasterLayer}(landscape)

\method{lsm_l_para_mn}{RasterStack}(landscape)

\method{lsm_l_para_mn}{RasterBrick}(landscape)

\method{lsm_l_para_mn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Mean perimeter-area ratio (class level)
}
\details{
\deqn{PARA_{MN} = mean(PARA[patch_{ij}]}
where \eqn{PARA[patch_{ij}]} is the perimeter area ratio of each patch.

PARA_MN is a 'Shape metric'. It summarises the landscape as the mean of
each patch in the landscape. The perimeter-area ratio describes the patch complexity
in a straightforward way. However, because it is not standarised to a certain shape
(e.g. a square), it is not scale independent, meaning that increasing the patch size
while not changing the patch form will change the ratio.

\subsection{Units}{None}
\subsection{Range}{PARA_MN > 0}
\subsection{Behaviour}{Approaches PARA_MN > 0 if PARA for each patch approaches PARA > 0,
i.e. the form approaches a rather small square. Increases, without limit, as PARA increases,
i.e. patches become more complex.}
}
\examples{
lsm_l_para_mn(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_para}},
\code{\link{mean}}, \cr
\code{\link{lsm_c_para_mn}},
\code{\link{lsm_c_para_sd}},
\code{\link{lsm_c_para_cv}}, \cr
\code{\link{lsm_l_para_sd}},
\code{\link{lsm_l_para_cv}}
}
