% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_para_sd.R
\name{lsm_l_para_sd}
\alias{lsm_l_para_sd}
\alias{lsm_l_para_sd}
\alias{lsm_l_para_sd.RasterLayer}
\alias{lsm_l_para_sd}
\alias{lsm_l_para_sd.RasterStack}
\alias{lsm_l_para_sd}
\alias{lsm_l_para_sd.RasterBrick}
\alias{lsm_l_para_sd}
\alias{lsm_l_para_sd.list}
\title{PARA_SD (landscape level)}
\usage{
lsm_l_para_sd(landscape)

\method{lsm_l_para_sd}{RasterLayer}(landscape)

\method{lsm_l_para_sd}{RasterStack}(landscape)

\method{lsm_l_para_sd}{RasterBrick}(landscape)

\method{lsm_l_para_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Standard deviation perimeter-area ratio (class level)
}
\details{
\deqn{PARA_{SD} = sd(PARA[patch_{ij}]}
where \eqn{PARA[patch_{ij}]} is the perimeter area ratio of each patch.

PARA_SD is a 'Shape metric'. It summarises the landscape as the standard deviation of
each patch belonging in the landscape. The perimeter-area ratio describes the patch complexity
in a straightforward way. However, because it is not standarised to a certain shape
(e.g. a square), it is not scale independent, meaning that increasing the patch size
while not changing the patch form will change the ratio.

\subsection{Units}{None}
\subsection{Range}{PARA_SD >= 0}
\subsection{Behaviour}{Equals PARA_SD = 0 if the perimeter-area ratio is identical for
all patches. Increases, without limit, as the variation of the perimeter-area ratio
increases.}
}
\examples{
lsm_l_para_sd(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_p_para}},
\code{\link{sd}}, \cr
\code{\link{lsm_c_para_mn}},
\code{\link{lsm_c_para_sd}},
\code{\link{lsm_c_para_cv}}, \cr
\code{\link{lsm_l_para_mn}},
\code{\link{lsm_l_para_cv}}
}
