% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_prd.R
\name{lsm_l_prd}
\alias{lsm_l_prd}
\alias{lsm_l_prd}
\alias{lsm_l_prd.RasterLayer}
\alias{lsm_l_prd}
\alias{lsm_l_prd.RasterStack}
\alias{lsm_l_prd}
\alias{lsm_l_prd.RasterBrick}
\alias{lsm_l_prd}
\alias{lsm_l_prd.list}
\title{Patch richness density}
\usage{
lsm_l_prd(landscape)

\method{lsm_l_prd}{RasterLayer}(landscape)

\method{lsm_l_prd}{RasterStack}(landscape)

\method{lsm_l_prd}{RasterBrick}(landscape)

\method{lsm_l_prd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Number of patch types (classes) in the landscape divided by the total
landscape area
}
\examples{
lsm_l_prd(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
