% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_rpr.R
\name{lsm_l_rpr}
\alias{lsm_l_rpr}
\title{Relative patch ritchness}
\usage{
lsm_l_rpr(landscape, classes_max = NULL)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{classes_max}{Maximum number of classes}
}
\value{
Value >= 1
}
\description{
Number of different classes divided by (potential) maximum number of classes
}
\examples{
lsm_l_rpr(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
