% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shape_sd.R
\name{lsm_l_shape_sd}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.RasterLayer}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.RasterStack}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.RasterBrick}
\alias{lsm_l_shape_sd}
\alias{lsm_l_shape_sd.list}
\title{Shape index  (class level)}
\usage{
lsm_l_shape_sd(landscape)

\method{lsm_l_shape_sd}{RasterLayer}(landscape)

\method{lsm_l_shape_sd}{RasterStack}(landscape)

\method{lsm_l_shape_sd}{RasterBrick}(landscape)

\method{lsm_l_shape_sd}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Perimeter divided by squareroot of area
}
\examples{
lsm_l_shape_sd(landscape)
lsm_l_shape_sd(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
