% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shdi.R
\name{lsm_l_shdi}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.RasterLayer}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.RasterStack}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.RasterBrick}
\alias{lsm_l_shdi}
\alias{lsm_l_shdi.list}
\title{Shannon's diversity index (landscape level)}
\usage{
lsm_l_shdi(landscape)

\method{lsm_l_shdi}{RasterLayer}(landscape)

\method{lsm_l_shdi}{RasterStack}(landscape)

\method{lsm_l_shdi}{RasterBrick}(landscape)

\method{lsm_l_shdi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Shannon's diversity index (landscape level)
}
\details{
Shannon's diversity index equals minus the sum of class proportions multipled by the natural
logarithm of class proportions
\deqn{SHDI = sum(proportion[patch_i] * log(proportion[patch_i]))}
\subsection{Units}{None}
\subsection{Range}{SHDI >= 0}
\subsection{Behaviour}{SHDI = 0 when only one patch and class is present and increases
without limit as the number of classes increases while the proportions are
equally distributed}
}
\examples{
lsm_l_shdi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
