% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shdi.R
\name{lsm_l_shdi}
\alias{lsm_l_shdi}
\title{Shannon's Diversity Index}
\usage{
lsm_l_shdi(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 0
}
\description{
Number of patch types (classes) in the landscape
}
\examples{
lsm_l_shdi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
