% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shei.R
\name{lsm_l_shei}
\alias{lsm_l_shei}
\alias{lsm_l_shei}
\alias{lsm_l_shei.RasterLayer}
\alias{lsm_l_shei}
\alias{lsm_l_shei.RasterStack}
\alias{lsm_l_shei}
\alias{lsm_l_shei.RasterBrick}
\alias{lsm_l_shei}
\alias{lsm_l_shei.list}
\title{SHEI (landscape level)}
\usage{
lsm_l_shei(landscape)

\method{lsm_l_shei}{RasterLayer}(landscape)

\method{lsm_l_shei}{RasterStack}(landscape)

\method{lsm_l_shei}{RasterBrick}(landscape)

\method{lsm_l_shei}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Shannons's evenness index (Diversity metric)
}
\details{
\deqn{SHEI = \frac{- \sum \limits_{i = 1} ^ {m} (P_{i} * \ln P_{i})} {\ln m}}
where \eqn{P_{i}} is the proportion of class i and \eqn{m} is the
number of classes.

SHEI is a 'Diversity metric'. It is the ratio between the actual Shannon's diversity index
and and the theoretical maximum of the Shannon diversity index. It can be understood as a
measure of dominance.

\subsection{Units}{None}
\subsection{Range}{0 <= SHEI < 1}
\subsection{Behaviour}{Equals SHEI = 0 when only one patch  present and equals SHEI = 1
when the proportion of classes is completely equally distributed}
}
\examples{
lsm_l_shei(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_c_pland}},
\code{\link{lsm_l_pr}}
}
