% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_shei.R
\name{lsm_l_shei}
\alias{lsm_l_shei}
\title{Shannon's Evenness Index}
\usage{
lsm_l_shei(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Evenness of patch types (classes) in the landscape
}
\examples{
lsm_l_shei(landscape)
lsm_l_shei(landscape_stack)
lsm_l_shei(as.list(landscape_stack))

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
