% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_sidi.R
\name{lsm_l_sidi}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.RasterLayer}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.RasterStack}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.RasterBrick}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.list}
\title{Simpson's diversity index (landscape level)}
\usage{
lsm_l_sidi(landscape)

\method{lsm_l_sidi}{RasterLayer}(landscape)

\method{lsm_l_sidi}{RasterStack}(landscape)

\method{lsm_l_sidi}{RasterBrick}(landscape)

\method{lsm_l_sidi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Simpson's diversity index (landscape level)
}
\details{
Simpson's diversity index equals 1 minus the sum of the squared class proportions
\deqn{SIDI = 1 - sum(proportion[patch_i] ^ 2)}
\subsection{Units}{None}
\subsection{Range}{0 <= SIDI < 1}
\subsection{Behaviour}{Simpson's diversity index approaches SIDI = 0 when only one patch and
class is present and approaches SIDI = 1 when the number of class types increases while the
proportions are equally distributed}
}
\examples{
lsm_l_sidi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
