% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_sidi.R
\name{lsm_l_sidi}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.RasterLayer}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.RasterStack}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.RasterBrick}
\alias{lsm_l_sidi}
\alias{lsm_l_sidi.list}
\title{Simpson's diversity index (landscape level)}
\usage{
lsm_l_sidi(landscape)

\method{lsm_l_sidi}{RasterLayer}(landscape)

\method{lsm_l_sidi}{RasterStack}(landscape)

\method{lsm_l_sidi}{RasterBrick}(landscape)

\method{lsm_l_sidi}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Simpson's diversity index (landscape level)
}
\examples{
lsm_l_sidi(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
