% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_siei.R
\name{lsm_l_siei}
\alias{lsm_l_siei}
\alias{lsm_l_siei}
\alias{lsm_l_siei.RasterLayer}
\alias{lsm_l_siei}
\alias{lsm_l_siei.RasterStack}
\alias{lsm_l_siei}
\alias{lsm_l_siei.RasterBrick}
\alias{lsm_l_siei}
\alias{lsm_l_siei.list}
\title{Simpson's evenness index (landscape level)}
\usage{
lsm_l_siei(landscape)

\method{lsm_l_siei}{RasterLayer}(landscape)

\method{lsm_l_siei}{RasterStack}(landscape)

\method{lsm_l_siei}{RasterBrick}(landscape)

\method{lsm_l_siei}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
Value >= 1
}
\description{
Simpson's evenness index (landscape level)
}
\examples{
lsm_l_siei(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
