% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_ta.R
\name{lsm_l_ta}
\alias{lsm_l_ta}
\title{Total class area (landcape level)}
\usage{
lsm_l_ta(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total area (landscape level)
}
\examples{
lsm_ta(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
