% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_te.R
\name{lsm_l_te}
\alias{lsm_l_te}
\title{Total Edge (TE)}
\usage{
lsm_l_te(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total Edge
}
\examples{
lsm_l_te(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
