% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_l_te.R
\name{lsm_l_te}
\alias{lsm_l_te}
\alias{lsm_l_te}
\alias{lsm_l_te.RasterLayer}
\alias{lsm_l_te}
\alias{lsm_l_te.RasterStack}
\alias{lsm_l_te}
\alias{lsm_l_te.RasterBrick}
\alias{lsm_l_te}
\alias{lsm_l_te.list}
\title{Total Edge  (landscape level)}
\usage{
lsm_l_te(landscape, count_boundary)

\method{lsm_l_te}{RasterLayer}(landscape, count_boundary = FALSE)

\method{lsm_l_te}{RasterStack}(landscape, count_boundary = FALSE)

\method{lsm_l_te}{RasterBrick}(landscape, count_boundary = FALSE)

\method{lsm_l_te}{list}(landscape, count_boundary = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{count_boundary}{???}
}
\value{
tibble
}
\description{
Total edge of class (landscape level)
}
\details{
Total edge equals the sum of the length of all edges in the landscape.
Total edge is an absolute measure making comparisons among landscapes with
different total areas difficult. ??? Landscape border included or not ???
\deqn{TE = sum(edges)}
\subsection{Units}{Meters}
\subsection{Range}{TE >= 0}
\subsection{Behaviour}{TE increases without limit as landscape becomes more patchy}
}
\examples{
lsm_l_te(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
