% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_area.R
\name{lsm_p_area}
\alias{lsm_p_area}
\alias{lsm_p_area}
\alias{lsm_p_area.RasterLayer}
\alias{lsm_p_area}
\alias{lsm_p_area.RasterStack}
\alias{lsm_p_area}
\alias{lsm_p_area.RasterBrick}
\alias{lsm_p_area}
\alias{lsm_p_area.list}
\title{Area  (patch level)}
\usage{
lsm_p_area(landscape)

\method{lsm_p_area}{RasterLayer}(landscape)

\method{lsm_p_area}{RasterStack}(landscape)

\method{lsm_p_area}{RasterBrick}(landscape)

\method{lsm_p_area}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Area of a patch (patch level)
}
\details{
Equals the area of the patch
\subsection{Units}{Square meter (assuming that the input cellsize is in meter)}
\subsection{Range}{AREA > 0 and increases without limit as the patch size increases}
}
\examples{
lsm_p_area(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
