% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_cai.R
\name{lsm_p_cai}
\alias{lsm_p_cai}
\alias{lsm_p_cai}
\alias{lsm_p_cai.RasterLayer}
\alias{lsm_p_cai}
\alias{lsm_p_cai.RasterStack}
\alias{lsm_p_cai}
\alias{lsm_p_cai.RasterBrick}
\alias{lsm_p_cai}
\alias{lsm_p_cai.list}
\title{Core area index  (patch level)}
\usage{
lsm_p_cai(landscape)

\method{lsm_p_cai}{RasterLayer}(landscape)

\method{lsm_p_cai}{RasterStack}(landscape)

\method{lsm_p_cai}{RasterBrick}(landscape)

\method{lsm_p_cai}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Core area divided by pathch area (patch level)
}
\examples{
lsm_p_cai(landscape)
lsm_p_cai(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
