% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_core.R
\name{lsm_p_core}
\alias{lsm_p_core}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterLayer}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterStack}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterBrick}
\alias{lsm_p_core}
\alias{lsm_p_core.list}
\title{Area of core areas  (patch level)}
\usage{
lsm_p_core(landscape)

\method{lsm_p_core}{RasterLayer}(landscape)

\method{lsm_p_core}{RasterStack}(landscape)

\method{lsm_p_core}{RasterBrick}(landscape)

\method{lsm_p_core}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Area of corea area
}
\examples{
lsm_p_core(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
