% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_core.R
\name{lsm_p_core}
\alias{lsm_p_core}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterLayer}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterStack}
\alias{lsm_p_core}
\alias{lsm_p_core.RasterBrick}
\alias{lsm_p_core}
\alias{lsm_p_core.list}
\title{Core area (patch level)}
\usage{
lsm_p_core(landscape, directions)

\method{lsm_p_core}{RasterLayer}(landscape, directions = 4)

\method{lsm_p_core}{RasterStack}(landscape, directions = 4)

\method{lsm_p_core}{RasterBrick}(landscape, directions = 4)

\method{lsm_p_core}{list}(landscape, directions = 4)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{directions}{...}
}
\value{
tibble
}
\description{
Area of core area of patch (patch level)
}
\details{
Equals the area within a patch that is not on the edge of the patch. In other words,
the area of a patch that has only neighbouring cells of the same type
\subsection{Units}{Square meter (assuming that the input cellsize is in meter)}
\subsection{Range}{CORE >= 0 and increases without limit as patch area increases
and patch shape simplifies. CORE = 0 when every cell in the patch is an edge}
}
\examples{
lsm_p_core(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
