% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_enn.R
\name{lsm_p_enn}
\alias{lsm_p_enn}
\alias{lsm_p_enn}
\alias{lsm_p_enn.RasterLayer}
\alias{lsm_p_enn}
\alias{lsm_p_enn.RasterStack}
\alias{lsm_p_enn}
\alias{lsm_p_enn.RasterBrick}
\alias{lsm_p_enn}
\alias{lsm_p_enn.list}
\title{Euclidean Nearest-Neighbor Distance (ENN)}
\usage{
lsm_p_enn(landscape)

\method{lsm_p_enn}{RasterLayer}(landscape)

\method{lsm_p_enn}{RasterStack}(landscape)

\method{lsm_p_enn}{RasterBrick}(landscape)

\method{lsm_p_enn}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Euclidean Nearest-Neighbor Distance of patch (patch level)
}
\details{
Equals the distance the the nearest neigbouring patch of the same patch type (shortest
edge-to-edge distance)
\subsection{Units}{Meters}
\subsection{Range}{ENN > 0}
\subsection{Behaviour}{???}
}
\examples{
lsm_p_enn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
