% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_enn.R
\name{lsm_p_enn}
\alias{lsm_p_enn}
\alias{lsm_p_enn}
\alias{lsm_p_enn.RasterLayer}
\title{Euclidean Nearest Neighbor Distance Distribution (ENN_MN)}
\usage{
lsm_p_enn(landscape)

\method{lsm_p_enn}{RasterLayer}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Euclidean Nearest Neighbor Distance Distribution (class level)
}
\examples{
lsm_p_enn(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
