% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_frac.R
\name{lsm_p_frac}
\alias{lsm_p_frac}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterLayer}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterStack}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterBrick}
\alias{lsm_p_frac}
\alias{lsm_p_frac.list}
\title{Perimeter area ratio  (patch level)}
\usage{
lsm_p_frac(landscape)

\method{lsm_p_frac}{RasterLayer}(landscape)

\method{lsm_p_frac}{RasterStack}(landscape)

\method{lsm_p_frac}{RasterBrick}(landscape)

\method{lsm_p_frac}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Ratio of patch perimeter and area
}
\examples{
lsm_p_frac(landscape)
lsm_p_frac(landscape_stack)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
