% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_frac.R
\name{lsm_p_frac}
\alias{lsm_p_frac}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterLayer}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterStack}
\alias{lsm_p_frac}
\alias{lsm_p_frac.RasterBrick}
\alias{lsm_p_frac}
\alias{lsm_p_frac.list}
\title{Fractal dimension index  (patch level)}
\usage{
lsm_p_frac(landscape)

\method{lsm_p_frac}{RasterLayer}(landscape)

\method{lsm_p_frac}{RasterStack}(landscape)

\method{lsm_p_frac}{RasterBrick}(landscape)

\method{lsm_p_frac}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Fractacl dimension index of patch (patch level)
}
\details{
The fractal dimension index equals two times the natural logarithm of a
quarter of the patch area divided by the natural logarithm of the patch area
\deqn{FRAC = 2 * ln(0.25 * perimeter[patch]) / ln(area[patch])}
\subsection{Units}{None}
\subsection{Range}{1 <= FRAC <= 2 \cr The fractal dimension index approaches FRAC = 1 for
simple patches and FRAC = 2 for irregular patches}
}
\examples{
lsm_p_frac(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
