% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_gyrate.R
\name{lsm_p_gyrate}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.RasterLayer}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.RasterStack}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.RasterBrick}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.list}
\title{GYRATE (patch level)}
\usage{
lsm_p_gyrate(landscape)

\method{lsm_p_gyrate}{RasterLayer}(landscape)

\method{lsm_p_gyrate}{RasterStack}(landscape)

\method{lsm_p_gyrate}{RasterBrick}(landscape)

\method{lsm_p_gyrate}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Radius of Gyration (Area and edge metric)
}
\details{
\deqn{GYRATE = \sum \limits_{r = 1}^{z} \frac{h_{ijr}} {z}}
where \eqn{h_{ijr}} is the distance from each cell to the centroid of the patch
and \eqn{z} is the number of cells.

GYRATE is an 'Area and edge metric'. The distance from each cell to the patch
centroid is based on cell center-to-cell center distances. The metrics characterises
both the patch area and compactness.

\subsection{Units}{Meters}
\subsection{Range}{GYRATE >= 0}
\subsection{Behaviour}{Approaches GYRATE = 0 if patch is a single cell. Increases,
without limit, when only one patch is present.}
}
\examples{
lsm_p_gyrate(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
\seealso{
\code{\link{lsm_c_gyrate_mn}},
\code{\link{lsm_c_gyrate_sd}},
\code{\link{lsm_c_gyrate_cv}}, \cr
\code{\link{lsm_l_gyrate_mn}},
\code{\link{lsm_l_gyrate_sd}},
\code{\link{lsm_l_gyrate_cv}}
}
