% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_gyrate.R
\name{lsm_p_gyrate}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.RasterLayer}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.RasterStack}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.RasterBrick}
\alias{lsm_p_gyrate}
\alias{lsm_p_gyrate.list}
\title{Radius of Gyration (GYRATE)}
\usage{
lsm_p_gyrate(landscape)

\method{lsm_p_gyrate}{RasterLayer}(landscape)

\method{lsm_p_gyrate}{RasterStack}(landscape)

\method{lsm_p_gyrate}{RasterBrick}(landscape)

\method{lsm_p_gyrate}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Radius of Gyration (patch)
}
\details{
Equals the mean distance of each cell centroid in a patch to the centroid of the whole patch (mean location of all cell centroids).
\subsection{Units}{Meters}
\subsection{Range}{GYRATE >= 0, without limit}
\subsection{Behaviour}{0 if single cell, maximum if patch occupies the entire landscape.}
}
\examples{
lsm_p_gyrate(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
