% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_ncore.R
\name{lsm_p_ncore}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.RasterLayer}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.RasterStack}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.RasterBrick}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.list}
\title{Number of core areas  (patch level)}
\usage{
lsm_p_ncore(landscape)

\method{lsm_p_ncore}{RasterLayer}(landscape)

\method{lsm_p_ncore}{RasterStack}(landscape)

\method{lsm_p_ncore}{RasterBrick}(landscape)

\method{lsm_p_ncore}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Number of core cells
}
\examples{
lsm_p_ncore(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
