% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_ncore.R
\name{lsm_p_ncore}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.RasterLayer}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.RasterStack}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.RasterBrick}
\alias{lsm_p_ncore}
\alias{lsm_p_ncore.list}
\title{Number of core areas (patch level)}
\usage{
lsm_p_ncore(landscape, ...)

\method{lsm_p_ncore}{RasterLayer}(landscape, ...)

\method{lsm_p_ncore}{RasterStack}(landscape, ...)

\method{lsm_p_ncore}{RasterBrick}(landscape, ...)

\method{lsm_p_ncore}{list}(landscape, ...)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{...}{Specific arguments for certain functions, if not provided they fall back to default.

#' @details
Equals the number of disjunct core areas. A core area is a 'patch within the patch' without
any edge cells. In other words, the number of patches within the patch that only have
neighbouring cells of the same type
\subsection{Units}{None}
\subsection{Range}{NCORE >= 0}
\subsection{Behaviour}{NCORE = 0 when CORE = 0, i.e. every cell in patches of class i is
an edge. NCORE increases without limit as core areas become more present, i.e. patches
becoming larger and less complex}}
}
\value{
tibble
}
\description{
Number of disjunct core areas (patch level)
}
\examples{
lsm_p_ncore(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
