% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_para.R
\name{lsm_p_para}
\alias{lsm_p_para}
\alias{lsm_p_para}
\alias{lsm_p_para.RasterLayer}
\alias{lsm_p_para}
\alias{lsm_p_para.RasterStack}
\alias{lsm_p_para}
\alias{lsm_p_para.RasterBrick}
\alias{lsm_p_para}
\alias{lsm_p_para.list}
\title{Perimeter-area ratio  (patch level)}
\usage{
lsm_p_para(landscape)

\method{lsm_p_para}{RasterLayer}(landscape)

\method{lsm_p_para}{RasterStack}(landscape)

\method{lsm_p_para}{RasterBrick}(landscape)

\method{lsm_p_para}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{details}{The perimeter-area ratio equals patch perimeter
divided by patch area (no unit) and is a simple measure of complexity
\deqn{PARA = perimeter[patch] / area[patch]}
\subsection{Units}{None}
\subsection{Range}{PARA = 1 for a square and
increases as the patch becomes more irregular}}
}
\value{
tibble
}
\description{
Ratio of patch perimeter and area (patch level)
}
\examples{
lsm_p_para(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
