% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_para.R
\name{lsm_p_para}
\alias{lsm_p_para}
\alias{lsm_p_para}
\alias{lsm_p_para.RasterLayer}
\alias{lsm_p_para}
\alias{lsm_p_para.RasterStack}
\alias{lsm_p_para}
\alias{lsm_p_para.RasterBrick}
\alias{lsm_p_para}
\alias{lsm_p_para.list}
\title{PARA (patch level)}
\usage{
lsm_p_para(landscape)

\method{lsm_p_para}{RasterLayer}(landscape)

\method{lsm_p_para}{RasterStack}(landscape)

\method{lsm_p_para}{RasterBrick}(landscape)

\method{lsm_p_para}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Perimeter-Area ratio (Shape metric)
}
\details{
\deqn{PARA = \frac{p_{ij}}{a_{ij}}}
where \eqn{p_{ij}} is the perimeter in meters and \eqn{a_{ij}} is the
area in square meters.

PARA is a 'Shape metric'. It describes the patch complexity in a straightforward way.
However, because it is not standarised to a certain shape (e.g. a square), it
is not scale independent.

\subsection{Units}{None}
\subsection{Range}{PARA > 0}
\subsection{Behaviour}{PARA increases, without limit, as the shape comlexity increases.}
}
\examples{
lsm_p_para(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_area}},
\code{\link{lsm_p_perim}}, \cr
\code{\link{lsm_c_para_mn}},
\code{\link{lsm_c_para_sd}},
\code{\link{lsm_c_para_cv}}, \cr
\code{\link{lsm_l_para_mn}},
\code{\link{lsm_l_para_sd}},
\code{\link{lsm_l_para_cv}}
}
