% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_perim.R
\name{lsm_p_perim}
\alias{lsm_p_perim}
\alias{lsm_p_perim}
\alias{lsm_p_perim.RasterLayer}
\alias{lsm_p_perim}
\alias{lsm_p_perim.RasterStack}
\alias{lsm_p_perim}
\alias{lsm_p_perim.RasterBrick}
\alias{lsm_p_perim}
\alias{lsm_p_perim.list}
\title{PERIM (patch level)}
\usage{
lsm_p_perim(landscape, directions)

\method{lsm_p_perim}{RasterLayer}(landscape, directions = 8)

\method{lsm_p_perim}{RasterStack}(landscape, directions = 8)

\method{lsm_p_perim}{RasterBrick}(landscape, directions = 8)

\method{lsm_p_perim}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.
#' @param directions The number of directions in which cells should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
tibble
}
\description{
Perimeter (Area and edge metric))
}
\details{
\deqn{PERIM = p_{ij}}
where \eqn{p_{ij}} is the perimeter in meters.

PERIM is an 'Area and edge metric'. It equals the perimeter of the patch
including also the edge to the landscape boundary. The metric describes
patch area (larger perimeter for larger patches), but also patch shape
(large perimeter for irregular shapes).

\subsection{Units}{Meters}
\subsection{Range}{PERIM > 0}
\subsection{Behaviour}{Increases, without limit, as patch size and
complexity increases.}
}
\examples{
lsm_p_perim(landscape)

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html
}
