% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_perim.R
\name{lsm_p_perim}
\alias{lsm_p_perim}
\alias{lsm_p_perim}
\alias{lsm_p_perim.RasterLayer}
\alias{lsm_p_perim}
\alias{lsm_p_perim.RasterStack}
\alias{lsm_p_perim}
\alias{lsm_p_perim.RasterBrick}
\alias{lsm_p_perim}
\alias{lsm_p_perim.list}
\title{Perimeter  (patch level)}
\usage{
lsm_p_perim(landscape)

\method{lsm_p_perim}{RasterLayer}(landscape)

\method{lsm_p_perim}{RasterStack}(landscape)

\method{lsm_p_perim}{RasterBrick}(landscape)

\method{lsm_p_perim}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Perimeter of patch (patch level)
}
\details{
Equals the perimeter of the patch.
\subsection{Units}{Meter (assuming that the input cellsize is in meter)}
\subsection{Range}{PERIm > 0 and increases without limit as the patch size increases}
}
\examples{
lsm_p_perim(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
