% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_shape.R
\name{lsm_p_shape}
\alias{lsm_p_shape}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterLayer}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterStack}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterBrick}
\alias{lsm_p_shape}
\alias{lsm_p_shape.list}
\title{SHAPE (patch level)}
\usage{
lsm_p_shape(landscape)

\method{lsm_p_shape}{RasterLayer}(landscape)

\method{lsm_p_shape}{RasterStack}(landscape)

\method{lsm_p_shape}{RasterBrick}(landscape)

\method{lsm_p_shape}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Shape index (Shape metric)
}
\details{
\deqn{SHAPE = \frac{0.25 * p_{ij}}{\sqrt{a_{ij}}}}
where \eqn{p_{ij}} is the perimeter in meters and \eqn{a_{ij}} is the area in
square meters.

SHAPE is a 'Shape metric'. The index is based on the patch perimeter and the patch area
and describes the patch complexity. Because it it standarised, it is independent
of the patch size.

\subsection{Units}{None}
\subsection{Range}{SHAPE >= 1}
\subsection{Behaviour}{Approaches SHAPE = 1 for a square and
increases, without limit, as the patch shape becomes more complex.}
}
\examples{
lsm_p_shape(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
\seealso{
\code{\link{lsm_p_perim}} and \code{\link{lsm_p_area}}
}
