% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_p_shape.R
\name{lsm_p_shape}
\alias{lsm_p_shape}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterLayer}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterStack}
\alias{lsm_p_shape}
\alias{lsm_p_shape.RasterBrick}
\alias{lsm_p_shape}
\alias{lsm_p_shape.list}
\title{Shape index  (patch level)}
\usage{
lsm_p_shape(landscape)

\method{lsm_p_shape}{RasterLayer}(landscape)

\method{lsm_p_shape}{RasterStack}(landscape)

\method{lsm_p_shape}{RasterBrick}(landscape)

\method{lsm_p_shape}{list}(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Shape index of patch (patch level)
}
\details{
The shape index equals a quater of the patch perimeter divided by the square root
of the patch area
\deqn{SHAPE = 0.25 * perimeter[patch] / sqrt(area[patch])}
\subsection{Units}{None}
\subsection{Range}{SHAPE >= 1}
\subsection{Behaviour}{SHAPE = 1 for a square and
increases as the patch becomes more irregular}
}
\examples{
lsm_p_shape(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
