% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_c_ta.R
\name{lsm_c_ta}
\alias{lsm_c_ta}
\alias{lsm_ta}
\title{Total class area (class level)}
\usage{
lsm_c_ta(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
tibble
}
\description{
Total area of class (class level)
}
\examples{
lsm_ta(landscape)

}
\references{
McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis
program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep.
PNW-351.
}
