% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_raster.R
\name{matrix_to_raster}
\alias{matrix_to_raster}
\title{matrix_to_raster}
\usage{
matrix_to_raster(matrix, landscape, to_disk = FALSE)
}
\arguments{
\item{matrix}{matrix with values.}

\item{landscape}{RasterLayer.}

\item{to_disk}{If TRUE raster will be saved to disk}
}
\value{
raster
}
\description{
Adding padding to raster
}
\details{
Converts \code{matrix} to a raster with same characteristics as \code{landscape}.
}
\examples{
test_matrix <- raster::as.matrix(augusta_nlcd)
matrix_to_raster(matrix = test_matrix, landscape = augusta_nlcd)

}
\keyword{internal}
