% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R, R/matrix_to_raster.R
\name{get_boundaries}
\alias{get_boundaries}
\alias{matrix_to_raster}
\title{get_boundaries}
\usage{
get_boundaries(landscape, directions = 4)

matrix_to_raster(matrix, landscape = NULL, landscape_empty = FALSE,
  x = NULL, resolution = NULL, crs = NULL, to_disk = FALSE)
}
\arguments{
\item{landscape}{RasterLayer.}

\item{directions}{4 (rooks case), 8 (queens case)}

\item{matrix}{matrix with values.}

\item{landscape_empty}{If true, RasterLayer is landscape_empty}

\item{x}{Extent of RasterLayer.}

\item{resolution}{Resolution of RasterLayer.}

\item{crs}{CRS of raster layer.}

\item{to_disk}{If TRUE raster will be saved to disk.}

\item{matrix}{matrix with values.}
}
\value{
matrix

raster
}
\description{
Streamlined version of raster::boundaries that takes a matrix as input

Adding padding to raster
}
\details{
Stripped down version of raster::boundaries so that we internally can parse
a matrix.

Converts \code{matrix} to a raster with same characteristics as \code{landscape}. Either
\code{landscape} or \code{x}, \code{resolution} and \code{crs} must be specified.
}
\examples{
test_matrix <- raster::as.matrix(landscape)
get_boundaries(test_matrix)

test_matrix <- raster::as.matrix(augusta_nlcd)
matrix_to_raster(matrix = test_matrix, landscape = augusta_nlcd)

}
\author{
Robert Hijmans
}
\keyword{internal}
