% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_raster.R
\name{pad_raster}
\alias{pad_raster}
\alias{pad_raster}
\alias{pad_raster.RasterLayer}
\alias{pad_raster}
\alias{pad_raster.RasterStack}
\alias{pad_raster}
\alias{pad_raster.RasterBrick}
\alias{pad_raster}
\alias{pad_raster.list}
\title{pad_raster}
\usage{
pad_raster(landscape, pad_raster_value, pad_raster_cells, global)

\method{pad_raster}{RasterLayer}(landscape, pad_raster_value = -999,
  pad_raster_cells = 1, global = FALSE)

\method{pad_raster}{RasterStack}(landscape, pad_raster_value = -999,
  pad_raster_cells = 1, global = FALSE)

\method{pad_raster}{RasterBrick}(landscape, pad_raster_value = -999,
  pad_raster_cells = 1, global = FALSE)

\method{pad_raster}{list}(landscape, pad_raster_value = -999,
  pad_raster_cells = 1, global = FALSE)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{pad_raster_value}{Value of cells added}

\item{pad_raster_cells}{Number of rows and columns added}

\item{global}{Only pad around the raster extent or also NA holes "inside"}
}
\value{
raster
}
\description{
Adding padding to raster
}
\details{
Adds equally (in all four directions) additional cells around the raster
}
\examples{
pad_raster(landscape)

}
\keyword{internal}
