% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_info.R
\name{proj_info}
\alias{proj_info}
\title{proj_info}
\usage{
proj_info(landscape)
}
\arguments{
\item{landscape}{RasterLayer}
}
\value{
raster
}
\description{
Projection info
}
\details{
Check projection of RasterLayer
}
\examples{
proj_info(terra::unwrap(landscape))
proj_info(terra::unwrap(augusta_nlcd))
proj_info(terra::unwrap(podlasie_ccilc))

}
\keyword{internal}
