% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_points.R
\name{raster_to_points}
\alias{raster_to_points}
\title{raster_to_points}
\usage{
raster_to_points(landscape, return_NA = TRUE)
}
\arguments{
\item{landscape}{A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.}

\item{return_NA}{If true, NA cells are also included}
}
\value{
matrix
}
\description{
Raster to points
}
\details{
Wrapper around terra::xyFromCell and terra::getValues to get raster_to_points
function including NA values
}
\examples{
landscape <- terra::rast(landscape)
raster_to_points(landscape)

}
\keyword{internal}
