% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_points.R
\name{raster_to_points}
\alias{raster_to_points}
\title{raster_to_points}
\usage{
raster_to_points(landscape)
}
\arguments{
\item{landscape}{Raster* Layer, Stack, Brick or a list of rasterLayers.}
}
\value{
matrix
}
\description{
Raster to points
}
\details{
Wrapper around raster::xyFromCell and raster::getValues to get raster_to_points
function including NA values
}
\examples{
raster_to_points(landscape)

}
\keyword{internal}
