% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_pairs}
\alias{rcpp_get_pairs}
\title{Adjacent cells values}
\usage{
rcpp_get_pairs(x, directions = 4L)
}
\arguments{
\item{x}{A matrix}

\item{directions}{The number of directions in which cells should be connected:
4 (rook's case) or 8 (queen's case)}
}
\description{
Extract values of adjacent cells.
An output is a two columns matrix, where the first column contains a
main cell value and the second column contains an adjacent cell value.
}
