% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_xy_from_matrix}
\alias{rcpp_xy_from_matrix}
\title{Coordinates from a matrix}
\usage{
rcpp_xy_from_matrix(x, cell = NULL)
}
\arguments{
\item{x}{A matrix}

\item{cell}{A vector of cell number.
If NULL, the coordinates will be calculated for the whole matrix}
}
\description{
This function gets coordinates (row and column numbers) of the matrix cells.
}
