% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_correlation.R
\name{show_correlation}
\alias{show_correlation}
\title{Show correlation}
\usage{
show_correlation(metrics, level, method = "pearson", text_size = 15)
}
\arguments{
\item{metrics}{tibble with results of as returned by the landscapemetrics package}

\item{method}{type of correlation. See \code{link{cor}} for details}

\item{text_size}{Text size of the plot}
}
\value{
ggplot
}
\description{
Show correlation
}
\details{
The functions calculates the correlation between all metrics. All metrics
of the input tibble must be from the same level. In order to calculate correlations,
for the landscape level more than one landscape needs to be present. All input
must be structured as returned by the \strong{landscapemetrics} package.
}
\examples{
metrics <- calculate_metrics(landscape, what = "patch")
show_correlation(metrics, level = "patch", method = "pearson")

}
