% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_patches.R
\name{show_patches}
\alias{show_patches}
\alias{show_patches}
\alias{show_patches.RasterLayer}
\alias{show_patches}
\alias{show_patches.RasterStack}
\alias{show_patches}
\alias{show_patches.RasterBrick}
\alias{show_patches}
\alias{show_patches.list}
\title{Plot connected labeld}
\usage{
show_patches(landscape)

\method{show_patches}{RasterLayer}(landscape)

\method{show_patches}{RasterStack}(landscape)

\method{show_patches}{RasterBrick}(landscape)

\method{show_patches}{list}(landscape)
}
\arguments{
\item{landscape}{*Raster object}
}
\value{
List with RasterLayer/RasterBrick
}
\description{
xxx
}
\details{
xxx
}
\examples{
show_patches(landscape)
show_patches(landscape_stack)

}
