% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_patches.R
\name{show_patches}
\alias{show_patches}
\alias{show_patches}
\alias{show_patches.RasterLayer}
\alias{show_patches}
\alias{show_patches.RasterStack}
\alias{show_patches}
\alias{show_patches.RasterBrick}
\alias{show_patches}
\alias{show_patches.list}
\title{Show patches}
\usage{
show_patches(landscape, directions)

\method{show_patches}{RasterLayer}(landscape, directions = 8)

\method{show_patches}{RasterStack}(landscape, directions = 8)

\method{show_patches}{RasterBrick}(landscape, directions = 8)

\method{show_patches}{list}(landscape, directions = 8)
}
\arguments{
\item{landscape}{*Raster object}

\item{directions}{The number of directions in which cells should be
connected: 4 (rook's case) or 8 (queen's case).}
}
\value{
ggplot
}
\description{
Show patches
}
\details{
The functions plots the landscape with the patches labelled with the
corresponding patch id.
}
\examples{
show_patches(landscape)

}
